AC* helper: define the individual's location given detection(s)

These function defines the relative plausibility possible locations of an individual given one or more detections, either across a SpatRaster (for acs()) or for selected particle positions (for pf_forward_2()).

.acs_given_detection_SpatRaster(
  .detections,
  .absences,
  .kernels,
  .zero_to_na = FALSE
)

.acs_given_detection_particles(.detections, .absences, .kernels, .particles)

Arguments

.detections

An integer vector of the receiver(s) that recorded detections at a given time step.

.absences

An integer vector of the remaining, overlapping receiver(s) that did not record a detection, from .acs_absences().

.kernels

A list from acs_setup_detection_kernels.

.zero_to_na

For .acs_given_detection_SpatRaster, .zero_to_na is a logical variable that defines whether or not to classify zeros as NAs. This should be FALSE for defining given_data in acs(), but TRUE for defining next_kernel in acs(), so that next_kernel is correctly buffered.

.particles

For .acs_given_detection_particles, .particles is a data.table witha cell_now column that defines particle locations on the grid.

Details

In the AC* algorithms, at the moment of detection, the probability kernels that describe the possible locations of an individual given the data depend on both the receivers that record detections and those that did not (eqn S5 in Lavender et al., 2023). This function solves eqn S5. For computational efficiency, the equation is solved in a stepwise manor such that the number of necessary operations is kept to a minimum.

Warning

For speed, these functions performs no internal checks.

Author

Edward Lavender