AC* set up: define detection container overlaps

This function identifies receivers with overlapping detection containers in space and time for the AC* algorithms.

acs_setup_detection_overlaps(.containers, .moorings, .services = NULL)

Source

This function is based on the get_detection_containers_overlaps function in the flapper package.

Arguments

.containers

A named list of SpatRasters that represent receiver detection containers, from acs_setup_detection_containers().

.moorings

A data.table that defines receiver deployments and associated information (see dat_moorings for an example). At a minimum, this must contain the following columns:

receiver_id---an integer vector of receiver IDs;

receiver_start---a Date vector that defines receiver deployment dates;

receiver_end---a Date vector that defines receiver retrieval dates;

.services

(optional) A data.table that defines receiver IDs and servicing Dates (times during the deployment period of a receiver when it was not active due to servicing) (see make_matrix_receivers()). If provided, this must contain the following columns:

receiver_id---an integer vector of receiver IDs;

service_start---a Date vector that defines receiver servicing start dates;

service_end---a Date vector that defines receiver servicing completion dates;

Value

The function returns a named list with two elements:

overlap_by_receiver is list, with one element for all integers from 1:max(.moorings$receiver_id). Any elements that do not correspond to receivers contain a NULL element. List elements that correspond to receivers contain a data.frame that defines, for each day over the deployment period (defined in timestamp) of that receiver (defined in receiver_id), whether (1) or not (0) that receiver overlapped in space with every other receiver (defined in the remaining columns by their receiver IDs).

overlap_by_date is a named list, with one element for each Date from the start until the end of the study (min(.moorings$receiver_start_date):max(.moorings$receiver_end_date)), that records an integer vector of all receivers with overlapping containers on that date. In this vector, each receiver overlaps with at least one other receiver (but not every receiver will necessarily overlap with every other receiver).

Details

In the AC* algorithms, at the moment of detection, the set of possible locations depends on the receiver(s) at which an individual is, and is not, detected. The outputs of this function are used to restrict the probability calculations to the set of receivers that overlap with the receiver(s) at which an individual is detected for improved efficiency.

Author

Edward Lavender

Examples

#### Example (1): Basic implementation
# Define detection containers
dat_moorings$receiver_range <- 500
containers <- acs_setup_detection_containers(dat_gebco(), dat_moorings)
# Identify receiver overlaps
overlaps <- acs_setup_detection_overlaps(containers, dat_moorings)
summary(overlaps)
#>                  Length Class  Mode
#> list_by_receiver  57    -none- list
#> list_by_date     508    -none- list