AC* set up: set up movement datasets

This function processes passive acoustic telemetry detections and/or archival time series for pf_forward().

acs_setup_obs(
  .acoustics = NULL,
  .archival = NULL,
  .trim = TRUE,
  .step,
  .period = NULL,
  .mobility,
  .detection_range
)

Arguments

.acoustics

(optional) A data.table that defines passive acoustic telemetry detections (see dat_acoustics for an example) for a single individual. At a minimum, this must contain two columns:

timestamp---an ordered, POSIXct vector that defines the times of detections;

receiver_id---an integer vector that defines the receiver(s) that recorded detections;

.archival

(optional) A data.table that defines depth (m) observations (see dat_archival for an example) for the same individual. At a minimum, this must contain two columns:

timestamp---an ordered, POSIXct vector that defines the times of observations (which may be regular or irregular);

depth---a positive-valued numeric vector that defines the individual's depth (m) below the surface at each time step;

.trim

If .archival is supplied, .trim is logical variable that defines whether or not to trim .acoustic and .archival time series to the time period for which they overlap.

.step

An character, passed to lubridate::period(), lubridate::round_date() and seq() that defines the duration between sequential time steps (e.g., "2 mins").

.period

(optional) A length-two POSIXct vector that defines the start and end time of the observations. If unprovided, this is defined internally according to .acoustics, .archival and .trim.

.mobility

A constant that defines the maximum (Euclidean) distance the individual could move in .step.

.detection_range

A constant that defines the detection range. A constant value across all receivers and time steps is assumed.

Value

The function returns a data.table with the following columns:

timestep---an integer that defines the time step;

timestamp---a regular sequences of POSIXct time stamps;

date---a character that defines the date;

mobility---a double that defines .mobility;

If .acoustics is provided, the following column(s) are also included:

detection_id---an integer vector that uniquely defines each detection;

detection---an integer that distinguishes the time steps at which detections were (1) or were not (0) recorded;

receiver_id---a list that defines the receiver(s) that recorded detection(s) at each time step;

receiver_id_next---a list that defines the receiver(s) that recorded the next detection(s);

buffer_past---a double that controls container growth from the past to the present;

buffer_future---a double that controls container shrinkage from the future to the present;

buffer_future_incl_gamma---a double (buffer_future + .detection_range) that controls container shrinkage for pf_forward();

If .archival is provided, the following column(s) are also included:

depth---a number that defines the individual's depth (m) at each time step;

Details

This function implements the following routines:

Acoustic progressing (if applicable);

Acoustic time series are rounded to the nearest .step;

All receivers that recorded detection(s) in each time interval are listed;

Duplicate detections (at the same receiver in the same time interval) are dropped;

Archival processing (if applicable);

Archival time series are rounded to the nearest .step;

Duplicate records (in the same time interval) are not permitted;

Alignment (if applicable)

Acoustic and archival time series are optionally trimmed to the time period for which they overlap;;

Time series

Acoustic and/or archival time series are mapped onto a regular time series;

Information required by pf_forward() is added

See also

Author

Edward Lavender

Examples

#### Define example datasets
acoustics <- dat_acoustics[dat_acoustics$individual_id == 25, ]
archival  <- dat_archival[dat_archival$individual_id == 25, ]

#### Example (1): Implement the function for acoustic time series only
obs <- acs_setup_obs(acoustics,
                     .step = "2 mins",
                     .mobility = 500,
                     .detection_range = 750)
utils::head(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:        1 2016-03-17 01:50:00 2016-03-17            1         1          26
#> 2:        2 2016-03-17 01:52:00 2016-03-17            2         1          26
#> 3:        3 2016-03-17 01:54:00 2016-03-17            3         1          26
#> 4:        4 2016-03-17 01:56:00 2016-03-17            3         0            
#> 5:        5 2016-03-17 01:58:00 2016-03-17            4         1          26
#> 6:        6 2016-03-17 02:00:00 2016-03-17            5         1          26
#>    receiver_id_next mobility buffer_past buffer_future buffer_future_incl_gamma
#> 1:               26      500         500           500                     1250
#> 2:               26      500         500           500                     1250
#> 3:               26      500         500          1000                     1750
#> 4:               26      500         500           500                     1250
#> 5:               26      500         500           500                     1250
#> 6:               26      500         500          1000                     1750

#### Example (2): Use alternative parameters
obs <- acs_setup_obs(acoustics,
                     .step = "4 mins",
                     .mobility = 1000,
                     .detection_range = 500)
utils::head(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:        1 2016-03-17 01:52:00 2016-03-17            1         1          26
#> 2:        2 2016-03-17 01:56:00 2016-03-17            2         1          26
#> 3:        3 2016-03-17 02:00:00 2016-03-17            3         1          26
#> 4:        4 2016-03-17 02:04:00 2016-03-17            4         1          26
#> 5:        5 2016-03-17 02:08:00 2016-03-17            5         1          26
#> 6:        6 2016-03-17 02:12:00 2016-03-17            6         1          26
#>    receiver_id_next mobility buffer_past buffer_future buffer_future_incl_gamma
#> 1:               26     1000        1000          1000                     1500
#> 2:               26     1000        1000          1000                     1500
#> 3:               26     1000        1000          1000                     1500
#> 4:               26     1000        1000          1000                     1500
#> 5:               26     1000        1000          1000                     1500
#> 6:               26     1000        1000          1000                     1500

#### Example (3): Implement the function for acoustic & archival time series
obs <- acs_setup_obs(acoustics, archival,
                     .step = "2 mins",
                     .mobility = 500,
                     .detection_range = 750)
utils::head(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:        1 2016-03-17 01:50:00 2016-03-17            1         1          26
#> 2:        2 2016-03-17 01:52:00 2016-03-17            2         1          26
#> 3:        3 2016-03-17 01:54:00 2016-03-17            3         1          26
#> 4:        4 2016-03-17 01:56:00 2016-03-17            3         0            
#> 5:        5 2016-03-17 01:58:00 2016-03-17            4         1          26
#> 6:        6 2016-03-17 02:00:00 2016-03-17            5         1          26
#>    receiver_id_next mobility buffer_past buffer_future buffer_future_incl_gamma
#> 1:               26      500         500           500                     1250
#> 2:               26      500         500           500                     1250
#> 3:               26      500         500          1000                     1750
#> 4:               26      500         500           500                     1250
#> 5:               26      500         500           500                     1250
#> 6:               26      500         500          1000                     1750
#>    depth
#> 1: 73.78
#> 2: 73.32
#> 3: 73.32
#> 4: 73.32
#> 5: 73.55
#> 6: 68.70

#### Example (4): Include full acoustic & archival time series
range(acoustics$timestamp)
#> [1] "2016-03-17 01:50:00 UTC" "2017-06-01 05:58:00 UTC"
range(archival$timestamp)
#> [1] "2016-03-16 00:00:00 UTC" "2016-04-19 17:18:00 UTC"
obs <- acs_setup_obs(acoustics, archival, .trim = FALSE,
                     .step = "2 mins",
                     .mobility = 500,
                     .detection_range = 750)
#> Warning: There are 25000/318420 (7.9 %) archival observations within the time series.
utils::head(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:        1 2016-03-16 00:00:00 2016-03-16           NA         0            
#> 2:        2 2016-03-16 00:02:00 2016-03-16           NA         0            
#> 3:        3 2016-03-16 00:04:00 2016-03-16           NA         0            
#> 4:        4 2016-03-16 00:06:00 2016-03-16           NA         0            
#> 5:        5 2016-03-16 00:08:00 2016-03-16           NA         0            
#> 6:        6 2016-03-16 00:10:00 2016-03-16           NA         0            
#>    receiver_id_next mobility buffer_past buffer_future buffer_future_incl_gamma
#> 1:               26      500         500        387500                   388250
#> 2:               26      500         500        387000                   387750
#> 3:               26      500         500        386500                   387250
#> 4:               26      500         500        386000                   386750
#> 5:               26      500         500        385500                   386250
#> 6:               26      500         500        385000                   385750
#>     depth
#> 1: 159.15
#> 2: 161.01
#> 3: 162.86
#> 4: 164.48
#> 5: 158.46
#> 6: 155.45
utils::tail(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:   318415 2017-06-01 05:48:00 2017-06-01        12489         1          47
#> 2:   318416 2017-06-01 05:50:00 2017-06-01        12489         0            
#> 3:   318417 2017-06-01 05:52:00 2017-06-01        12489         0            
#> 4:   318418 2017-06-01 05:54:00 2017-06-01        12489         0            
#> 5:   318419 2017-06-01 05:56:00 2017-06-01        12490         1          47
#> 6:   318420 2017-06-01 05:58:00 2017-06-01        12491         1          47
#>    receiver_id_next mobility buffer_past buffer_future buffer_future_incl_gamma
#> 1:               47      500         500          2000                     2750
#> 2:               47      500         500          1500                     2250
#> 3:               47      500         500          1000                     1750
#> 4:               47      500         500           500                     1250
#> 5:               47      500         500           500                     1250
#> 6:               NA      500         500           500                     1250
#>    depth
#> 1:    NA
#> 2:    NA
#> 3:    NA
#> 4:    NA
#> 5:    NA
#> 6:    NA