Assemble observations

These functions assemble a timeline and observations for the particle filter (pf_filter()).

assemble_timeline(.datasets = list(), .step, .trim = FALSE)

assemble_acoustics(.timeline, .detections, .moorings, .services = NULL)

assemble_acoustics_containers(
  .timeline,
  .acoustics,
  .mobility,
  .map = NULL,
  .threshold = NULL
)

assemble_archival(.timeline, .archival)

assemble_custom(.timeline, .dataset)

Arguments

.datasets, .step, .trim

Arguments for assemble_timeline().

.datasets—A list of data.tables, one for each data type, each containing a timestamp column;

.step—A character (such as "2 mins"), passed to lubridate::round_date() and seq.POSIXt(), that defines the resolution of the timeline;

.trim—A logical variable that defines whether or not to trim the timeline to the overlapping period between datasets;

.timeline

A POSIXct vector of regularly spaced time stamps that defines the timeline for the simulation (optionally from assemble_timeline()). Here, timeline is used to:

Define the resolution of observations;

.detections, .moorings, .services

The data.tables for assemble_acoustics() (see pat_setup_data()).

.detections is a data.table of acoustic detections for a single individual. This must contain the receiver_id and timestamp columns.

.moorings is a data.table of acoustic receiver deployments. This must contain the receiver_id, receiver_start, and receiver_end columns, plus (optional) additional parameter columns.

(optional) .services is a data.table of servicing events. This must contain the receiver_id, service_start and service_end columns.

.acoustics, .mobility, .map, .threshold

Arguments for assemble_acoustics_containers().

.acoustics is a data.table of acoustic observations, from assemble_acoustics().

.mobility is the maximum movement distance (m) between two time steps (and sets the rate of container contraction).

.map, .threshold are distance threshold options (see Details). Specify .map or .threshold.

.map is a two-column matrix of the four coordinates of the study area or a SpatRaster or SpatVector from which such a matrix can be obtained. If provided, .threshold is set automatically based on the distances between receivers and the boundaries of the study area.

Otherwise, .threshold is a double that defines the distance threshold.

.archival

For assemble_archival(), .archival is a data.table of depth observations for a single individual with timestamp and obs columns (see .dataset, below).

.dataset

For assemble_custom(), .dataset is a data.table of observations (such as depth measurements) for a single individual. This must contain timestamp and obs columns plus (optional) additional parameter columns.

Value

assemble_timeline() returns a POSIXct vector;

assemble_acoustics(), assemble_archival() and assemble_custom() return a data.table for pf_filter();

assemble_acoustics_containers() returns a named list, with one element (data.table) for (a) the forward and (b) the backward runs of pf_filter();

Details

assemble_timeline() is a simple function that defines a regular timeline, of resolution .step, from a list of input datasets.

If .trim = FALSE, this defines a sequence of regular time stamps across the full range of time stamps in the input datasets.

If .trim = TRUE, the timeline is trimmed to the overlapping period between datasets.

assemble_{dataset}() functions are helper routines that prepare timelines observations for different data types as required for the particle filter (pf_filter()). The filter expects a named list of datasets (one for each data type). Each dataset must contain the following columns: timestamp, sensor_id, obs and additional columns with the parameters of the observation model (see glossary).

assemble_acoustics() prepares a timeline of acoustic observations as required by the filter for a single individual. This function expects a 'standard' detection dataset (that is, a data.table like dat_detections but for a single individual) that defines detections at receivers alongside a moorings dataset (like dat_moorings) that defines receiver deployment periods and, optionally, a data.table of servicing events (when receiver(s) were non-operational). assemble_acoustics() uses these datasets to assemble a complete time series of acoustic observations; that is, a data.table of time stamps and receivers that defines, for each time step and each operational receiver whether (1L) or not (0L) a detection was recorded at that time step. Duplicate observations (that is, detections at the same receiver in the same time step) are dropped. If available in .moorings, additional columns (receiver_alpha, receiver_beta and receiver_gamma) are included as required for the default acoustic observation model (that is, ModelObsAcousticLogisTrunc). If observation model parameters vary both by receiver and through time, simply amend these columns as required.

assemble_acoustics_containers() prepares a dataset of acoustic containers, given the acoustic time series from assemble_acoustics(). Acoustic containers define the region within which an individual must be located at a given time step according to the receiver(s) at which it was next detected. Each container is a circular region, of radius \(r\), around a receiver that recorded a detection at the next time step. The radius depends on the time until the next detection, the maximum movement speed and the detection range around the receiver at the time of the detection. For example, if an individual can move up to .mobility = 500 m per time step, and two time steps elapse between the first and second detections, then at the moment of the first detection the individual must be within 1000 m of the detection range (say, receiver_gamma = 750 m) of the second receiver; that is, at the moment of first detection, the maximum possible distance of the individual from the receiver that recorded the next detection is 1750 m. As time passes, the container shrinks towards the receiver(s) that recorded the next detection(s), in line with the individual's .mobility. Acoustic containers, coupled with regular re-sampling, facilitate convergence in the particle filter: at each time step, only particles within an acoustic container are duplicated (other particles are killed), which encourages particles to move towards the next receiver(s) by the time of the next detection(s). In practice, this works as follows. assemble_acoustics_containers() assembles a data.table that defines the maximum distance (radius) of the individual from the receiver(s) that recorded the next detection. For computational efficiency, this data.table only includes containers with a radius < .threshold. If .map is supplied, the .threshold is set to the maximum distance between each receiver and the furthest corner of the study area. Otherwise, set the .threshold to the desired value. The data.table from assemble_acoustics_containers() is used to instantiate a Vector of ModelObsAcousticContainer sub-types in Julia. In the particle filter, for each particle, we compute the log-probability of the particle from the distance of the particle from the relevant receiver (0.0 or -Inf). By re-sampling particles with replacement, particles that move in a way that is incompatible with the location of the next detection(s) are killed. The bottom line is that if have acoustic observations, you should also include acoustic containers in the list of 'observations' for the particle filter (pf_filter()). This function requires the tidyr::nest(), tidyr::unnest() and zoo::na.locf() functions (suggested dependencies).

assemble_archival() prepares a timeline of archival functions for a single individual. This simply wraps assemble_custom() and is informally deprecated.

assemble_custom() prepares a timeline of observations for other data types, as required by the filter. This function expects a data.table that includes, at a minimum, the timestamp and obs columns. The latter defines the observations. The sensor_id column (if unspecified) is simply set to 1L. The function re-expresses time stamps at the resolution specified by timeline. Duplicate observations (that is, multiple measurements in the same time step) throw a warning.

In Julia, datasets are translated into a hash-table (Dict) of observations (via Patter.assemble_yobs()). For each time stamp with an observation, this includes a Vector of Tuples, each containing the observation and the associated ModelObs instance that defines the parameters of the observation model. The particle filter (Patter.particle_filter()) iterates over each time step in the timeline, uses a movement model to simulate animal movement and, for the time stamps with observations, evaluates the likelihood of those observations for the simulated locations (particles).

assemble_*() routines are only required for real-world analyses.

See also

Author

Edward Lavender

Examples

library(data.table)
library(dtplyr)
library(dplyr, warn.conflicts = FALSE)

#### Define example dataset(s) for a selected individual
# Study area map
map <- dat_gebco()
# Acoustic detection time series
# * Observation model parameters are defined in `.moorings`
det <-
  dat_detections |>
  filter(individual_id == 25L) |>
  select("timestamp", "receiver_id") |>
  as.data.table()
# Archival time series
# * Observation model parameters must be included
# * Here, we define parameters for `?ModelObsDepthNormalTrunc`
arc <-
  dat_archival |>
  filter(individual_id == 25L) |>
  select("timestamp", obs = "depth") |>
  mutate(depth_sigma = 50, depth_deep_eps = 20) |>
  as.data.table()

#### Example (1): Define a timeline
# Define a timeline manually
timeline <- seq(as.POSIXct("2016-03-01 00:00:00", tz = "UTC"),
                as.POSIXct("2016-04-01 00:00:00"),
                by = "2 mins")
# Use `assemble_timeline()` with `.trim = FALSE`
timeline <- assemble_timeline(list(det, arc), .step = "2 mins")
range(timeline)
#> [1] "2016-03-16 00:00:00 UTC" "2017-06-01 05:58:00 UTC"
# Use `assemble_timeline()` with `.trim = TRUE`
timeline <- assemble_timeline(list(det, arc), .step = "2 mins", .trim = TRUE)
timeline <- timeline[1:1440]
range(timeline)
#> [1] "2016-03-17 01:50:00 UTC" "2016-03-19 01:48:00 UTC"

#### Example (2): Assemble an acoustic timeline
# Assemble a timeline of acoustic observations (0, 1) and model parameters
# * The default acoustic observation model parameters are taken from `.moorings`
# * But can be modified or added afterwards for custom observation models
acoustics <- assemble_acoustics(.timeline = timeline,
                                .detections = det,
                                .moorings = dat_moorings)
head(acoustics)
#>              timestamp sensor_id   obs receiver_x receiver_y receiver_alpha
#>                 <POSc>     <int> <int>      <num>      <num>          <num>
#> 1: 2016-03-17 01:50:00         3     0   706442.1    6254007              4
#> 2: 2016-03-17 01:50:00         4     0   709742.1    6267707              4
#> 3: 2016-03-17 01:50:00         7     0   708742.1    6269107              4
#> 4: 2016-03-17 01:50:00         9     0   706042.1    6254307              4
#> 5: 2016-03-17 01:50:00        11     0   707542.1    6267707              4
#> 6: 2016-03-17 01:50:00        12     0   710042.1    6267307              4
#>    receiver_beta receiver_gamma
#>            <num>          <num>
#> 1:         -0.01            750
#> 2:         -0.01            750
#> 3:         -0.01            750
#> 4:         -0.01            750
#> 5:         -0.01            750
#> 6:         -0.01            750

#### Example (3): Assemble corresponding acoustic containers
containers <- assemble_acoustics_containers(.timeline = timeline,
                                            .acoustics = acoustics,
                                            .mobility = 750,
                                            .map = map)
# This function returns a list:
summary(containers)
#>          Length Class      Mode
#> forward  6      data.table list
#> backward 6      data.table list
# Use the `forward` element for `pf_filter()` with `.direction = "forward"`
head(containers$forward)
#>              timestamp   obs sensor_id receiver_x receiver_y radius
#>                 <POSc> <int>     <int>      <num>      <num>  <num>
#> 1: 2016-03-17 01:50:00     1        26   709242.1    6253107   1500
#> 2: 2016-03-17 01:52:00     1        26   709242.1    6253107   1500
#> 3: 2016-03-17 01:54:00     1        26   709242.1    6253107   2250
#> 4: 2016-03-17 01:56:00     1        26   709242.1    6253107   1500
#> 5: 2016-03-17 01:58:00     1        26   709242.1    6253107   1500
#> 6: 2016-03-17 02:00:00     1        26   709242.1    6253107   2250
# Use the `backward` for `pf_filter()` with `.direction = "backward"`
head(containers$backward)
#>              timestamp   obs sensor_id receiver_x receiver_y radius
#>                 <POSc> <int>     <int>      <num>      <num>  <num>
#> 1: 2016-03-17 01:52:00     1        26   709242.1    6253107   1500
#> 2: 2016-03-17 01:54:00     1        26   709242.1    6253107   1500
#> 3: 2016-03-17 01:56:00     1        26   709242.1    6253107   1500
#> 4: 2016-03-17 01:58:00     1        26   709242.1    6253107   2250
#> 5: 2016-03-17 02:00:00     1        26   709242.1    6253107   1500
#> 6: 2016-03-17 02:02:00     1        26   709242.1    6253107   1500

#### Example (4): Assemble an archival timeline
# Assemble a timeline of archival observations and model parameters
archival <- assemble_archival(.timeline = timeline,
                              .archival = arc)
head(archival)
#>              timestamp sensor_id   obs depth_sigma depth_deep_eps
#>                 <POSc>     <int> <num>       <num>          <num>
#> 1: 2016-03-17 01:50:00         1 73.78          50             20
#> 2: 2016-03-17 01:52:00         1 73.32          50             20
#> 3: 2016-03-17 01:54:00         1 73.32          50             20
#> 4: 2016-03-17 01:56:00         1 73.32          50             20
#> 5: 2016-03-17 01:58:00         1 73.55          50             20
#> 6: 2016-03-17 02:00:00         1 68.70          50             20

#### Example (5): Assemble custom datasets
temperature <-
  data.table(timestamp = c(as.POSIXct("2016-03-17 01:50:30", tz = "UTC"),
                           as.POSIXct("2016-03-17 02:00:30 UTC", tz = "UTC")),
             obs = c(7.6, 7.7))
temperature <- assemble_custom(.timeline = timeline,
                               .dataset = temperature)
head(temperature)
#>              timestamp sensor_id   obs
#>                 <POSc>     <int> <num>
#> 1: 2016-03-17 01:50:00         1   7.6
#> 2: 2016-03-17 02:00:00         1   7.7

#### Example (6): Implement particle filter
# Use `pf_filter()` to implement the particle filter
# A list of assembled datasets is passed to the `yobs` argument
# The corresponding `ModelObs` sub-types must also be specified