Check input datasets

These functions check input datasets.

check_moorings(.moorings, .class = "data.table")

check_services(.services, .moorings)

check_acoustics(.acoustics)

check_archival(.archival)

Arguments

.moorings, .services, .acoustics, .archival

User inputs. NULL inputs are allowed for .services and .archival.

Value

The functions invisibly return the inputted object if all checks are passed, with the following adjustments:

.moorings$receiver_id is silently coerced to an integer if numeric;

Author

Edward Lavender