Parallelisation helpers (internal)

A set of internal wrappers for parallel::parallel functions that facilitate the implementation of parallel routines in functions via pbapply::pblapply().

Usage

cl_lapply_chunk(.x, .fun, ..., .chunk_fun, .cl, .varlist, .envir)

cl_lapply_elm(.x, .fun, ..., .cl, .varlist, .envir)

cl_check(.cl = NULL, .varlist = NULL)

cl_check_chunk(.fun, .cl, .chunk, .chunk_fun)

cl_cores(.cl = NULL)

cl_chunks(.cl = NULL, .length, .nout = getOption("pboptions")$nout)

cl_export(.cl = NULL, .varlist = NULL, .envir = .GlobalEnv)

cl_stop(.cl = NULL)

Value

cl_lapply_chunk() returns a list.

cl_lapply_elm() returns a list.

cl_cores() returns an integer.

cl_chunks() returns a list of integers.

cl_check(), cl_check_chunk(), cl_export() and cl_stop() return invisible(NULL).

Details

cl_lapply() is the main function. This is a wrapper for pbapply::pblapply() that handles cluster checking, set up, batch processing and cluster closure, using the following functions:

cl_lapply_chunk() implements a function over chunks;

cl_lapply_elm() implements a function over individual elements;

cl_check() checks .cl and .varlist arguments, as inputted to a parent function. For example, if .cl = NULL, .varlist should also be NULL.

cl_check_chunk() checks .chunk arguments.

cl_cores() identifies the number of cores specified.

cl_chunks() defines a list, with one element for each 'chunk' that contains an integer vector of the positions of an object over which to iterate serially in each chunk.

cl_export() implements parallel::clusterExport() if both .cl and .varlist are specified

cl_stop() implements parallel::stopCluster() if .cl is a cluster object from parallel::makeCluster().

These routines evolved from flapper::cl_*() functions.

Author

Edward Lavender