Datasets: example algorithm outputs

These functions load example outputs from key patter functions. They are included in the package to streamline function examples and tests.

dat_dlist()

dat_obs()

dat_pff()

dat_pff_src()

dat_coa()

Source

For full details on the algorithm parameters used to generate these datasets, see https://github.com/edwardlavender/patter/blob/main/data-raw/003-add-data-algorithms.R.

Value

The functions return a dataset or a character string that defines the path to a dataset. See the corresponding function documentation for full details.

Details

dat_dlist() reads example output from pat_setup_data();

dat_pff() reads an example output from pf_setup_obs();

dat_pff() reads an example output from pf_forward();

dat_pff_src() is the directory in which example .parquet files from pf_forward() are stored;

dat_coa() reads an example output from coa();

See also

Author

Edward Lavender

Examples

# Load inbuilt datasets
dat_dlist() |> summary()
#>           Length Class  Mode
#> data      3      -none- list
#> spatial   1      -none- list
#> pars      2      -none- list
#> algorithm 0      -none- list
dat_obs() |> dplyr::glimpse()
#> Rows: 25
#> Columns: 8
#> $ timestep  <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 1…
#> $ timestamp <dttm> 2016-03-17 01:50:00, 2016-03-17 01:52:00, 2016-03-17 01:54:…
#> $ mobility  <dbl> 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, …
#> $ array_id  <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ detection <int> 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, …
#> $ acoustics <list> <1, 0, 0, 0>, <1, 0, 0, 0>, <1, 0, 0, 0>, <NULL>, <1, 0, 0,…
#> $ container <list> [<<matrix[1 x 2]>>, 1250], [<<matrix[1 x 2]>>, 1250], [<<ma…
#> $ depth     <dbl> 73.78, 73.32, 73.32, 73.32, 73.55, 68.70, 72.16, 72.85, 73.5…
dat_pff() |> summary()
#>             Length Class  Mode   
#> history     25     -none- list   
#> convergence  1     -none- logical
#> time         3     -none- list   
dat_coa() |> summary()
#>    timestamp                       coa_x            coa_y        
#>  Min.   :2016-03-17 01:00:00   Min.   :702842   Min.   :6252991  
#>  1st Qu.:2016-03-24 01:00:00   1st Qu.:706442   1st Qu.:6253107  
#>  Median :2016-03-27 21:00:00   Median :709057   Median :6253174  
#>  Mean   :2016-03-28 04:02:24   Mean   :707749   Mean   :6257412  
#>  3rd Qu.:2016-03-29 13:00:00   3rd Qu.:709242   3rd Qu.:6265407  
#>  Max.   :2016-04-16 05:00:00   Max.   :709430   Max.   :6267707  

# Directory of inbuilt parquet files from `pf_forward()`
dat_pff_src()
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward"
pf_files(dat_pff_src())
#> [[1]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/1.parquet"
#> 
#> [[2]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/2.parquet"
#> 
#> [[3]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/3.parquet"
#> 
#> [[4]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/4.parquet"
#> 
#> [[5]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/5.parquet"
#> 
#> [[6]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/6.parquet"
#> 
#> [[7]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/7.parquet"
#> 
#> [[8]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/8.parquet"
#> 
#> [[9]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/9.parquet"
#> 
#> [[10]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/10.parquet"
#> 
#> [[11]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/11.parquet"
#> 
#> [[12]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/12.parquet"
#> 
#> [[13]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/13.parquet"
#> 
#> [[14]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/14.parquet"
#> 
#> [[15]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/15.parquet"
#> 
#> [[16]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/16.parquet"
#> 
#> [[17]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/17.parquet"
#> 
#> [[18]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/18.parquet"
#> 
#> [[19]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/19.parquet"
#> 
#> [[20]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/20.parquet"
#> 
#> [[21]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/21.parquet"
#> 
#> [[22]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/22.parquet"
#> 
#> [[23]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/23.parquet"
#> 
#> [[24]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/24.parquet"
#> 
#> [[25]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpCT5JiX/temp_libpath713b1a760015/patter/extdata/acpf/forward/25.parquet"
#>