DC: a template DC model

This function is an example DC model, of the kind required by .model in dc().

dc_setup_model(
  .obs,
  .t,
  .bathy,
  .unwrap = ifelse(inherits(.bathy, "SpatRaster"), FALSE, TRUE),
  ...
)

Arguments

.obs

The .obs data.table from dc().

.t

An integer that defines the time step (used to index .obs).

.bathy

The .bathy SpatRaster, posibly as a PackedSpatRaster, from dc().

.unwrap

A logical variable that defines whether or not to unwrap .bathy (via terra::unwrap()). This is required if .bathy is passed to this function as a PackedSpatRaster, which is the case for implementations of dc() using socket clusters.

...

Additional arguments passed from dc() (unused here).

Value

The function returns a normalised SpatRaster.

Details

This function assigns equal probability to all locations between two depth limits (as defined by depth_shallow and depth_deep columns in .obs).

Warning

This function is used to streamline examples and does not represent a generically suitable model.

The function does not check user inputs.

Author

Edward lavender

Examples

# For examples, see `?dc()`