AC* helper: Define active, overlapping receivers with absences

Given a detection at one or more receivers on a given date, this function defines the set of remaining, active, overlapping receivers that did not record detections.

.acs_absences(.date, .detections, .overlaps)

Arguments

.date

A character that defines the date (e.g., obs$date[t] internally in acs()).

.detections

An integer vector that defines the receiver(s) that recorded detection(s) at the current time step (e.g., recs_current or recs_next internally in acs()).

.overlaps

A named list from acs_setup_detection_overlaps() that defines receiver overlaps (e.g., detection_overlaps internally in acs()). NULL is permitted.

Value

The function returns an integer vector that defines the set of receivers that overlap with .detections but did not record detections. NULL indicates no overlapping receivers.

Details

In the AC* algorithms, at the moment of detection, the probability kernels that describe the possible locations of an individual given the data depend on both the receivers that record detections and those that did not (eqn S5 in Lavender et al., 2023). This function is used to restrict the set of receivers to which eqn S5 needs to be applied.

Warning

For speed, this function performs no internal checks.

See also

Author

Edward Lavender