AC* helper: define the individual's location given detection(s)

These function defines the relative plausibility possible locations of an individual given one or more detections or for selected particle positions (for pf_forward()).

.acs_given_detection_particles(.detections, .absences, .kernels, .particles)

Arguments

.detections

An integer vector of the receiver(s) that recorded detections at a given time step.

.absences

An integer vector of the remaining, overlapping receiver(s) that did not record a detection, from .acs_absences().

.kernels

A list from acs_setup_detection_kernels.

.particles

A data.table witha cell_now column that defines particle locations on the grid.

Details

In the AC* algorithms, at the moment of detection, likelihood of the acoustic data given a particle sample depends on depends on both the receivers that record detections and those that did not (eqn S5 in Lavender et al., 2023). This function solves eqn S5.

Warning

For speed, these functions performs no internal checks.

Author

Edward Lavender