Study metrics: receiver deployments

This function creates a matrix that, for each time step (matrix row) in a sequence of user-defined times, defines whether or not each receiver (matrix column) was active, optionally accounting for servicing dates. The timeline is defined from a user-provided start and end date (or the range of deployment times in the input data) and an interval (.delta_t) parameter that defines the duration between successive time steps.

make_matrix_receivers(
  .dlist,
  .start = NULL,
  .end = NULL,
  .delta_t = "days",
  .as_POSIXct = NULL,
  .set_names = TRUE
)

Source

This function is based on flapper::make_matrix_receivers. Other flapper::make_matrix_*() routines are not currently available in patter.

Arguments

.dlist

A named list of data and parameters from pat_setup_data(). This function requires:

.dlist$data$moorings, with the following columns: receiver_id, receiver_start, and receiver_end.

(optional) .dlist$data$services, with the following columns: receiver_id, service_start and service_end.

.start, .end

Date or POSIXct objects that define the start and end time. If unspecified, these are taken from the range of deployment times in .moorings.

.delta_t

A numeric or character input that defines the time interval between successive time steps. This is passed to the by argument of base::seq.POSIXt() or base::seq.Date() (depending on .as_POSIXct, below).

.as_POSIXct

(optional) A function that coerces supplied any supplied times (.dlist$data$moorings$receiver_start, .dlist$data$$moorings$receiver_end, .dlist$data$services$service_start, .dlist$data$services$service_end, start and end) that are not POSIXct objects to POSIXct objects. This can be suppressed via .as_POSIXct = NULL if supplied times are Date objects and .delta_t >= one day.

.set_names

A logical variable that defines whether or not to set the row and column names of the matrix to the time steps and the receiver IDs respectively.

Value

The function returns a matrix with one row for each time step and one column for each receiver. Each cell defines whether (1) or not (0) each receiver was at active during that time step. A bins attribute is included, which defines the time steps as a Date or POSIXct vector.

See also

Author

Edward Lavender