patter for passive acoustic telemetry

patter is a re-implementation of the flapper family of algorithms for passive acoustic telemetry that is simpler, faster and better tested than its predecessor. The aim of the package is to facilitate the reconstruction of movement paths and patterns of space use in passive acoustic telemetry systems.

Vignettes

For an introduction to patter, use:

vignette("workflow", package = "patter") for an overview of the analytical workflow;

vignette("faqs", package = "patter") for FAQs;

For a full list of all functions, see help(package = "patter").

Datasets

For example datasets from the Movement Ecology of Flapper Skate project (datasets-mefs), see:

dat_moorings for receiver locations and associated information;

dat_acoustics for acoustic time series;

dat_archival for archival (depth) time series;

dat_gebco() for a bathymetry grid;

For example algorithm outputs (datasets-algorithms), see:

dat_obs() for an example output dataset from acs_setup_obs();

dat_overlaps() for an example output from acs_setup_detection_overlaps();

dat_kernels() for an example output from acs_setup_detection_kernels();

dat_pff() for an example output from pf_forward();

dat_pfb() for an example output from pf_backward_killer();

dat_pfp() for an example output from pf_path();

Simulation

To simulate passive acoustic telemetry data, see:

sim_array() to simulate an acoustic array;

sim_path_walk() to simulate a movement path from a walk model, with the help of:

rtruncgamma() and rlen() to simulate step lengths;

rwn(), rangrw() and rangcrw() to simulate turning angles;

sim_detections() to simulate detections at receivers, with the help of:

calc_detection_pr_logistic() and calc_detection_pr(), which represent example detection probability models;

To evaluate model skill in reconstructing simulated patterns, see skill_() functions, specifically:

skill_mb() to calculate mean bias;

skill_me() to calculate mean error;

skill_rmse() to calculate root mean squared error;

skill_R() to calculate Spearman's rank correlation coefficient;

skill_d() to calculate the index of agreement;

Data preparation

For help with data acquisition, processing, checking and preliminary analyses, see the flapper package. This facilitates:

Simulation;

Data preparation;

Spatial operations;

Distance calculations;

Movement analyses;

Please submit a feature request if you would like functions from flapper in patter.

Modelling workflow

The main thrust of patter is the provision of an integrated modelling workflow for reconstructing animal movement paths and emergent patterns of space use in passive acoustic telemetry systems.

To implement the centre-of-activity algorithm use:

coa() to calculate centres of activity;

To implement the acoustic-centroid* (AC*) algorithm(s) (e.g., AC and ACDC), use:

acs_setup_obs() to set up observations;

acs_setup_detection_overlaps() to set up detection overlaps;

acs_setup_detection_pr() to define a detection probability model;

acs_setup_detection_kernels() to set up detection kernels;

To implement particle filtering (PF), use:

pf_files() to set up files;

pf_forward() to implement the forward simulation;

pf_backward_*() functions implement the backward pass;

pf_coords() to collate particle coordinates;

pf_path() (and pf_path_pivot()) to reconstruct movement paths;

pf_map_pou() to map probability-of-use;

pf_map_dens() to create smooth maps;

For home ranges, use:

get_hr_prop() for a custom range;

get_hr_core() for the 'core' range;

get_hr_home() for the 'home' range;

get_hr_full() for the full range;

Miscellaneous helpers

The following convenience functions are also made available to users of patter:

Operational statistics. See:

make_matrix_receivers() to matricise receiver deployment time series;

Distance calculations. See:

dist_along_path() to calculate distances along a movement path;

degrees() to create circular angles;

terra helpers. See:

spatTemplate() to create a template SpatRaster;

spatNormalise() to normalise a SpatRaster;

as.im.SpatRaster and as.owin.SpatRaster for interfaces to spatstat function(s);

Parallelisation. See:

cl_lapply() and associated cl_*() functions for parallelisation routines;

References

Lavender, E. et al. (2023). An integrative modelling framework for passive acoustic telemetry. Methods in Ecology and Evolution. https://doi.org/10.1111/2041-210X.14193.

See also

Author

Edward Lavender (ORCID)