PF: pre-calculate densities for the backward pass

This function is used to pre-calculate densities for the backward pass.

pf_backward_dens(
  .history,
  .dens_step,
  ...,
  .in_memory = TRUE,
  .collect = 1e+09,
  .store = NULL,
  .cl = NULL,
  .varlist = NULL,
  .verbose = TRUE,
  .txt = ""
)

Arguments

.history

Particle samples from the forward simulation.

If .in_memory = TRUE, this can be provided as:

A list of data.tables that define cell samples; i.e., the history element of a pf object.

An ordered list of file paths (from pf_setup_files()) that define the directories in which particle samples were written from the forward simulation (as parquet files).

If .in_memory = FALSE, this must be a character string that defines the directory within which parquet files are stored.

.dens_step, ...

A function, and associated arguments, to calculate the density of movements between locations (see pf_backward() and dstep()). This must accept the arguments described in pf_backward():

.data_now---a data.table of particle locations;

.data_past---a data.table of paired particle locations;

(optional) ...---additional arguments;

However, note that the exact form of .data_now and .data_past differ depending on the implementation (see Details):

For implementation options (1) and (2A), densities are calculated between all cell pairs (i.e., two multi-row data.tables), by default via dstep(). Internally, this uses terra::distance() so you must specify lonlat and pairwise = TRUE.

For implementation option (2B) .data_now is a single row (as described in pf_backward()). Under the default implementation, lonlat is required but pairwise is optional.

For custom functions, note that coordinates are stored in each data.table in x_now and y_now columns. Other columns are dropped, with some exceptions.

.in_memory

A logical variable that defines whether or not to begin processing in memory (see Details).

.collect

If .in_memory = FALSE, .collect is an integer that controls whether or not particle pairs are brought into memory for density calculations (see Details). If the number of particle pairs exceeds .collect, processing continues without bringing the data fully into memory.

.store

If .in_memory = FALSE and .collect is exceeded, .store is required. This is a character string that defines the path to an empty directory within which to write files.

.cl, .varlist

Parallelisation options for implementation 2B (see Details and cl_lapply()).

.verbose, .txt

Arguments to monitor function progress (see pf_forward_1()).

Value

The function returns:

Implementations 1 and 2A: a named, nested list

Implementation 2B: invisible(NULL) (see Details)

Details

The backward pass sweeps backwards in time through particle samples to reconstruct movement paths (see pf_backward()). For each particle, at each time step this algorithm samples a preceding location according to the probability densities connecting that particle to all particles from the preceding time step. The purpose of this function is to identify the set of unique transitions between particle pairs and pre-calculate probability densities for pf_backward().

There are five steps:

Identify cells. The first step is to identify set of unique locations (grid cells) sampled (by the forward filter) at each time step.

Identify transitions. The second step is to reconstruct the set of possible transitions between cells at each time step. This requires a cross-join operation (at each time step, each (unique) cell is paired against all (unique) cells from the previous time step). This is memory intensive and can be optionally implemented using sparklyr (see below).

Identify distinct transitions. The third step is to identify the set of unique cell combinations for which probability densities are required. This is hopefully considerably fewer less than the total number of transitions.

Calculate densities. The fourth step is to calculate densities (e.g., via dstep()).

Return outputs. The final step is to return densities in a format that can be efficiently accessed in pf_backward() (e.g., via dstep_lookup() or dstep_read()).

There are three implementation options:

.in_memory is TRUE. Under this option, steps 1--5 are implemented in memory. A list is returned that can be accessed in pf_backward() via dstep_lookup().

.in_memory is FALSE. Under this option, steps 1--5 are implemented without bringing the data (fully) into memory. Steps 1--3 are achieved using sparklyr. The function establishes a connection to Spark, implements 1--3 using Spark, and counts the resultant number of unique transitions for which densities are required (n). What happens next depends on whether or not n <= .collect:

(A) n <= .collect. If n < .collect, data are brought into memory for steps 4--5. A list is returned that can be accessed in pf_backward() via dstep_lookup().

(B) n > .collect. If n > .collect, steps 4--5 are implemented in batches:

For each grid cell, a parquet file is written to {.store}/pairs/{cell}.parquet with the coordinates of cell pairs;

Parquet files are iteratively read into memory, density calculations are implemented and, for each grid cell, a list of densities to paired cells is written to {.store}/density/{cell}.qs. This step can be optionally parallelised. Limited experimentation suggested that this approach is faster than a full Spark implementation. The resultant files can be accessed in pf_backward() via dstep_read(). The function itself returns invisible(NULL).

See also

Author

Edward Lavender

Examples

# See ?`pf_backward()` for examples