PF: backward killer

This function implements backward pruning (killing) of particle samples.

pf_backward_killer(
  .history,
  .record = pf_opt_record(),
  .verbose = getOption("patter.verbose")
)

Arguments

.history

Particle samples from the forward simulation, provided in any format accepted by .pf_history_list(). Particle samples must contain cell_past and cell_now columns.

.record

A named list of output options, from pf_opt_record().

.verbose

User output control (see patter-progress for supported options).

Value

The function returns a pf_particles object.

Details

This function removes 'dead ends' from particle samples. If you imagine the forward simulation, as implemented by pf_forward(), as a branching process, like the growth of a fungal network between two points in space (representing the start and end of the time series), dead ends are the side branches that emerge during this process that do not reach the destination (because sooner-or-later they are rendered incompatible with the data). This function prunes dead-ends from the time series by running a fast matching process backwards in time and retaining the subset of particle samples that lead to the destination particle samples. This process is very fast, and you can use the results to reconstruct movement paths (via pf_path()) and maps of space use (via map_*() functions), but crude.

There are two, related limitations with the 'prune' methodology. The first is that the removal of dead ends tends to bias particle samples, because early samples (which invariably sooner-or-later end up on a dead-end) are more likely to get killed than later samples. This is known as particle degeneracy. Use the pf_backward_killer_diagnostics() function to evaluate trends in the effective sample size through time and examine whether this is an issue. The second is that while particles from the forward simulation are contingent upon the past (a marginal distribution), they do not embody information from the future (the joint distribution).

To reconstruct the joint distribution of particle samples given all data (i.e., 'proper' movement trajectories), the backward sampler is required instead (pf_backward_sampler()). However, this is much more expensive.

This function replaces functionality in flapper::pf_simplify().

See also

Author

Edward Lavender

Examples

#### Set up examples
# Run the forward simulation (see ?`pf_forward()`)
# We will use example output objects:
out_pff    <- dat_pff()
pff_folder <- dat_pff_src(.folder = "history")

#### Example (1): Implement backward pass from object in memory
out_pfb <- pf_backward_killer(.history = out_pff$history,
                              .record = pf_opt_record(.save = TRUE))
#> `patter::pf_backward_killer()` called @ 2024-01-19 11:42:30... 
#> ... Time step 25: 
#> ... ... Recording (cleaned) outputs for `.history[[25]]`... 
#> ... Time step 24: 
#> ... ... Cleaning `.history[[23]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 100 rows in `.history[[t]]`... 
#> ... ... ... Filtering 12 dead ends (88 remain)... 
#> ... ... ... Output: 88 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[24]]`... 
#> ... Time step 23: 
#> ... ... Cleaning `.history[[22]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 88 rows in `.history[[t]]`... 
#> ... ... ... Filtering 22 dead ends (78 remain)... 
#> ... ... ... Output: 78 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[23]]`... 
#> ... Time step 22: 
#> ... ... Cleaning `.history[[21]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 78 rows in `.history[[t]]`... 
#> ... ... ... Filtering 21 dead ends (79 remain)... 
#> ... ... ... Output: 79 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[22]]`... 
#> ... Time step 21: 
#> ... ... Cleaning `.history[[20]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 79 rows in `.history[[t]]`... 
#> ... ... ... Filtering 56 dead ends (44 remain)... 
#> ... ... ... Output: 44 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[21]]`... 
#> ... Time step 20: 
#> ... ... Cleaning `.history[[19]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 44 rows in `.history[[t]]`... 
#> ... ... ... Filtering 41 dead ends (59 remain)... 
#> ... ... ... Output: 59 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[20]]`... 
#> ... Time step 19: 
#> ... ... Cleaning `.history[[18]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 59 rows in `.history[[t]]`... 
#> ... ... ... Filtering 50 dead ends (50 remain)... 
#> ... ... ... Output: 50 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[19]]`... 
#> ... Time step 18: 
#> ... ... Cleaning `.history[[17]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 50 rows in `.history[[t]]`... 
#> ... ... ... Filtering 38 dead ends (62 remain)... 
#> ... ... ... Output: 62 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[18]]`... 
#> ... Time step 17: 
#> ... ... Cleaning `.history[[16]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 62 rows in `.history[[t]]`... 
#> ... ... ... Filtering 30 dead ends (70 remain)... 
#> ... ... ... Output: 70 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[17]]`... 
#> ... Time step 16: 
#> ... ... Cleaning `.history[[15]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 70 rows in `.history[[t]]`... 
#> ... ... ... Filtering 17 dead ends (83 remain)... 
#> ... ... ... Output: 83 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[16]]`... 
#> ... Time step 15: 
#> ... ... Cleaning `.history[[14]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 83 rows in `.history[[t]]`... 
#> ... ... ... Filtering 17 dead ends (83 remain)... 
#> ... ... ... Output: 83 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[15]]`... 
#> ... Time step 14: 
#> ... ... Cleaning `.history[[13]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 83 rows in `.history[[t]]`... 
#> ... ... ... Filtering 28 dead ends (72 remain)... 
#> ... ... ... Output: 72 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[14]]`... 
#> ... Time step 13: 
#> ... ... Cleaning `.history[[12]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 72 rows in `.history[[t]]`... 
#> ... ... ... Filtering 27 dead ends (73 remain)... 
#> ... ... ... Output: 73 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[13]]`... 
#> ... Time step 12: 
#> ... ... Cleaning `.history[[11]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 73 rows in `.history[[t]]`... 
#> ... ... ... Filtering 26 dead ends (74 remain)... 
#> ... ... ... Output: 74 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[12]]`... 
#> ... Time step 11: 
#> ... ... Cleaning `.history[[10]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 74 rows in `.history[[t]]`... 
#> ... ... ... Filtering 41 dead ends (59 remain)... 
#> ... ... ... Output: 59 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[11]]`... 
#> ... Time step 10: 
#> ... ... Cleaning `.history[[9]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 59 rows in `.history[[t]]`... 
#> ... ... ... Filtering 39 dead ends (61 remain)... 
#> ... ... ... Output: 61 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[10]]`... 
#> ... Time step 9: 
#> ... ... Cleaning `.history[[8]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 61 rows in `.history[[t]]`... 
#> ... ... ... Filtering 44 dead ends (56 remain)... 
#> ... ... ... Output: 56 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[9]]`... 
#> ... Time step 8: 
#> ... ... Cleaning `.history[[7]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 56 rows in `.history[[t]]`... 
#> ... ... ... Filtering 53 dead ends (47 remain)... 
#> ... ... ... Output: 47 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[8]]`... 
#> ... Time step 7: 
#> ... ... Cleaning `.history[[6]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 47 rows in `.history[[t]]`... 
#> ... ... ... Filtering 40 dead ends (60 remain)... 
#> ... ... ... Output: 60 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[7]]`... 
#> ... Time step 6: 
#> ... ... Cleaning `.history[[5]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 60 rows in `.history[[t]]`... 
#> ... ... ... Filtering 41 dead ends (59 remain)... 
#> ... ... ... Output: 59 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[6]]`... 
#> ... Time step 5: 
#> ... ... Cleaning `.history[[4]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 59 rows in `.history[[t]]`... 
#> ... ... ... Filtering 44 dead ends (56 remain)... 
#> ... ... ... Output: 56 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[5]]`... 
#> ... Time step 4: 
#> ... ... Cleaning `.history[[3]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 56 rows in `.history[[t]]`... 
#> ... ... ... Filtering 40 dead ends (60 remain)... 
#> ... ... ... Output: 60 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[4]]`... 
#> ... Time step 3: 
#> ... ... Cleaning `.history[[2]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 60 rows in `.history[[t]]`... 
#> ... ... ... Filtering 35 dead ends (65 remain)... 
#> ... ... ... Output: 65 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[3]]`... 
#> ... Time step 2: 
#> ... ... Cleaning `.history[[1]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 65 rows in `.history[[t]]`... 
#> ... ... ... Filtering 51 dead ends (49 remain)... 
#> ... ... ... Output: 49 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[2]]`... 
#> ... Time step 1: 
#> ... ... Recording (cleaned) outputs for `.history[[1]]`... 
#> `patter::pf_backward_killer()` call ended @ 2024-01-19 11:42:30 (duration: ~0 sec(s)). 

#### Example (2): Implement backward pass from parquet files
out_pfb <- pf_backward_killer(.history = pf_files(pff_folder),
                              .record = pf_opt_record(.save = TRUE))
#> `patter::pf_backward_killer()` called @ 2024-01-19 11:42:30... 
#> ... Time step 25: 
#> ... ... Reading `.history[[25]]`... 
#> ... ... Reading `.history[[24]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[25]]`... 
#> ... Time step 24: 
#> ... ... Reading `.history[[23]]`... 
#> ... ... Cleaning `.history[[23]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[24]]`... 
#> ... Time step 23: 
#> ... ... Reading `.history[[22]]`... 
#> ... ... Cleaning `.history[[22]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[23]]`... 
#> ... Time step 22: 
#> ... ... Reading `.history[[21]]`... 
#> ... ... Cleaning `.history[[21]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[22]]`... 
#> ... Time step 21: 
#> ... ... Reading `.history[[20]]`... 
#> ... ... Cleaning `.history[[20]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[21]]`... 
#> ... Time step 20: 
#> ... ... Reading `.history[[19]]`... 
#> ... ... Cleaning `.history[[19]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[20]]`... 
#> ... Time step 19: 
#> ... ... Reading `.history[[18]]`... 
#> ... ... Cleaning `.history[[18]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[19]]`... 
#> ... Time step 18: 
#> ... ... Reading `.history[[17]]`... 
#> ... ... Cleaning `.history[[17]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[18]]`... 
#> ... Time step 17: 
#> ... ... Reading `.history[[16]]`... 
#> ... ... Cleaning `.history[[16]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[17]]`... 
#> ... Time step 16: 
#> ... ... Reading `.history[[15]]`... 
#> ... ... Cleaning `.history[[15]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[16]]`... 
#> ... Time step 15: 
#> ... ... Reading `.history[[14]]`... 
#> ... ... Cleaning `.history[[14]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[15]]`... 
#> ... Time step 14: 
#> ... ... Reading `.history[[13]]`... 
#> ... ... Cleaning `.history[[13]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[14]]`... 
#> ... Time step 13: 
#> ... ... Reading `.history[[12]]`... 
#> ... ... Cleaning `.history[[12]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[13]]`... 
#> ... Time step 12: 
#> ... ... Reading `.history[[11]]`... 
#> ... ... Cleaning `.history[[11]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[12]]`... 
#> ... Time step 11: 
#> ... ... Reading `.history[[10]]`... 
#> ... ... Cleaning `.history[[10]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[11]]`... 
#> ... Time step 10: 
#> ... ... Reading `.history[[9]]`... 
#> ... ... Cleaning `.history[[9]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[10]]`... 
#> ... Time step 9: 
#> ... ... Reading `.history[[8]]`... 
#> ... ... Cleaning `.history[[8]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[9]]`... 
#> ... Time step 8: 
#> ... ... Reading `.history[[7]]`... 
#> ... ... Cleaning `.history[[7]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[8]]`... 
#> ... Time step 7: 
#> ... ... Reading `.history[[6]]`... 
#> ... ... Cleaning `.history[[6]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[7]]`... 
#> ... Time step 6: 
#> ... ... Reading `.history[[5]]`... 
#> ... ... Cleaning `.history[[5]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[6]]`... 
#> ... Time step 5: 
#> ... ... Reading `.history[[4]]`... 
#> ... ... Cleaning `.history[[4]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[5]]`... 
#> ... Time step 4: 
#> ... ... Reading `.history[[3]]`... 
#> ... ... Cleaning `.history[[3]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[4]]`... 
#> ... Time step 3: 
#> ... ... Reading `.history[[2]]`... 
#> ... ... Cleaning `.history[[2]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[3]]`... 
#> ... Time step 2: 
#> ... ... Reading `.history[[1]]`... 
#> ... ... Cleaning `.history[[1]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[2]]`... 
#> ... Time step 1: 
#> ... ... Recording (cleaned) outputs for `.history[[1]]`... 
#> `patter::pf_backward_killer()` call ended @ 2024-01-19 11:42:31 (duration: ~1 sec(s)). 

#### Example (3): Write history to file (as in `pf_forward()`)
con        <- file.path(tempdir(), "patter")
pfb_folder <- file.path(con, "backward", "killer")
dir.create(pfb_folder, recursive = TRUE)
out_pfb <- pf_backward_killer(.history = pf_files(pff_folder),
                              .record = pf_opt_record(.sink = pfb_folder))
#> `patter::pf_backward_killer()` called @ 2024-01-19 11:42:31... 
#> ... Time step 25: 
#> ... ... Reading `.history[[25]]`... 
#> ... ... Reading `.history[[24]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[25]]`... 
#> ... Time step 24: 
#> ... ... Reading `.history[[23]]`... 
#> ... ... Cleaning `.history[[23]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[24]]`... 
#> ... Time step 23: 
#> ... ... Reading `.history[[22]]`... 
#> ... ... Cleaning `.history[[22]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[23]]`... 
#> ... Time step 22: 
#> ... ... Reading `.history[[21]]`... 
#> ... ... Cleaning `.history[[21]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[22]]`... 
#> ... Time step 21: 
#> ... ... Reading `.history[[20]]`... 
#> ... ... Cleaning `.history[[20]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[21]]`... 
#> ... Time step 20: 
#> ... ... Reading `.history[[19]]`... 
#> ... ... Cleaning `.history[[19]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[20]]`... 
#> ... Time step 19: 
#> ... ... Reading `.history[[18]]`... 
#> ... ... Cleaning `.history[[18]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[19]]`... 
#> ... Time step 18: 
#> ... ... Reading `.history[[17]]`... 
#> ... ... Cleaning `.history[[17]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[18]]`... 
#> ... Time step 17: 
#> ... ... Reading `.history[[16]]`... 
#> ... ... Cleaning `.history[[16]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[17]]`... 
#> ... Time step 16: 
#> ... ... Reading `.history[[15]]`... 
#> ... ... Cleaning `.history[[15]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[16]]`... 
#> ... Time step 15: 
#> ... ... Reading `.history[[14]]`... 
#> ... ... Cleaning `.history[[14]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[15]]`... 
#> ... Time step 14: 
#> ... ... Reading `.history[[13]]`... 
#> ... ... Cleaning `.history[[13]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[14]]`... 
#> ... Time step 13: 
#> ... ... Reading `.history[[12]]`... 
#> ... ... Cleaning `.history[[12]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[13]]`... 
#> ... Time step 12: 
#> ... ... Reading `.history[[11]]`... 
#> ... ... Cleaning `.history[[11]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[12]]`... 
#> ... Time step 11: 
#> ... ... Reading `.history[[10]]`... 
#> ... ... Cleaning `.history[[10]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[11]]`... 
#> ... Time step 10: 
#> ... ... Reading `.history[[9]]`... 
#> ... ... Cleaning `.history[[9]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[10]]`... 
#> ... Time step 9: 
#> ... ... Reading `.history[[8]]`... 
#> ... ... Cleaning `.history[[8]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[9]]`... 
#> ... Time step 8: 
#> ... ... Reading `.history[[7]]`... 
#> ... ... Cleaning `.history[[7]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[8]]`... 
#> ... Time step 7: 
#> ... ... Reading `.history[[6]]`... 
#> ... ... Cleaning `.history[[6]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[7]]`... 
#> ... Time step 6: 
#> ... ... Reading `.history[[5]]`... 
#> ... ... Cleaning `.history[[5]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[6]]`... 
#> ... Time step 5: 
#> ... ... Reading `.history[[4]]`... 
#> ... ... Cleaning `.history[[4]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[5]]`... 
#> ... Time step 4: 
#> ... ... Reading `.history[[3]]`... 
#> ... ... Cleaning `.history[[3]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[4]]`... 
#> ... Time step 3: 
#> ... ... Reading `.history[[2]]`... 
#> ... ... Cleaning `.history[[2]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[3]]`... 
#> ... Time step 2: 
#> ... ... Reading `.history[[1]]`... 
#> ... ... Cleaning `.history[[1]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[2]]`... 
#> ... Time step 1: 
#> ... ... Recording (cleaned) outputs for `.history[[1]]`... 
#> `patter::pf_backward_killer()` call ended @ 2024-01-19 11:42:31 (duration: ~0 sec(s)). 
head(list.files(pfb_folder))
#> character(0)

#### Example (4): Adjust standard `patter-progress` options
# Use a log.txt file
log.txt <- tempfile(fileext = ".txt")
out_pfb <- pf_backward_killer(pf_files(pff_folder),
                              .record = pf_opt_record(.save = TRUE),
                              .verbose = log.txt)
head(readLines(log.txt), 15)
#>  [1] "`patter::pf_backward_killer()` called @ 2024-01-19 11:42:31... "                                         
#>  [2] "... Time step 25: "                                                                                      
#>  [3] "... ... Reading `.history[[25]]`... "                                                                    
#>  [4] "... ... Reading `.history[[24]]`... "                                                                    
#>  [5] "... ... Recording (cleaned) outputs for `.history[[25]]`... "                                            
#>  [6] "... Time step 24: "                                                                                      
#>  [7] "... ... Reading `.history[[23]]`... "                                                                    
#>  [8] "... ... Cleaning `.history[[23]]`... "                                                                   
#>  [9] "... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... "
#> [10] "... ... ... Input: 0 rows in `.history[[t]]`... "                                                        
#> [11] "... ... Recording (cleaned) outputs for `.history[[24]]`... "                                            
#> [12] "... Time step 23: "                                                                                      
#> [13] "... ... Reading `.history[[22]]`... "                                                                    
#> [14] "... ... Cleaning `.history[[22]]`... "                                                                   
#> [15] "... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... "
unlink(log.txt)
# Suppress `.verbose`
out_pfb <- pf_backward_killer(pf_files(pff_folder),
                              .record = pf_opt_record(.save = TRUE),
                              .verbose = FALSE)
# Suppress progress bar
pbo <- pbapply::pboptions(type = "n")
out_pfb <- pf_backward_killer(pf_files(pff_folder),
                              .record = pf_opt_record(.save = TRUE))
#> `patter::pf_backward_killer()` called @ 2024-01-19 11:42:31... 
#> ... Time step 25: 
#> ... ... Reading `.history[[25]]`... 
#> ... ... Reading `.history[[24]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[25]]`... 
#> ... Time step 24: 
#> ... ... Reading `.history[[23]]`... 
#> ... ... Cleaning `.history[[23]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[24]]`... 
#> ... Time step 23: 
#> ... ... Reading `.history[[22]]`... 
#> ... ... Cleaning `.history[[22]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[23]]`... 
#> ... Time step 22: 
#> ... ... Reading `.history[[21]]`... 
#> ... ... Cleaning `.history[[21]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[22]]`... 
#> ... Time step 21: 
#> ... ... Reading `.history[[20]]`... 
#> ... ... Cleaning `.history[[20]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[21]]`... 
#> ... Time step 20: 
#> ... ... Reading `.history[[19]]`... 
#> ... ... Cleaning `.history[[19]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[20]]`... 
#> ... Time step 19: 
#> ... ... Reading `.history[[18]]`... 
#> ... ... Cleaning `.history[[18]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[19]]`... 
#> ... Time step 18: 
#> ... ... Reading `.history[[17]]`... 
#> ... ... Cleaning `.history[[17]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[18]]`... 
#> ... Time step 17: 
#> ... ... Reading `.history[[16]]`... 
#> ... ... Cleaning `.history[[16]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[17]]`... 
#> ... Time step 16: 
#> ... ... Reading `.history[[15]]`... 
#> ... ... Cleaning `.history[[15]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[16]]`... 
#> ... Time step 15: 
#> ... ... Reading `.history[[14]]`... 
#> ... ... Cleaning `.history[[14]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[15]]`... 
#> ... Time step 14: 
#> ... ... Reading `.history[[13]]`... 
#> ... ... Cleaning `.history[[13]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[14]]`... 
#> ... Time step 13: 
#> ... ... Reading `.history[[12]]`... 
#> ... ... Cleaning `.history[[12]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[13]]`... 
#> ... Time step 12: 
#> ... ... Reading `.history[[11]]`... 
#> ... ... Cleaning `.history[[11]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[12]]`... 
#> ... Time step 11: 
#> ... ... Reading `.history[[10]]`... 
#> ... ... Cleaning `.history[[10]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[11]]`... 
#> ... Time step 10: 
#> ... ... Reading `.history[[9]]`... 
#> ... ... Cleaning `.history[[9]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[10]]`... 
#> ... Time step 9: 
#> ... ... Reading `.history[[8]]`... 
#> ... ... Cleaning `.history[[8]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[9]]`... 
#> ... Time step 8: 
#> ... ... Reading `.history[[7]]`... 
#> ... ... Cleaning `.history[[7]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[8]]`... 
#> ... Time step 7: 
#> ... ... Reading `.history[[6]]`... 
#> ... ... Cleaning `.history[[6]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[7]]`... 
#> ... Time step 6: 
#> ... ... Reading `.history[[5]]`... 
#> ... ... Cleaning `.history[[5]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[6]]`... 
#> ... Time step 5: 
#> ... ... Reading `.history[[4]]`... 
#> ... ... Cleaning `.history[[4]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[5]]`... 
#> ... Time step 4: 
#> ... ... Reading `.history[[3]]`... 
#> ... ... Cleaning `.history[[3]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[4]]`... 
#> ... Time step 3: 
#> ... ... Reading `.history[[2]]`... 
#> ... ... Cleaning `.history[[2]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[3]]`... 
#> ... Time step 2: 
#> ... ... Reading `.history[[1]]`... 
#> ... ... Cleaning `.history[[1]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 0 rows in `.history[[t]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[2]]`... 
#> ... Time step 1: 
#> ... ... Recording (cleaned) outputs for `.history[[1]]`... 
#> `patter::pf_backward_killer()` call ended @ 2024-01-19 11:42:31 (duration: ~0 sec(s)). 

# Clean up
unlink(con, recursive = TRUE)