PF: particle coordinates

This function collects particle samples and extracts coordinates.

pf_coord(.history, .bathy, .obs = NULL, .cols = NULL)

Arguments

.history

Particle samples, provided in any format accepted by .pf_history_dt(). Particle samples may be sourced from:

pf_forward() (a marginal distribution);

pf_backward_killer() (a 'partial' joint distribution);

pf_backward_sampler_v() (the full joint distribution);

Particle samples must contain the timestep and cell_now columns.

.bathy

The bathymetry SpatRaster.

.obs, .cols

(optional) A data.table and a character vector of column names in .obs to match onto the output. .obs must contain a timestep column for matching.

Value

The function returns a data.table that defines time steps, sampled locations and, optionally, information from .obs, with the following columns:

timestep---an integer vector that defines the time step;

cell_id---an integer vector that defines the cell ID on .bathy;

cell_x, cell_y, cell_z---doubles that define cell coordinates;

If .obs is supplied, the output also contains any columns specified in .cols.

Details

The cell_now column is used to extract coordinates on .bathy.

This function is not memory safe, since the entire time series of coordinates is collected in memory.

See also

Author

Edward Lavender

Examples

#### Example (1): Use particle samples in memory or on file
# Particles can be provided in any format accepted by `?.pf_history_dt()`
# Here, we use precomputed samples:
gebco <- dat_gebco()
pf_coord(dat_pff(), .bathy = gebco)
#>       timestep cell_id   cell_x  cell_y    cell_z
#>          <int>   <int>    <num>   <num>     <num>
#>    1:        1   36618 709242.1 6253807 112.27689
#>    2:        1   36807 709142.1 6253707 114.06328
#>    3:        1   36810 709442.1 6253707  76.37878
#>    4:        1   36995 708942.1 6253607 126.04160
#>    5:        1   36997 709142.1 6253607 105.22495
#>   ---                                            
#> 2496:       25   38713 709742.1 6252707  36.09760
#> 2497:       25   39280 709442.1 6252407  37.93974
#> 2498:       25   39280 709442.1 6252407  37.93974
#> 2499:       25   38903 709742.1 6252607  36.44192
#> 2500:       25   38899 709342.1 6252607  41.63695
pf_coord(dat_pff()$history, .bathy = gebco)
#>       timestep cell_id   cell_x  cell_y    cell_z
#>          <int>   <int>    <num>   <num>     <num>
#>    1:        1   36618 709242.1 6253807 112.27689
#>    2:        1   36807 709142.1 6253707 114.06328
#>    3:        1   36810 709442.1 6253707  76.37878
#>    4:        1   36995 708942.1 6253607 126.04160
#>    5:        1   36997 709142.1 6253607 105.22495
#>   ---                                            
#> 2496:       25   38713 709742.1 6252707  36.09760
#> 2497:       25   39280 709442.1 6252407  37.93974
#> 2498:       25   39280 709442.1 6252407  37.93974
#> 2499:       25   38903 709742.1 6252607  36.44192
#> 2500:       25   38899 709342.1 6252607  41.63695
pf_coord(dat_pff_src(), .bathy = gebco)
#>       timestep cell_id   cell_x  cell_y    cell_z
#>          <int>   <int>    <num>   <num>     <num>
#>    1:        1   36618 709242.1 6253807 112.27689
#>    2:        1   36807 709142.1 6253707 114.06328
#>    3:        1   36810 709442.1 6253707  76.37878
#>    4:        1   36995 708942.1 6253607 126.04160
#>    5:        1   36997 709142.1 6253607 105.22495
#>   ---                                            
#> 2496:       25   38713 709742.1 6252707  36.09760
#> 2497:       25   39280 709442.1 6252407  37.93974
#> 2498:       25   39280 709442.1 6252407  37.93974
#> 2499:       25   38903 709742.1 6252607  36.44192
#> 2500:       25   38899 709342.1 6252607  41.63695
pf_coord(pf_files(dat_pff_src()), .bathy = gebco)
#>       timestep cell_id   cell_x  cell_y    cell_z
#>          <int>   <int>    <num>   <num>     <num>
#>    1:        1   36618 709242.1 6253807 112.27689
#>    2:        1   36807 709142.1 6253707 114.06328
#>    3:        1   36810 709442.1 6253707  76.37878
#>    4:        1   36995 708942.1 6253607 126.04160
#>    5:        1   36997 709142.1 6253607 105.22495
#>   ---                                            
#> 2496:       25   38713 709742.1 6252707  36.09760
#> 2497:       25   39280 709442.1 6252407  37.93974
#> 2498:       25   39280 709442.1 6252407  37.93974
#> 2499:       25   38903 709742.1 6252607  36.44192
#> 2500:       25   38899 709342.1 6252607  41.63695

#### Example (2): Use particle samples from the forward run or the backward pass
# (TO DO)

#### Example (3): Include additional columns in the output
obs <- dat_obs()
pxy <- pf_coord(dat_pff(), .bathy = gebco,
                .obs = obs, .cols = c("detection", "depth"))

#### Example (4): Use particle coordinates for mapping
# See `map_*()` functions e.g. `?map_dens()`