PF: particle coordinates

This function collects particle samples and extracts coordinates.

pf_coords(.history, .bathy, .obs = NULL, .cols = NULL)

Arguments

.history

Particle samples from the particle filter, provided either as:

A list of data.tables that define cell samples; i.e., the history element of a pf object. This must contain a column that defines accepted cell samples at each time step (cell_now).

An ordered list of file paths (from pf_setup_files()) that define the directories in which particle samples were written from the forward simulation (as parquet files).

.bathy

The bathymetry SpatRaster.

.obs, .cols

(optional) A data.table and a character vector of column names in .obs to match onto the output. .obs must contain a timestep column for matching.

Value

The function returns a data.table that defines timesteps, sampled locations and, optionally, information from .obs, with the following columns:

timestep---an integer vector that defines the time step;

cell_id---an integer vector that defines the cell ID on .bathy;

cell_x, cell_y, cell_z---doubles that define cell coordinates;

If .obs is supplied, the output also contains any columns specified in .cols.

Details

This function is not memory safe.

See also

Author

Edward Lavender

Examples

p <- dat_pfb()
pf_coords(p$history, dat_gebco())
#>        timestep cell_id   cell_x  cell_y    cell_z
#>     1:        1   37376 709042.1 6253407  99.90396
#>     2:        1   37757 709142.1 6253207  73.27666
#>     3:        1   37568 709242.1 6253307  68.53316
#>     4:        1   39089 709342.1 6252507  40.22162
#>     5:        1   37564 708842.1 6253307 113.02628
#>    ---                                            
#> 44408:       50   38702 708642.1 6252707  74.05189
#> 44409:       50   38890 708442.1 6252607  74.70188
#> 44410:       50   38703 708742.1 6252707  68.56325
#> 44411:       50   38135 708942.1 6253007  74.56544
#> 44412:       50   38701 708542.1 6252707  79.41030