PF: count particles

This function counts particle (cell) samples. Its primary use is to process outputs from forward filter for pf_backward_sampler().

pf_count(
  .history,
  .record = pf_opt_record(),
  .verbose = getOption("patter.verbose"),
  .cl = NULL,
  .cl_varlist = NULL,
  .cl_chunk = ifelse(is.null(.cl), FALSE, TRUE)
)

Arguments

.history

Particle samples, provided in any format accepted by .pf_history_list(). Particle samples must contain the cell_now column.

.record

A named list of output options, from pf_opt_record().

.verbose

User output control (see patter-progress for supported options).

.cl, .cl_varlist, .cl_chunk

Parallelisation options, passed to cl_lapply().

Value

The function returns a pf_particles object.

Details

pf_count() iterates over time steps, counts the number of copies of each cell and saves a data.table of counts in memory and/or writes them to file for pf_backward_sampler(). This is designed to improve speed in pf_backward_sampler().

Author

Edward Lavender