PF: particle diagnostics

These functions collate particle diagnostics from pf_forward() and pf_backward_*().

pf_diag_summary(.history, ...)

Arguments

.history

For pf_diag_summary(), .history expects particle samples, provided in any format accepted by .pf_history_dt().

...

Additional arguments.

In pf_diag_convergence(), ... is passed to arrow::open_dataset().

In pf_diag_summary(), ... is passed to .pf_history_dt(). .collect, if used, is necessarily TRUE and should not be specified.

Value

pf_diag_*() functions return data.tables.

pf_diag_convergence() returns the diagnostics

data.table of a pf_particles-class object.

pf_diag_summary returns a summary data.table with the following columns:

timestep---an integer that defines the time step;

n---an integer that defines the number of particles;

nu---an integer that defines the number of unique location samples (see .pf_diag_nu());

ess---a double that defines the effective sample size or NA_real_ if the lik column is unavailable (see .pf_diag_ess());

Details

Particle diagnostics are fully described in pf_diag-internal.

pf_diag_convergence() collates convergence diagnostic outputs from pf_forward(). If pf_forward() is implemented with .record$save = TRUE, the outputted pf_particles object from pf_forward() includes a pre-compiled data.table of diagnostics (see pf_opt_record()). pf_diag_convergence() accepts a pf_particles object, but there is really no need for this, as a diagnostics element is already present in this object. Instead, pf_diag_convergence() is primarily designed to collate particle diagnostics on file. pf_forward() writes particle diagnostics to {.record$sink}/diagnostics (see pf_opt_record()). You can supply {.record$sink} or {.record$sink}/diagnostics/ to this function. The individual data.tables are collated to match the form in which they are provided by pf_forward() when .record$save = TRUE.

pf_diag_summary() summarises particle diagnostics directly from (accepted) particle samples. This can be used for pf_forward() or pf_backward_*().

See also

Examples

#### Example (1): Use particle samples from pf_forward() or pf_backward_*()
pf_diag_summary(.history = dat_pff())
#>     timestep     n    nu      ess
#>        <int> <int> <int>    <num>
#>  1:        1    78    78 69.44444
#>  2:        2    47    47 32.89474
#>  3:        3    42    42 27.47253
#>  4:        4    39    39 26.45503
#>  5:        5    41    41 27.32240
#>  6:        6    45    45 33.33333
#>  7:        7    40    40 28.57143
#>  8:        8    44    44 26.59574
#>  9:        9    45    45 30.30303
#> 10:       10    48    48 31.05590
#> 11:       11    45    45 28.73563
#> 12:       12    43    43 28.40909
#> 13:       13    46    46 28.40909
#> 14:       14    47    47 26.31579
#> 15:       15    36    36 25.00000
#> 16:       16    42    42 26.45503
#> 17:       17    47    47 30.12048
#> 18:       18    44    44 25.51020
#> 19:       19    44    44 30.48780
#> 20:       20    49    49 36.49635
#> 21:       21    37    37 24.87562
#> 22:       22    53    53 34.48276
#> 23:       23    51    51 40.32258
#> 24:       24    56    56 38.46154
#> 25:       25    67    67 52.08333
#>     timestep     n    nu      ess
# (TO REVISE)

#### Example (2): Use particle samples in memory or on file
# Particles can be provided in any format accepted by `?.pf_history_dt()`
d1 <- pf_diag_summary(.history = dat_pff())
d2 <- pf_diag_summary(.history = dat_pff()$history)
d3 <- pf_diag_summary(.history = dat_pff_src())
d4 <- pf_diag_summary(.history = pf_files(dat_pff_src()))
stopifnot(isTRUE(all.equal(d1, d2)))
stopifnot(isTRUE(all.equal(d1, d3)))
stopifnot(isTRUE(all.equal(d1, d4)))

## (C) Examine trends in particle diagnostics
# (TO REVISE)