PF: list files

pf_files() creates an ordered list of the parquet files that contain particle samples (e.g., from pf_forward() or pf_backward_killer()).

pf_files_size() calculates the total size of all files.

pf_files(.sink, ...)

pf_files_size(.sink, .folder = NULL, ..., .unit = c("MB", "GB", "TB"))

Arguments

.sink

A character string that defines the directory in which files are located.

...

A placeholder for additional arguments passed to list.files(), such as pattern, excluding full.names.

.folder

(optional) For pf_files_size(), a character string that defines the name of a sub-folder for which to summarise file sizes.

.unit

For pf_files_size(), .unit is a character string that defines the units of the output (MB, GB, TB).

Value

pf_files() returns an ordered list of file paths.

pf_files_size() returns a number.

Warning

At the time of writing, pf_files() cannot be used to list particle-diagnostic files from pf_forward() and should only be used for particle samples. However, pf_files_size() can be used to estimate the total file size in any directory.

See also

Author

Edward Lavender

Examples

# Use `pf_files()` to list files from `pf_forward()`
pff_folder <- dat_pff_src(.folder = "history")
files <- pf_files(pff_folder)

# Use `pf_files()` to list files from `pf_backward_killer()`
pfbk_folder <- dat_pfbk_src()
files <- pf_files(pfbk_folder)

# Use `pf_files_size()` to calculate file size
pf_files_size(pfbk_folder)
#> [1] 0.102697
pf_files_size(pfbk_folder, .unit = "GB")
#> [1] 0.000102697
pf_files_size(pfbk_folder, .unit = "TB")
#> [1] 1.02697e-07