PF: run the forward simulation

This function simulates the possible locations of an animal forwards in time. This function optionally incorporates AC* dynamics alongside the movement model in the simulation, unlike pf_forward_1() which takes the outputs of an AC* algorithm as the starting point for the forward simulation.

pf_forward_2(
  .obs,
  .origin = NULL,
  .bathy,
  .lonlat = FALSE,
  .moorings = NULL,
  .detection_overlaps = NULL,
  .detection_kernels = NULL,
  .update_ac = NULL,
  .kick,
  ...,
  .n = 100L,
  .save_history = FALSE,
  .write_history = NULL,
  .progress = TRUE,
  .verbose = TRUE,
  .txt = ""
)

Arguments

.obs

A data.table that defines the time series of observations for a selected individual (e.g., from acs_setup_obs()). At a minimum, this must contain the following columns:

timestep--an integer that defines the time step;

Columns required by the AC algorithm, if implemented; i.e.:

date---a character column that defines the date;

detection---an integer column that uniquely defines each detection;

detection_id---an integer column that distinguishes the time steps at which detections were (1) or were not (0) recorded;

receiver_id---a list column that defines the receiver(s) that recorded detections at each time step;

receiver_id_next---a list column that define receiver(s) that recorded the next detection(s);

buffer_future_incl_gamma---a numeric column that defines the maximum distance of the individual from receiver_id_next;

Columns required by .update_ac;

Columns required by .kick;

.origin

(optional) A data.table that defines a set of 'quadrature points' from which the first .n particles are sampled (with replacement, according to their weights). If supplied, .origin must include three columns:

cell_now---an integer vector that defines the cell ID (on .bathy);

x_now---a numeric vector that defines the x coordinates;

y_now---a numeric vector that defines y coordinates;

If un-supplied, quadrature points are sampled from within acoustic containers using .bathy (if applicable) or from .bathy at large (see Details).

.bathy

A SpatRaster over the region of interest. This is used to:

Coerce receiver coordinates onto the grid, if necessary (see .moorings);

Define quadrature points, if necessary (see .origin);

Filter particle samples from inhospitable habitats (define by NA), if necessary;

(optional) Define or update particle weights (via .update_ac);

(optional) Simulate movement (via .kick);

Coerce simulated positions onto the grid;

.lonlat

A logical variable that defines whether or not spatial inputs (namely .bathy and .moorings coordinates) are defined in terms of longitude/latitude or a planar coordinate reference system.

.moorings, .detection_overlaps, .detection_kernels

(optional) AC* arguments. These arguments are used to implement the AC* algorithm on the fly.

moorings is a data.table that defines receiver deployments. This must contain the following columns:

receiver_id---an integer vector of receiver IDs;

receiver_easting and receiver_northing (if .lonlat = FALSE) or receiver_lon and receiver_lat (if.lonlat = FALSE) ---numeric vectors that define receiver coordinates;

detection_overlaps is a list of detection container overlaps (see acs()).

detection_kernels is a named list of detection kernels (see acs()).

.moorings, .detection_overlaps and .detection_gaps can be NULL if the AC algorithm is not implemented.

.update_ac

A list of function(s) used to calculate, or update, particle weights. Each function in the list must accept four named arguments, even if unused:

.particles---a data.table that defines particle locations, with columns cell_now, x_now and y_now;

.obs---the .obs data.table;

.t---an integer that defines the time step (used to index .obs);

.bathy---the .bathy SpatRaster

The function must return a numeric vector of weights (one for each particle). The weights from the AC* algorithm (if applicable) and the functions in .update_ac are combined internally and used to (re-)sample particles.

.kick, ...

A function, and associated inputs, used to 'kick' particles into new (proposal) locations (see pf_forward_1()).

.n

An integer that defines the number of particle samples at each time step.

.save_history, .write_history

Arguments to save particle samples in memory or to file (see pf_forward_1()).

.progress, .verbose, .txt

Controls of function prompts and messages (see pf_forward_1()).

Value

The function returns a pf object.

Details

The forward simulation is implemented as follows:

At each time step, at quadrature points (t = 1) or proposal locations (t > 1), weights are calculated according to the AC* algorithm (if specified) and/or models for ancillary data as specified in .update_ac. Unlike pf_forward_1(), weights are calculated at particle locations rather than across .bathy as a whole. Weights are calculated in four steps:

Land filter. Quadrature points/particles on land (in NA cells on .bathy) are dropped.

(optional) Container filter. Particle proposals that are incompatible with AC dynamics are dropped.

(optional) AC weights. For the set of remaining particles, AC weights are calculated (using .detection_overlaps and .detection_kernels).

Updated weights. Weights are calculated, or updated, based on additional models (e.g., to account for depth time series).

.n locations (particles) are sampled from the quadrature points (t = 1) or resampled from a set of proposals (t > 1), with replacement, in line with calculated weights;

The previous locations (NA for t = 1) and the (accepted) current locations are recorded;

Each particle is 'kicked' into a new (proposal) location by the movement model;

Steps 1-4 are repeated until the end of the time series.

The essential difference between pf_forward_1() and pf_forward_2() is that former is part of a coupled workflow in which (a) an AC* algorithm is used to calculate weights across the entire grid and (b) the weights are provided as an argument to pf_forward_1() which simulates movement trajectories, whereas the is an integrated implementation in which weights are calculated on the fly. In both cases, AC* weights are defined based on pre-computed SpatRasters (namely .detection_kernels) and movements are represented across the same grid.

See also

Author

Edward Lavender

Examples

#### Set up examples
# Use pre-prepared datasets (see also `?acs()`)
# * Note that the default grid is low resolution
# * The grid should be resampled for real analyses to higher resolution
moorings  <- dat_moorings
obs       <- dat_obs()
gebco     <- dat_gebco()
overlaps  <- dat_overlaps()
kernels   <- dat_kernels()

#### Example (1): Implement the ACPF algorithm using default options
# * AC algorithm is implemented on the fly using overlaps & kernels
# * These pre-prepared datasets assume a particular detection probability model
# * PF is implemented using the default movement model
out_pff <- pf_forward_2(obs,
                        .bathy = gebco,
                        .moorings = moorings,
                        .detection_overlaps = overlaps,
                        .detection_kernels = kernels,
                        .kick = pf_kick,
                        .save_history = TRUE)
#> patter::pf_forward_2() called (@ 2023-11-24 20:46:45.640059)... 
#> ... Setting up simulation... 
#> ... ... Defining filter(s)... 
#> Warning: `.moorings` coordinates coerced onto `.bathy` grid.
#> ... ... Defining time indicators... 
#> ... ... Initiating simulation... 
#> 
#> ... Time step 1: 
#> ... ... Defining starting surface (quadrature points)... 
#> ... ... ... Building acoustic container(s)... 
#> ... ... ... Sampling quadature points... 
#> ... ... Calculating weights... 
#> ... ... Detection 1: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 2: 
#> ... ... Calculating weights... 
#> ... ... Detection 2: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 3: 
#> ... ... Calculating weights... 
#> ... ... Detection 3: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 4: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 5: 
#> ... ... Calculating weights... 
#> ... ... Detection 4: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 6: 
#> ... ... Calculating weights... 
#> ... ... Detection 5: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 7: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 8: 
#> ... ... Calculating weights... 
#> ... ... Detection 6: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 9: 
#> ... ... Calculating weights... 
#> ... ... Detection 7: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 10: 
#> ... ... Calculating weights... 
#> ... ... Detection 8: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 11: 
#> ... ... Calculating weights... 
#> ... ... Detection 9: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 12: 
#> ... ... Calculating weights... 
#> ... ... Detection 10: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 13: 
#> ... ... Calculating weights... 
#> ... ... Detection 11: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 14: 
#> ... ... Calculating weights... 
#> ... ... Detection 12: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 15: 
#> ... ... Calculating weights... 
#> ... ... Detection 13: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 16: 
#> ... ... Calculating weights... 
#> ... ... Detection 14: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 17: 
#> ... ... Calculating weights... 
#> ... ... Detection 15: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 18: 
#> ... ... Calculating weights... 
#> ... ... Detection 16: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 19: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 20: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 21: 
#> ... ... Calculating weights... 
#> ... ... Detection 17: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 22: 
#> ... ... Calculating weights... 
#> ... ... Detection 18: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 23: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 24: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 25: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 26: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 27: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 28: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 29: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 30: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 31: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 32: 
#> ... ... Calculating weights... 
#> ... ... Detection 19: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 33: 
#> ... ... Calculating weights... 
#> ... ... Detection 20: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 34: 
#> ... ... Calculating weights... 
#> ... ... Detection 21: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 35: 
#> ... ... Calculating weights... 
#> ... ... Detection 22: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 36: 
#> ... ... Calculating weights... 
#> ... ... Detection 23: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 37: 
#> ... ... Calculating weights... 
#> ... ... Detection 24: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 38: 
#> ... ... Calculating weights... 
#> ... ... Detection 25: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 39: 
#> ... ... Calculating weights... 
#> ... ... Detection 26: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 40: 
#> ... ... Calculating weights... 
#> ... ... Detection 27: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 41: 
#> ... ... Calculating weights... 
#> ... ... Detection 28: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 42: 
#> ... ... Calculating weights... 
#> ... ... Detection 29: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 43: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 44: 
#> ... ... Calculating weights... 
#> ... ... Detection 30: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 45: 
#> ... ... Calculating weights... 
#> ... ... Detection 31: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 46: 
#> ... ... Calculating weights... 
#> ... ... Detection 32: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 47: 
#> ... ... Calculating weights... 
#> ... ... Detection 33: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 48: 
#> ... ... Calculating weights... 
#> ... ... Detection 34: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 49: 
#> ... ... Calculating weights... 
#> ... ... Detection 35: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 50: 
#> ... ... Calculating weights... 
#> ... ... Detection 36: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> patter::pf_forward_2() call ended (@ 2023-11-24 20:46:46.179255). 
# The function returns a list as `pf_forward_1()`:
summary(out_pff)
#>         Length Class  Mode
#> history 50     -none- list
#> time     3     -none- list

#### Example (2): Implement the DCPF algorithm
# Define depth-error model (see also `?acs()` and ?dc()`)
# * We imagine the individual's depth is known ± 25 m with equal probability
calc_depth_error   <- function(...) matrix(c(-25, 25), nrow = 2)
# Define shallow/deep depth limits
obs$depth_shallow <- obs$depth + calc_depth_error(obs$depth)[1, ]
obs$depth_deep    <- obs$depth + calc_depth_error(obs$depth)[2, ]
# Define function that calculates DC weights
update_ac <- function(.particles, .bathy, .obs, .t, ...) {
  # Extract depth of seabed at particle positions
  .particles$bathy <- terra::extract(.bathy, as.matrix(.particles[, c("x_now", "y_now")]))
  # Weight = 1 in locations where bathymetric depth is within possible limits, otherwise 0
  (.particles$bathy  >= .obs$depth_shallow[.t] & .particles$bathy <= .obs$depth_deep[.t]) + 0
}
# Implement DCPF
out_pff <- pf_forward_2(obs,
                        .bathy = gebco,
                        .kick = pf_kick,
                        .update_ac = update_ac,
                        .save_history = TRUE)
#> patter::pf_forward_2() called (@ 2023-11-24 20:46:46.181395)... 
#> ... Setting up simulation... 
#> ... ... Defining filter(s)... 
#> ... ... Defining time indicators... 
#> ... ... Initiating simulation... 
#> 
#> ... Time step 1: 
#> ... ... Defining starting surface (quadrature points)... 
#> ... ... ... Sampling quadrature points from `.bathy`... 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 2: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 3: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 4: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 5: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 6: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 7: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 8: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 9: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 10: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 11: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 12: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 13: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 14: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 15: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 16: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 17: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 18: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 19: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 20: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 21: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 22: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 23: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 24: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 25: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 26: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 27: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 28: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 29: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 30: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 31: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 32: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 33: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 34: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 35: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 36: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 37: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 38: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 39: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 40: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 41: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 42: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 43: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 44: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 45: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 46: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 47: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 48: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 49: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 50: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> patter::pf_forward_2() call ended (@ 2023-11-24 20:46:46.498653). 
# Extract coordinates & validate implementation of depth-error model
out_pfxy <- pf_coords(out_pff$history, gebco,
                      .obs = obs, .cols = c("depth_shallow", "depth_deep"))
stopifnot(all(out_pfxy$depth >= out_pfxy$depth_shallow - 25 &
                out_pfxy$depth <= out_pfxy$depth_shallow + 25))

#### Example (3): Implement the ACDCPF algorithm
# Note that this algorithm may fail to converge on this low resolution grid
out_pff <- pf_forward_2(obs,
                        .bathy = gebco,
                        .moorings = moorings,
                        .detection_overlaps = overlaps,
                        .detection_kernels = kernels,
                        .update_ac = update_ac,
                        .kick = pf_kick,
                        .save_history = TRUE)
#> patter::pf_forward_2() called (@ 2023-11-24 20:46:46.506069)... 
#> ... Setting up simulation... 
#> ... ... Defining filter(s)... 
#> Warning: `.moorings` coordinates coerced onto `.bathy` grid.
#> ... ... Defining time indicators... 
#> ... ... Initiating simulation... 
#> 
#> ... Time step 1: 
#> ... ... Defining starting surface (quadrature points)... 
#> ... ... ... Building acoustic container(s)... 
#> ... ... ... Sampling quadature points... 
#> ... ... Calculating weights... 
#> ... ... Detection 1: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 2: 
#> ... ... Calculating weights... 
#> ... ... Detection 2: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 3: 
#> ... ... Calculating weights... 
#> ... ... Detection 3: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 4: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 5: 
#> ... ... Calculating weights... 
#> ... ... Detection 4: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 6: 
#> ... ... Calculating weights... 
#> ... ... Detection 5: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 7: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 8: 
#> ... ... Calculating weights... 
#> ... ... Detection 6: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 9: 
#> ... ... Calculating weights... 
#> ... ... Detection 7: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 10: 
#> ... ... Calculating weights... 
#> ... ... Detection 8: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 11: 
#> ... ... Calculating weights... 
#> ... ... Detection 9: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 12: 
#> ... ... Calculating weights... 
#> ... ... Detection 10: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 13: 
#> ... ... Calculating weights... 
#> ... ... Detection 11: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 14: 
#> ... ... Calculating weights... 
#> ... ... Detection 12: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 15: 
#> ... ... Calculating weights... 
#> ... ... Detection 13: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 16: 
#> ... ... Calculating weights... 
#> ... ... Detection 14: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 17: 
#> ... ... Calculating weights... 
#> ... ... Detection 15: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 18: 
#> ... ... Calculating weights... 
#> ... ... Detection 16: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 19: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 20: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 21: 
#> ... ... Calculating weights... 
#> ... ... Detection 17: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 22: 
#> ... ... Calculating weights... 
#> ... ... Detection 18: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 23: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 24: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 25: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 26: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 27: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 28: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 29: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 30: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 31: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 32: 
#> ... ... Calculating weights... 
#> ... ... Detection 19: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 33: 
#> ... ... Calculating weights... 
#> ... ... Detection 20: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 34: 
#> ... ... Calculating weights... 
#> ... ... Detection 21: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 35: 
#> ... ... Calculating weights... 
#> ... ... Detection 22: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 36: 
#> ... ... Calculating weights... 
#> ... ... Detection 23: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 37: 
#> ... ... Calculating weights... 
#> ... ... Detection 24: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 38: 
#> ... ... Calculating weights... 
#> ... ... Detection 25: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 39: 
#> ... ... Calculating weights... 
#> ... ... Detection 26: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 40: 
#> ... ... Calculating weights... 
#> ... ... Detection 27: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 41: 
#> ... ... Calculating weights... 
#> ... ... Detection 28: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 42: 
#> ... ... Calculating weights... 
#> ... ... Detection 29: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 43: 
#> ... ... Calculating weights... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 44: 
#> ... ... Calculating weights... 
#> ... ... Detection 30: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 45: 
#> ... ... Calculating weights... 
#> ... ... Detection 31: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 46: 
#> ... ... Calculating weights... 
#> ... ... Detection 32: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 47: 
#> ... ... Calculating weights... 
#> ... ... Detection 33: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 48: 
#> ... ... Calculating weights... 
#> ... ... Detection 34: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 49: 
#> ... ... Calculating weights... 
#> ... ... Detection 35: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> 
#> ... Time step 50: 
#> ... ... Calculating weights... 
#> ... ... Detection 36: 
#> ... ... * Defining location given data... 
#> ... ... (Re-)sampling particles... 
#> ... ... Kicking particles into proposal location(s)... 
#> patter::pf_forward_2() call ended (@ 2023-11-24 20:46:47.012471). 

#### Example (4): Write history to file
# See `pf_forward_1()`

#### Example (4): Customise verbose options
# See `pf_forward_1()`

#### Example (5): Customise movement model (as in `pf_forward_1()`)
# See `pf_forward_1()`

#### Example (6): Implement the backward pass
# `pf_forward_*()` should be followed by `pf_backward()`