PF: forward run diagnostics

This function collates particle diagnostics from pf_forward().

pf_forward_diagnostics(.sink, ...)

Arguments

.sink

A character string that defines the directory in which pf_forward() outputs are located. Either {.record$sink}, as specified in pf_forward(), or {.record$sink}/diagnostics/ is acceptable. A pf_particles object is also accepted.

...

Arguments passed to arrow::open_dataset().

Value

The function returns the diagnostics data.table (see pf_particles).

Details

Particle diagnostics are fully described in pf_diag.

pf_forward_diagnostics() is designed to collate particle diagnostics on file. pf_forward() writes particle diagnostics to {.record$sink}/diagnostics (see pf_opt_record()). You can supply {.record$sink} or {.record$sink}/diagnostics/ to this function. The individual data.tables are collated to match the form in which they are provided by pf_forward() when .record$save = TRUE.

If pf_forward() is implemented with .record$save = TRUE, the outputted pf_particles object from pf_forward() includes a pre-compiled data.table of diagnostics (see pf_opt_record()). pf_forward_diagnostics() accepts a pf_particles object, but there is really no need for this, as a diagnostics element is already present in this object.

To calculate particle diagnostics for the backward pass, see pf_backward_killer_diagnostics().

See also

Author

Edward Lavender

Examples

# For outputs stored in memory, extract the diagnostics element
dat_pff()$diagnostics
#>      iter_m iter_i timestep                component trial   n n_u         ess
#>   1:      1      1        1                 proposal    NA 203 203 0.004926108
#>   2:      1      1        1      lik-acs_filter_land    NA 203 203 0.004926108
#>   3:      1      1        1 lik-acs_filter_container    NA 203 203 0.004926108
#>   4:      1      1        1            lik-pf_lik_ac    NA 177 177 0.078771993
#>   5:      1      1        1            sample-origin     1 100  57 0.050732899
#>  ---                                                                          
#> 145:      1      1       25                 proposal     1 100  80 0.010000000
#> 146:      1      1       25      lik-acs_filter_land     1 100  80 0.010000000
#> 147:      1      1       25 lik-acs_filter_container     1 100  80 0.010000000
#> 148:      1      1       25            lik-pf_lik_ac     1 100  80 0.015502579
#> 149:      1      1       25              sample-kick     1 100  47 0.013213661

# pf_forward_diagnostics() also works but is unnecessary
pf_forward_diagnostics(dat_pff())
#>      iter_m iter_i timestep                component trial   n n_u         ess
#>   1:      1      1        1                 proposal    NA 203 203 0.004926108
#>   2:      1      1        1      lik-acs_filter_land    NA 203 203 0.004926108
#>   3:      1      1        1 lik-acs_filter_container    NA 203 203 0.004926108
#>   4:      1      1        1            lik-pf_lik_ac    NA 177 177 0.078771993
#>   5:      1      1        1            sample-origin     1 100  57 0.050732899
#>  ---                                                                          
#> 145:      1      1       25                 proposal     1 100  80 0.010000000
#> 146:      1      1       25      lik-acs_filter_land     1 100  80 0.010000000
#> 147:      1      1       25 lik-acs_filter_container     1 100  80 0.010000000
#> 148:      1      1       25            lik-pf_lik_ac     1 100  80 0.015502579
#> 149:      1      1       25              sample-kick     1 100  47 0.013213661

# For outputs stored on file, use pf_forward_diagnostics()
pff_folder <- dat_pff_src(.folder = NULL)
pf_forward_diagnostics(pff_folder)
#>      iter_m iter_i timestep                component trial   n n_u         ess
#>   1:      1      1        1                 proposal    NA 203 203 0.004926108
#>   2:      1      1        1      lik-acs_filter_land    NA 203 203 0.004926108
#>   3:      1      1        1 lik-acs_filter_container    NA 203 203 0.004926108
#>   4:      1      1        1            lik-pf_lik_ac    NA 177 177 0.078771993
#>   5:      1      1        1            sample-origin     1 100  57 0.050732899
#>  ---                                                                          
#> 145:      1      1       25                 proposal     1 100  80 0.010000000
#> 146:      1      1       25      lik-acs_filter_land     1 100  80 0.010000000
#> 147:      1      1       25 lik-acs_filter_container     1 100  80 0.010000000
#> 148:      1      1       25            lik-pf_lik_ac     1 100  80 0.015502579
#> 149:      1      1       25              sample-kick     1 100  47 0.013213661
pf_forward_diagnostics(file.path(pff_folder, "diagnostics"))
#>      iter_m iter_i timestep                component trial   n n_u         ess
#>   1:      1      1        1                 proposal    NA 203 203 0.004926108
#>   2:      1      1        1      lik-acs_filter_land    NA 203 203 0.004926108
#>   3:      1      1        1 lik-acs_filter_container    NA 203 203 0.004926108
#>   4:      1      1        1            lik-pf_lik_ac    NA 177 177 0.078771993
#>   5:      1      1        1            sample-origin     1 100  57 0.050732899
#>  ---                                                                          
#> 145:      1      1       25                 proposal     1 100  80 0.010000000
#> 146:      1      1       25      lik-acs_filter_land     1 100  80 0.010000000
#> 147:      1      1       25 lik-acs_filter_container     1 100  80 0.010000000
#> 148:      1      1       25            lik-pf_lik_ac     1 100  80 0.015502579
#> 149:      1      1       25              sample-kick     1 100  47 0.013213661

# Outputs are identical
stopifnot(all.equal(
  dat_pff()$diagnostics,
  pf_forward_diagnostics(pff_folder)
))