PF: access particle histories

These functions function access or collate particle histories.

.pf_history_list() lists particle histories or file pointers;

.pf_history_read() distinguishes particle histories in memory versus on file;

.pf_history_write() defines whether or not to write particle histories to file;

.pf_history_cols() validates input and output columns;

.pf_history_elm() accesses particle histories for a specific list element;

.pf_history_dt() collates particle histories into a single data.table;

.pf_history_list(.history, ...)

.pf_history_read(.history)

.pf_history_write(.record)

.pf_history_cols(.history, .record, .input_cols, .output_cols = NULL)

.pf_history_elm(
  .history,
  .elm,
  .read = .pf_history_read(.history),
  .cols = NULL,
  ...
)

.pf_history_dt(.history, ..., .collect = TRUE)

Arguments

.history

The particle samples. In .pf_history_list() and .pf_history_dt(), these can be supplied as:

A pf_particles object;

The history element of a pf_particles object;

A character string that defines the directory containing parquet files;

A list of file paths (e.g., from pf_files());

For .pf_history_read(), .pf_history_elm() and .pf_history_cols(), .history should be a list.

All functions also silently accept (and return) NULL.

...

Additional arguments:

In .pf_history_list(), ... is passed to pf_files() if .history is a directory;

In .pf_history_elm(), ... is passed to arrow::read_parquet() is .read = TRUE. col_select is controlled via .cols and not permitted;

In .pf_history_dt(): ... is passed to arrow::open_dataset() if .history is a directory.

.record, .input_cols, .output_columns

Additional arguments for .pf_history_write() and .pf_history_cols():

.record is a named list of output options, from pf_opt_record().

.input_cols and .output_cols are character vectors that define essential input and output columns;

.elm, .read, .cols

Additional arguments for .pf_history_elm():

.elm is an integer that defines the list index.

.read is a logical variable that defines whether or not to read particle samples from file.

(optional) .cols is a character vector of columns to select, for .pf_history_elm().

.collect

For .pf_history_dt(), if .history is a directory, .collect is a logical variable that defines whether or not to collect the dataset in memory.

Value

If .history = NULL, NULL is returned. Otherwise:

.pf_history_list() returns a list of data.tables (options 1 and 2) or file paths (options 3 and 4);

.pf_history_read() and .pf_history_write() each return a logical variable;

.pf_history_cols() returns a named list;

.pf_history_elm() returns a data.table for particle samples for a single time step;

.pf_history_dt() returns a data.table or an arrow::FileSystemDataset (if .history is a directory and .collect = FALSE).

List particle samples

.pf_history_list() lists particle samples or file pointers. There are four implementation options:

If .history is a pf_particles object, the function returns the history element.

If .history is the history element of a pf_particles object, it is returned unchanged.

If .history is a folder, pf_files() is used to return a list of file paths.

If .history is a list of file paths, they are returned unchanged.

This function is used in iterative routines, supported by .pf_history_elm().

Collate particle samples

.pf_history_dt() collates all particle samples. There are three implementation options:

If .history is a pf_particles object, or the history element of such an object, data.table::rbindlist(..., fill = TRUE) is used to combine particle samples.

If .history is a folder, arrow::open_dataset() plus (optionally) dplyr::collect() is used. Additional arguments, such as schema can be passed to arrow::open_dataset() via ....

If .history is a list of parquet files, files are iteratively read into memory (via arrow::read_parquet()) and combined (via data.table::rbindlist(..., fill = TRUE). This option is much slower than option 2.

Author

Edward Lavender

Examples

if (FALSE) {

# .pf_history_dt() examples

require(arrow)

# Use pf_particles-class object
.pf_history_dt(dat_pfbk())

# Use `history` element
.pf_history_dt(dat_pfbk()$history)

# Use directory (and optionally select a subset of columns)
pfbk_folder <- dat_pfbk_
sch <- schema(timestep = int32(),
              cell_now = int32(),
              x_now = double(),
              y_now = double())
.pf_history_dt(pfbk_folder, schema = sch)

# Use directory with `.collect = FALSE`
.pf_history_dt(pfbk_folder, schema = sch, .collect = FALSE)

# Use a file list
.pf_history_dt(pf_files(pfbk_folder))
}