PF: likelihood functions

These are likelihood functions for the forward simulation. pf_forward() expects a named list of likelihood functions that evaluate the likelihood of each dataset given location proposals. Convenience functions are provided that calculate the likelihood of acoustic and archival data, given location proposals, as resolved by the acoustic-container and depth-contour algorithms presented by Lavender et al. (2023).

acs_filter_land(.particles, .obs, .t, .dlist)

acs_filter_container(.particles, .obs, .t, .dlist)

pf_lik_ac(.particles, .obs, .t, .dlist)

pf_lik_dc(.particles, .obs, .t, .dlist)

Arguments

.particles

A data.table that defines proposal locations. This contains the following columns:

timestep---an integer that defines the time step;

cell_now---an integer that defines the grid cell(s) of proposal location(s);

x_now,y_now---doubles that define proposal location coordinates;

lik---a double that defines the likelihood. At each time step, this begins as a vector of 1s that should be progressively updated by each likelihood function.

.obs, .t

(optional) The .obs data.table (from pf_forward()) and an integer that defines the time step (used to index .obs).

.dlist

(optional) The .dlist list (from pf_forward()).

For acs_filter_land(), .dlist should contain .dlist$spatial$bathy and .dlist$pars$spatna.

For acs_filter_container(), .dlist should contain .dlist$data$moorings and .dlist$pars$lonlat.

For pf_lik_ac(), .dlist should contain .dlist$algorithm$detection_overlaps (from acs_setup_detection_overlaps()) and .dlist$algorithm$detection_kernels (from acs_setup_detection_kernels().

For pf_lik_dc(), .dlist should contain .dlist$spatial$bathy.

For custom likelihood functions, .dlist may require other datasets.

Value

Functions return the .particle

data.table for valid proposal locations and updated likelihoods (in the lik column).

Details

In pf_forward(), the likelihood of the data given proposal locations is evaluated using a named list of user-defined functions (one for each dataset). Each function must accept a data.table of proposal locations (.particles) alongside the arguments named above (even if they are ignored), evaluate the likelihood of the data and return an updated data.table for the subset of valid proposals and updated likelihoods (in the lik column). For computational efficiency, is is desirable that functions are ordered by the extent to which they invalidate proposal locations and that each function drops invalid proposals (since this reduces the number of subsequent likelihood calculations). For faster evaluations, it might also pay to group likelihood terms under a single wrapper function (since this eliminates the need to loop over individual terms in .pf_lik()).

The following convenience functions are provided:

acs_filter_land() is a binary 'habitability' (land/water) filter. This is useful when the 'stochastic kick' methodology is used to generate proposal locations in systems that include inhospitable habitats. The function calculates the likelihood (0, 1) of the 'hospitable' data given sampled particles. Location proposals in NA cells on the bathymetry grid (.dlist$spatial$bathy) are dropped.

acs_filter_container() is recommended for acoustic time series. This function calculates the approximate likelihood of the next acoustic detection, given particle proposals. This is a binary filter that excludes location proposals that are incompatible with acoustic container dynamics. This is strictly optional but facilitates algorithm convergence by filtering location proposals that are inconsistent with the location of the next detection.

pf_lik_ac() calculates the likelihood of acoustic data (detection or non detection at each operational receiver), given location proposals.

pf_lik_dc() calculates the likelihood of archival (depth) data, given particle proposals. This is based on Lavender et al.'s (2023) depth-contour algorithm in which depth observations are considered valid in locations whose depth's lie between a shallow and deep limit (.obs$depth_shallow[.t] and .obs$depth_deep[.t]) but impossible otherwise.

In pf_forward(), acs_filter_container() and pf_lik_ac() effectively replace the role of flapper::ac() function. pf_lik_dc() effectively replaces the role of flapper::dc().

See also

Author

Edward Lavender