PF: simulation options

These functions define selected function arguments for pf_forward().

pf_opt_trial(
  .trial_origin_crit = 1L,
  .trial_origin = 1L,
  .trial_kick_crit = 1L,
  .trial_kick = 1L,
  .trial_sampler_crit = 10L,
  .trial_sampler = 1L,
  .trial_revert_crit = 1L,
  .trial_revert_steps = 10L,
  .trial_revert = 2L
)

pf_opt_record(.save = FALSE, .cols = NULL, .sink = NULL)

pf_opt_control(.sampler_batch_size = 2L)

pf_opt_rerun_from(.rerun, .revert = 25L)

Arguments

.trial_origin_crit, .trial_origin, .trial_kick_crit, .trial_kick, .trial_sampler_crit, .trial_sampler, .trial_revert_crit, .trial_revert_steps, .trial_revert

pf_opt_trial() arguments used to tune convergence properties. All arguments expect integers.

.trial_{step} arguments define the number of times to trial a stochastic process at each time step (before giving up).

.trial_{step}_crit arguments define the number of valid, unique proposal locations (grid cells) required to trigger a repeated trial.

For example, if after kicking previous particles into proposal locations and evaluating the likelihood of each location only 10 unique particles remain, the stochastic process is repeated up to .trial_kick times or until the number of valid proposals exceeds .trial_kick_crit (see Details). In full:

.trial_origin_crit is the critical threshold for the starting samples. If the number of unique, valid starting locations is <= .trial_origin_crit, sampling is repeated up to .trial_origin times.

trial_kick_crit is the critical threshold for stochastic kicks. If the number of unique, valid proposals is <= trial_kick_crit, the process of kicking and sampling particles (via .rpropose and .sample) is repeated up to .trial_kick times. Use .trial_kick = 0L to suppress stochastic kicks.

.trial_sampler_crit is the critical threshold for directed sampling. Following stochastic kicks, if the number of unique, valid proposals remains <= .trial_sampler_crit, directed sampling is implemented. Samples are redrawn up to .trial_sampler times. Use .trial_sampler_crit = 0L to suppress directed sampling.

.trial_revert_crit is the critical threshold for a reversion. If the number of unique, valid proposal locations is <= .trial_revert_crit, the algorithm reverts by .trial_revert_steps time steps to an earlier time step (time step two or greater). .trial_revert is the total number of reversions permitted. This is reset on algorithm reruns (see .rerun).

.save, .sink, .cols

pf_opt_record() arguments, passed to .record in pf_forward().

.save---a logical variable that defines whether or not to save particle samples and diagnostics in memory. Use .save = TRUE with caution.

.sink---a character string that defines a (usually) empty directory in which to write particle samples and diagnostics. {.sink}/history/and {.sink}/diagnostics directories are created (if necessary) to store particle samples and diagnostics respectively.

.cols---a character vector that defines the names of the columns in particle-sample data.tables to save and/or write to file at each time step. This reduces the space occupied by outputs. For pf_backward_killer(), you need to retain timestep, cell_past and cell_now. For pf_backward_sampler(), you need timestep, cell_now, x_now and y_now for the backward sampler. For calculation of effective sample size, lik is required. NULL retains all columns.

At least one of .save = TRUE and .sink must be set.

.sampler_batch_size

pf_opt_control() arguments, passed to .control in pf_forward().

.sampler_batch_size---an integer that controls the batch size (the number of particles processed simultaneously) in directed sampling. Increase the batch size to improve speed; decrease the batch size to avoid memory constraints. The appropriate batch size depends on grid resolution and memory availability.

.rerun, .revert

pf_opt_rerun_from() arguments.

.rerun is a named list of algorithm outputs from a previous run.

.revert is an integer that defines the number of steps to revert.

Value

pf_opt_trial() returns a named list;

pf_opt_record() returns a named list;

pf_opt_control() returns a named list;

pf_opt_rerun_from() returns an integer;

Details

These functions are defined separately for convenience of documentation. Note that they do not define global options and must be passed to pf_forward() arguments.

See also

Author

Edward lavender

Examples

pf_opt_trial()
#> $trial_origin_crit
#> [1] 1
#> 
#> $trial_origin
#> [1] 1
#> 
#> $trial_kick_crit
#> [1] 1
#> 
#> $trial_kick
#> [1] 1
#> 
#> $trial_sampler_crit
#> [1] 10
#> 
#> $trial_sampler
#> [1] 1
#> 
#> $trial_revert_crit
#> [1] 1
#> 
#> $trial_revert_steps
#> [1] 10
#> 
#> $trial_revert
#> [1] 2
#> 
pf_opt_trial(.trial_kick_crit = 10L, .trial_kick = 2L)
#> $trial_origin_crit
#> [1] 1
#> 
#> $trial_origin
#> [1] 1
#> 
#> $trial_kick_crit
#> [1] 10
#> 
#> $trial_kick
#> [1] 2
#> 
#> $trial_sampler_crit
#> [1] 10
#> 
#> $trial_sampler
#> [1] 1
#> 
#> $trial_revert_crit
#> [1] 1
#> 
#> $trial_revert_steps
#> [1] 10
#> 
#> $trial_revert
#> [1] 2
#> 
pf_opt_record(.save = TRUE)
#> $save
#> [1] TRUE
#> 
#> $cols
#> NULL
#> 
#> $sink
#> NULL
#> 
pf_opt_control()
#> $sampler_batch_size
#> [1] 2
#> 
pf_opt_rerun_from(dat_pff(), .revert = 10L)
#> [1] 15