PF: particle internals

These internal functions facilitate proposals, likelihood calculations and sampling in pf_forward().

.pf_rpropose_origin(.particles = NULL, .obs, .t = 1L, .dlist)

.pf_sample_origin(.particles, .n, .sample, .trial_crit, .trial_count)

.pf_particles_origin(
  .particles = NULL,
  .obs,
  .t = 1L,
  .dlist,
  .rpropose = NULL,
  .dpropose = NULL,
  .likelihood,
  .sample,
  .n,
  .trial_crit,
  .trial_count,
  .control
)

.pf_lik(
  .particles,
  .obs,
  .t,
  .dlist,
  .stack,
  .diagnostics = list(),
  .trial = NA_integer_
)

.pf_lik_ac_detection(.particles, .kernels, .detections, .absences)

.pf_particles_kick(
  .particles,
  .obs,
  .t,
  .dlist,
  .rpropose,
  .dpropose = NULL,
  .likelihood,
  .sample,
  .n,
  .trial_crit,
  .trial_count,
  .control
)

.pf_particles_sampler(
  .particles,
  .obs,
  .t,
  .dlist,
  .rpropose = NULL,
  .dpropose,
  .likelihood,
  .sample,
  .n,
  .trial_crit,
  .trial_count,
  .control
)

Details

Function hierarchy is as follows:

pf_rpropose_*() functions generate proposals;

pf_lik_*() functions calculate likelihoods;

pf_sample_*() functions sample proposals;

pf_particle_*() functions wrap proposals, likelihood calculations and sampling in an iterative framework;

Proposal and likelihood functions must accept the following arguments:

.particles

.obs

.t

.dlist

.pf_lik() is a wrapper for specified likelihood functions (see pf_lik). The function returns a data.table that defines valid proposal locations, likelihoods and weights. A diagnostics attribute contains proposal diagnostics. This also requires:

.stack---a named list of likelihood functions (see pf_lik).

.diagnostics---an empty list used to store likelihood diagnostics, or NULL.

.trial---an integer that distinguishes trials.

Most likelihood functions are directly exported but some internals routines are also documented here for convenience, namely .pf_lik_ac_detection(). See also acs_*() functions.

Sampling functions must accept:

.particles

.n

Wrapper functions must accept:

Proposal function arguments (.particles, .obs, .t, .dlist);

Proposal functions (.rpropose,.dpropose);

Additional likelihood arguments (.likelihood);

Sampling arguments (.sample, .n);

Iteration arguments (.trial_crit, .trial_count);

Control arguments (.control);

Internal functions may support additional arguments as required.

At the first time step:

.pf_rpropose_origin() 'proposes' starting locations (quadrature points);

.pf_sample_origin() samples starting locations;

.pf_lik() calculates likelihoods;

.pf_particles_origin() integrates .pf_rpropose_origin(), .pf_lik() and .pf_particles_origin();

At subsequent time steps:

The exported functions pf_rpropose_kick() and pf_rpropose_reachable propose locations;

.pf_lik() calculates likelihoods;

.pf_particles_kick() and .pf_particles_sampler() are the integration functions;

pf_particles_*() functions are similar internally, but currently separated for convenience.

Author

Edward Lavender