PF: pivot path matrices

This function converts paths in 'wide format' to paths in 'long format'.

pf_path_pivot(.mat, .bathy = NULL, .obs = NULL, .cols = NULL)

Arguments

.mat

A wide-format data.table, from pf_path() with .return = "wide", in which:

columns (x1, x2, etc.) represent time steps;

rows represent paths;

.bathy

(optional) The bathymetry SpatRaster (used to extract cell coordinates).

.obs, .cols

(optional) A data.table and a character vector of column names in .obs to match onto the output. .obs must contain a timestep column for matching.

Value

The function returns a data.table with at least three columns:

path_id---An integer that defines each path (row in .mat);

timestep---An integer that defines each time step (column in .mat);

cell_id---An integer that defines each cell (value in .mat);

If .bathy is supplied, the data.table contains three additional columns:

cell_x---A double that defines the cell's x coordinate;

cell_y---A double that defines the cell's y coordinate;

cell_z---A double that defines the cell's z coordinate;

If .obs is supplied, the output also contains any columns specified in .cols.

Details

This function uses the fast collapse::pivot() function.

Author

Edward Lavender

Examples

# Define a hypothetical set of paths in wide-format
require(data.table)
mat <- matrix(1:25, ncol = 5)
mat <- as.data.table(mat)
colnames(mat) <- c("x1", "x2", "x3", "x4", "x5")

# Convert to long format
pf_path_pivot(mat)
#>     path_id timestep cell_id
#>  1:       1        1       1
#>  2:       1        2       6
#>  3:       1        3      11
#>  4:       1        4      16
#>  5:       1        5      21
#>  6:       2        1       2
#>  7:       2        2       7
#>  8:       2        3      12
#>  9:       2        4      17
#> 10:       2        5      22
#> 11:       3        1       3
#> 12:       3        2       8
#> 13:       3        3      13
#> 14:       3        4      18
#> 15:       3        5      23
#> 16:       4        1       4
#> 17:       4        2       9
#> 18:       4        3      14
#> 19:       4        4      19
#> 20:       4        5      24
#> 21:       5        1       5
#> 22:       5        2      10
#> 23:       5        3      15
#> 24:       5        4      20
#> 25:       5        5      25
#>     path_id timestep cell_id