PF: proposal functions

These are proposal functions for pf_forward().

pf_rpropose_kick(.particles, .obs, .t, .dlist, .rstep = rstep, ...)

pf_rpropose_reachable(.particles, .obs, .t, .dlist)

pf_dpropose(.particles, .obs, .t, .dlist, .dstep = dstep, ...)

Arguments

.particles, .obs, .t, .dlist

Required arguments for all .propose functions.

.particles---a data.table of particle samples from the previous time step;

.obs---the .obs data.table from pf_forward();

.t---an integer that defines the time step;

.dlist---the .dlist list from pf_forward();

For pf_rpropose_kick(), .dlist must contain .dlist$spatial$bathy and .dlist$pars$lonlat;

For pf_rpropose_reachable(), .dlist must contain .dlist$spatial$bathy;

for pf_dpropose(), .dlist must contain .dlist$pars$lonlat;

.rstep, .dstep, ...

Additional arguments for pf_rpropose_kick() and pf_dpropose() respectively. By default, in pf_forward() arguments passed via ... are passed to these arguments.

.rstep is a function, like rstep(), that simulates new locations;

.dstep is a function, like dstep(), that calculates the probability density of movements between locations;

At the time of writing, these functions must accept the following named arguments:

.xy0---a two-column object (matrix, data.frame, data.table) of accepted (x, y) coordinates from the previous time step. In .rstep, movement is simulated from .xy0 into new locations;

.xy1---For .dstep only, .xy1 is a two-column matrix of accepted coordinates for the current time step;

.lonlat---A logical variable that defines whether or not coordinates are in longitude/latitude format or planar;

...---Additional arguments;

Details

In pf_forward(), proposal functions are used to generate (propose) new, candidate locations for the individual's position, contingent upon previous positions (particle samples). Proposal locations are generated from previous locations via stochastic kicks and directed sampling.

The .rpropose argument in pf_forward() expects a stochastic-kick routine and pf_rpropose_kick() is the default. This is used to simulate proposal locations by 'kicking' particles into new locations as specified by a movement model (.rstep).

pf_rpropose_kick is a simple wrapper for rstep() that passes the .particles data.table to the function as required and then updates .particles with simulated locations. Simulated coordinates are redefined on the grid (.dlist$spatial$bathy) and grid cell IDs are included in the output. Argument passed via ... arguments are passed to .rstep, which under default settings means rstep()'s .rlen and .rang arguments, or additional arguments passed to those functions. This facilitates simulation of a wide variety of random walks. At the time of writing, correlated random walks are not easy for users to implement, but this should improve in future.

In pf_forward(), if stochastic kicks fail to produce a sufficient number of valid particle samples, pf_rpropose_reachable() may be called under-the-hood for directed sampling (see pf_opt_trial()). For selected particles, this function identifies the set of reachable locations. We evaluate the likelihood of reachable locations and the probability density of moving into each location and then sample locations according to the (normalised) product of these two variables. The .dpropose argument is required to calculate the probability density of moving between locations. This function must accept the usual .particles, .obs, .t, .dlist and ... arguments and return a data.table, for the subset of valid locations, with a dens column that defines probability densities. The function must be able to handle empty data.tables, which are passed down the call stack if all proposal (reachable) locations have zero likelihood.

pf_dpropose() is the default .dpropose routine. Under default settings, this is a simple wrapper for dstep() that handles empty data.tables or passes the relevant coordinate columns, for the accepted locations from the previous time step ((x_past, y_past) in .particles) and the proposal locations for the current time step ((x_now, y_now) in .particles) to dstep(). In this situation, ... arguments are passed to dstep().

See also

Author

Edward Lavender