PF: sampling functions

These are particle sampling functions for pf_forward().

pf_sample_multinomial(.particles, .n)

pf_sample_systematic(.particles, .n)

Arguments

.particles

A data.table that defines particle samples. This includes a weight column that should be used for (re)sampling.

.n

An integer that defines the number of particle samples at each time step.

Value

The functions return a data.table, as inputted, but containing .n rows, one for each sampled particle.

Details

pf_forward() expects a sampling function in the .sample argument. You can write your own function or use an exported one:

pf_sample_multinomial implements multinomial (re)sampling.

pf_sample_systematic() implements systematic (re)sampling.

Sampling functions must accept .particles and .n arguments and return a data.table of .n particle samples.

Systematic particle sampling (pf_sample_systematic()) is generally recommended.

See also

Author

Edward Lavender

Examples

require(data.table)
p <- data.table(cell_now = c(1, 2, 3), weight = c(0.5, 0.25, 0.25))
pf_sample_multinomial(p, .n = 10L)
#>     cell_now weight
#>  1:        3   0.25
#>  2:        1   0.50
#>  3:        3   0.25
#>  4:        3   0.25
#>  5:        2   0.25
#>  6:        1   0.50
#>  7:        2   0.25
#>  8:        1   0.50
#>  9:        2   0.25
#> 10:        1   0.50
pf_sample_systematic(p, .n = 10L)
#>     cell_now weight
#>  1:        1   0.50
#>  2:        1   0.50
#>  3:        1   0.50
#>  4:        1   0.50
#>  5:        1   0.50
#>  6:        2   0.25
#>  7:        2   0.25
#>  8:        2   0.25
#>  9:        3   0.25
#> 10:        3   0.25