PF set up: list files for PF

This function creates an ordered list of files for pf_backward_*().

pf_files(.root, ...)

Arguments

.root

A string that defines the directory in which files are located.

...

Additional arguments passed to list.files(), such as pattern, excluding full.names.

Value

The function returns an ordered list of file paths.

See also

Author

Edward Lavender

Examples

# Quick implementation of pf_forward()
con <- tempdir()
pff_folder <- file.path(con, "patter", "pf", "forward")
dir.create(pff_folder, recursive = TRUE)
#> Warning: '/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/forward' already exists
out_pff <- pf_forward(.obs = dat_obs(),
                      .bathy = dat_gebco(),
                      .moorings = dat_moorings, .detection_overlaps = dat_overlaps(),
                      .detection_kernels = dat_kernels(),
                      .record_opts = list(sink = pff_folder))
#> `patter::pf_forward()` called @ 2023-12-12 21:41:22... 
#> ... Setting up simulation... 
#> Warning: The directory '/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/forward/history' is not empty.
#> Warning: The directory '/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/forward/diagnostics' is not empty.
#> Warning: `.moorings` coordinates coerced onto `.bathy` grid.
#> ... Defining origin... 
#>   |====                                                                            | 4 % elapsed=00s, remaining~00s... Initiating simulation... 
#> ... ... Time step 2: 
#>   |====                                                                            | 4 % elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 3: 
#>   |=====                                                                           | 6 % elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 4: 
#>   |=======                                                                         | 8 % elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 5: 
#>   |========                                                                        | 10% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 6: 
#>   |==========                                                                      | 12% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 7: 
#>   |============                                                                    | 14% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 8: 
#>   |=============                                                                   | 16% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 9: 
#>   |===============                                                                 | 18% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 10: 
#>   |================                                                                | 20% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 11: 
#>   |==================                                                              | 22% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 12: 
#>   |====================                                                            | 24% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 13: 
#>   |=====================                                                           | 26% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 14: 
#>   |=======================                                                         | 28% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 15: 
#>   |========================                                                        | 30% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 16: 
#>   |==========================                                                      | 32% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 17: 
#>   |============================                                                    | 34% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 18: 
#>   |=============================                                                   | 36% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 19: 
#>   |===============================                                                 | 38% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 20: 
#>   |================================                                                | 40% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 21: 
#>   |==================================                                              | 42% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 22: 
#>   |====================================                                            | 44% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 23: 
#>   |=====================================                                           | 46% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 24: 
#>   |=======================================                                         | 48% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 25: 
#>   |========================================                                        | 50% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 26: 
#>   |==========================================                                      | 52% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 27: 
#>   |============================================                                    | 54% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 28: 
#>   |=============================================                                   | 56% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 29: 
#>   |===============================================                                 | 58% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 30: 
#>   |================================================                                | 60% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 31: 
#>   |==================================================                              | 62% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 32: 
#>   |====================================================                            | 64% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 33: 
#>   |=====================================================                           | 66% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 34: 
#>   |=======================================================                         | 68% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 35: 
#>   |========================================================                        | 70% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 36: 
#>   |==========================================================                      | 72% elapsed=00s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 37: 
#>   |============================================================                    | 74% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 38: 
#>   |=============================================================                   | 76% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 39: 
#>   |===============================================================                 | 78% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 40: 
#>   |================================================================                | 80% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 41: 
#>   |==================================================================              | 82% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 42: 
#>   |====================================================================            | 84% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 43: 
#>   |=====================================================================           | 86% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 44: 
#>   |=======================================================================         | 88% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 45: 
#>   |========================================================================        | 90% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 46: 
#>   |==========================================================================      | 92% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 47: 
#>   |============================================================================    | 94% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 48: 
#>   |=============================================================================   | 96% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 49: 
#>   |=============================================================================== | 98% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> ... ... Time step 50: 
#>   |================================================================================| 100% elapsed=01s, remaining~00s... ... ... Kicking particles... 
#> 
#> `patter::pf_forward()` call ended @ 2023-12-12 21:41:23 (duration: ~1 sec(s)). 

# List files for pf_backward_*()
files <- pf_files(file.path(pff_folder, "history"))

# Quick implementation of pf_backward_killer()
pfb_folder <- file.path(con, "patter", "pf", "backward")
dir.create(pfb_folder, recursive = TRUE)
#> Warning: '/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward' already exists
out_pfb <- pf_backward_killer(files,
                              .write_history = list(sink = pfb_folder))
#> Warning: `.write_history$sink` ('/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward') is not an empty directory.
#> patter::pf_backward_*() called (@ 2023-12-12 21:41:23.825633)... 
#>   |                                                                                | 0 % elapsed=00s     |==                                                                              | 2 % elapsed=00s, remaining~00s... Time step 50: 
#> ... ... Reading `.history[[50]]`... 
#> ... ... Reading `.history[[49]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[50]]`... 
#>   |====                                                                            | 4 % elapsed=00s, remaining~00s... Time step 49: 
#> ... ... Reading `.history[[48]]`... 
#> ... ... Cleaning `.history[[48]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 100 rows in `.history[[t]]`... 
#> ... ... ... Filtering 13 dead ends (87 remain)... 
#> ... ... ... Output: 87 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[49]]`... 
#>   |=====                                                                           | 6 % elapsed=00s, remaining~00s... Time step 48: 
#> ... ... Reading `.history[[47]]`... 
#> ... ... Cleaning `.history[[47]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 87 rows in `.history[[t]]`... 
#> ... ... ... Filtering 27 dead ends (73 remain)... 
#> ... ... ... Output: 73 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[48]]`... 
#>   |=======                                                                         | 8 % elapsed=00s, remaining~00s... Time step 47: 
#> ... ... Reading `.history[[46]]`... 
#> ... ... Cleaning `.history[[46]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 73 rows in `.history[[t]]`... 
#> ... ... ... Filtering 15 dead ends (85 remain)... 
#> ... ... ... Output: 85 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[47]]`... 
#>   |========                                                                        | 10% elapsed=00s, remaining~00s... Time step 46: 
#> ... ... Reading `.history[[45]]`... 
#> ... ... Cleaning `.history[[45]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 85 rows in `.history[[t]]`... 
#> ... ... ... Filtering 29 dead ends (71 remain)... 
#> ... ... ... Output: 71 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[46]]`... 
#>   |==========                                                                      | 12% elapsed=00s, remaining~00s... Time step 45: 
#> ... ... Reading `.history[[44]]`... 
#> ... ... Cleaning `.history[[44]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 71 rows in `.history[[t]]`... 
#> ... ... ... Filtering 36 dead ends (64 remain)... 
#> ... ... ... Output: 64 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[45]]`... 
#>   |============                                                                    | 14% elapsed=00s, remaining~00s... Time step 44: 
#> ... ... Reading `.history[[43]]`... 
#> ... ... Cleaning `.history[[43]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 64 rows in `.history[[t]]`... 
#> ... ... ... Filtering 55 dead ends (45 remain)... 
#> ... ... ... Output: 45 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[44]]`... 
#>   |=============                                                                   | 16% elapsed=00s, remaining~00s... Time step 43: 
#> ... ... Reading `.history[[42]]`... 
#> ... ... Cleaning `.history[[42]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 45 rows in `.history[[t]]`... 
#> ... ... ... Filtering 35 dead ends (65 remain)... 
#> ... ... ... Output: 65 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[43]]`... 
#>   |===============                                                                 | 18% elapsed=00s, remaining~00s... Time step 42: 
#> ... ... Reading `.history[[41]]`... 
#> ... ... Cleaning `.history[[41]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 65 rows in `.history[[t]]`... 
#> ... ... ... Filtering 37 dead ends (63 remain)... 
#> ... ... ... Output: 63 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[42]]`... 
#>   |================                                                                | 20% elapsed=00s, remaining~00s... Time step 41: 
#> ... ... Reading `.history[[40]]`... 
#> ... ... Cleaning `.history[[40]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 63 rows in `.history[[t]]`... 
#> ... ... ... Filtering 34 dead ends (66 remain)... 
#> ... ... ... Output: 66 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[41]]`... 
#>   |==================                                                              | 22% elapsed=00s, remaining~00s... Time step 40: 
#> ... ... Reading `.history[[39]]`... 
#> ... ... Cleaning `.history[[39]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 66 rows in `.history[[t]]`... 
#> ... ... ... Filtering 33 dead ends (67 remain)... 
#> ... ... ... Output: 67 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[40]]`... 
#>   |====================                                                            | 24% elapsed=00s, remaining~00s... Time step 39: 
#> ... ... Reading `.history[[38]]`... 
#> ... ... Cleaning `.history[[38]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 67 rows in `.history[[t]]`... 
#> ... ... ... Filtering 18 dead ends (82 remain)... 
#> ... ... ... Output: 82 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[39]]`... 
#>   |=====================                                                           | 26% elapsed=00s, remaining~00s... Time step 38: 
#> ... ... Reading `.history[[37]]`... 
#> ... ... Cleaning `.history[[37]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 82 rows in `.history[[t]]`... 
#> ... ... ... Filtering 17 dead ends (83 remain)... 
#> ... ... ... Output: 83 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[38]]`... 
#>   |=======================                                                         | 28% elapsed=00s, remaining~00s... Time step 37: 
#> ... ... Reading `.history[[36]]`... 
#> ... ... Cleaning `.history[[36]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 83 rows in `.history[[t]]`... 
#> ... ... ... Filtering 19 dead ends (81 remain)... 
#> ... ... ... Output: 81 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[37]]`... 
#>   |========================                                                        | 30% elapsed=00s, remaining~00s... Time step 36: 
#> ... ... Reading `.history[[35]]`... 
#> ... ... Cleaning `.history[[35]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 81 rows in `.history[[t]]`... 
#> ... ... ... Filtering 30 dead ends (70 remain)... 
#> ... ... ... Output: 70 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[36]]`... 
#>   |==========================                                                      | 32% elapsed=00s, remaining~00s... Time step 35: 
#> ... ... Reading `.history[[34]]`... 
#> ... ... Cleaning `.history[[34]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 70 rows in `.history[[t]]`... 
#> ... ... ... Filtering 19 dead ends (81 remain)... 
#> ... ... ... Output: 81 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[35]]`... 
#>   |============================                                                    | 34% elapsed=00s, remaining~00s... Time step 34: 
#> ... ... Reading `.history[[33]]`... 
#> ... ... Cleaning `.history[[33]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 81 rows in `.history[[t]]`... 
#> ... ... ... Filtering 33 dead ends (67 remain)... 
#> ... ... ... Output: 67 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[34]]`... 
#>   |=============================                                                   | 36% elapsed=00s, remaining~00s... Time step 33: 
#> ... ... Reading `.history[[32]]`... 
#> ... ... Cleaning `.history[[32]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 67 rows in `.history[[t]]`... 
#> ... ... ... Filtering 22 dead ends (78 remain)... 
#> ... ... ... Output: 78 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[33]]`... 
#>   |===============================                                                 | 38% elapsed=00s, remaining~00s... Time step 32: 
#> ... ... Reading `.history[[31]]`... 
#> ... ... Cleaning `.history[[31]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 78 rows in `.history[[t]]`... 
#> ... ... ... Filtering 94 dead ends (6 remain)... 
#> ... ... ... Output: 6 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[32]]`... 
#>   |================================                                                | 40% elapsed=00s, remaining~00s... Time step 31: 
#> ... ... Reading `.history[[30]]`... 
#> ... ... Cleaning `.history[[30]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 6 rows in `.history[[t]]`... 
#> ... ... ... Filtering 96 dead ends (4 remain)... 
#> ... ... ... Output: 4 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[31]]`... 
#>   |==================================                                              | 42% elapsed=00s, remaining~00s... Time step 30: 
#> ... ... Reading `.history[[29]]`... 
#> ... ... Cleaning `.history[[29]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 4 rows in `.history[[t]]`... 
#> ... ... ... Filtering 94 dead ends (6 remain)... 
#> ... ... ... Output: 6 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[30]]`... 
#>   |====================================                                            | 44% elapsed=00s, remaining~00s... Time step 29: 
#> ... ... Reading `.history[[28]]`... 
#> ... ... Cleaning `.history[[28]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 6 rows in `.history[[t]]`... 
#> ... ... ... Filtering 91 dead ends (9 remain)... 
#> ... ... ... Output: 9 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[29]]`... 
#>   |=====================================                                           | 46% elapsed=00s, remaining~00s... Time step 28: 
#> ... ... Reading `.history[[27]]`... 
#> ... ... Cleaning `.history[[27]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 9 rows in `.history[[t]]`... 
#> ... ... ... Filtering 91 dead ends (9 remain)... 
#> ... ... ... Output: 9 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[28]]`... 
#>   |=======================================                                         | 48% elapsed=00s, remaining~00s... Time step 27: 
#> ... ... Reading `.history[[26]]`... 
#> ... ... Cleaning `.history[[26]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 9 rows in `.history[[t]]`... 
#> ... ... ... Filtering 89 dead ends (11 remain)... 
#> ... ... ... Output: 11 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[27]]`... 
#>   |========================================                                        | 50% elapsed=00s, remaining~00s... Time step 26: 
#> ... ... Reading `.history[[25]]`... 
#> ... ... Cleaning `.history[[25]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 11 rows in `.history[[t]]`... 
#> ... ... ... Filtering 83 dead ends (17 remain)... 
#> ... ... ... Output: 17 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[26]]`... 
#>   |==========================================                                      | 52% elapsed=00s, remaining~00s... Time step 25: 
#> ... ... Reading `.history[[24]]`... 
#> ... ... Cleaning `.history[[24]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 17 rows in `.history[[t]]`... 
#> ... ... ... Filtering 73 dead ends (27 remain)... 
#> ... ... ... Output: 27 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[25]]`... 
#>   |============================================                                    | 54% elapsed=00s, remaining~00s... Time step 24: 
#> ... ... Reading `.history[[23]]`... 
#> ... ... Cleaning `.history[[23]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 27 rows in `.history[[t]]`... 
#> ... ... ... Filtering 71 dead ends (29 remain)... 
#> ... ... ... Output: 29 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[24]]`... 
#>   |=============================================                                   | 56% elapsed=00s, remaining~00s... Time step 23: 
#> ... ... Reading `.history[[22]]`... 
#> ... ... Cleaning `.history[[22]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 29 rows in `.history[[t]]`... 
#> ... ... ... Filtering 73 dead ends (27 remain)... 
#> ... ... ... Output: 27 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[23]]`... 
#>   |===============================================                                 | 58% elapsed=00s, remaining~00s... Time step 22: 
#> ... ... Reading `.history[[21]]`... 
#> ... ... Cleaning `.history[[21]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 27 rows in `.history[[t]]`... 
#> ... ... ... Filtering 54 dead ends (46 remain)... 
#> ... ... ... Output: 46 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[22]]`... 
#>   |================================================                                | 60% elapsed=00s, remaining~00s... Time step 21: 
#> ... ... Reading `.history[[20]]`... 
#> ... ... Cleaning `.history[[20]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 46 rows in `.history[[t]]`... 
#> ... ... ... Filtering 59 dead ends (41 remain)... 
#> ... ... ... Output: 41 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[21]]`... 
#>   |==================================================                              | 62% elapsed=00s, remaining~00s... Time step 20: 
#> ... ... Reading `.history[[19]]`... 
#> ... ... Cleaning `.history[[19]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 41 rows in `.history[[t]]`... 
#> ... ... ... Filtering 45 dead ends (55 remain)... 
#> ... ... ... Output: 55 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[20]]`... 
#>   |====================================================                            | 64% elapsed=00s, remaining~00s... Time step 19: 
#> ... ... Reading `.history[[18]]`... 
#> ... ... Cleaning `.history[[18]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 55 rows in `.history[[t]]`... 
#> ... ... ... Filtering 37 dead ends (63 remain)... 
#> ... ... ... Output: 63 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[19]]`... 
#>   |=====================================================                           | 66% elapsed=00s, remaining~00s... Time step 18: 
#> ... ... Reading `.history[[17]]`... 
#> ... ... Cleaning `.history[[17]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 63 rows in `.history[[t]]`... 
#> ... ... ... Filtering 29 dead ends (71 remain)... 
#> ... ... ... Output: 71 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[18]]`... 
#>   |=======================================================                         | 68% elapsed=00s, remaining~00s... Time step 17: 
#> ... ... Reading `.history[[16]]`... 
#> ... ... Cleaning `.history[[16]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 71 rows in `.history[[t]]`... 
#> ... ... ... Filtering 16 dead ends (84 remain)... 
#> ... ... ... Output: 84 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[17]]`... 
#>   |========================================================                        | 70% elapsed=00s, remaining~00s... Time step 16: 
#> ... ... Reading `.history[[15]]`... 
#> ... ... Cleaning `.history[[15]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 84 rows in `.history[[t]]`... 
#> ... ... ... Filtering 23 dead ends (77 remain)... 
#> ... ... ... Output: 77 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[16]]`... 
#>   |==========================================================                      | 72% elapsed=00s, remaining~00s... Time step 15: 
#> ... ... Reading `.history[[14]]`... 
#> ... ... Cleaning `.history[[14]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 77 rows in `.history[[t]]`... 
#> ... ... ... Filtering 27 dead ends (73 remain)... 
#> ... ... ... Output: 73 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[15]]`... 
#>   |============================================================                    | 74% elapsed=00s, remaining~00s... Time step 14: 
#> ... ... Reading `.history[[13]]`... 
#> ... ... Cleaning `.history[[13]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 73 rows in `.history[[t]]`... 
#> ... ... ... Filtering 20 dead ends (80 remain)... 
#> ... ... ... Output: 80 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[14]]`... 
#>   |=============================================================                   | 76% elapsed=00s, remaining~00s... Time step 13: 
#> ... ... Reading `.history[[12]]`... 
#> ... ... Cleaning `.history[[12]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 80 rows in `.history[[t]]`... 
#> ... ... ... Filtering 27 dead ends (73 remain)... 
#> ... ... ... Output: 73 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[13]]`... 
#>   |===============================================================                 | 78% elapsed=00s, remaining~00s... Time step 12: 
#> ... ... Reading `.history[[11]]`... 
#> ... ... Cleaning `.history[[11]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 73 rows in `.history[[t]]`... 
#> ... ... ... Filtering 36 dead ends (64 remain)... 
#> ... ... ... Output: 64 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[12]]`... 
#>   |================================================================                | 80% elapsed=00s, remaining~00s... Time step 11: 
#> ... ... Reading `.history[[10]]`... 
#> ... ... Cleaning `.history[[10]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 64 rows in `.history[[t]]`... 
#> ... ... ... Filtering 30 dead ends (70 remain)... 
#> ... ... ... Output: 70 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[11]]`... 
#>   |==================================================================              | 82% elapsed=00s, remaining~00s... Time step 10: 
#> ... ... Reading `.history[[9]]`... 
#> ... ... Cleaning `.history[[9]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 70 rows in `.history[[t]]`... 
#> ... ... ... Filtering 34 dead ends (66 remain)... 
#> ... ... ... Output: 66 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[10]]`... 
#>   |====================================================================            | 84% elapsed=00s, remaining~00s... Time step 9: 
#> ... ... Reading `.history[[8]]`... 
#> ... ... Cleaning `.history[[8]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 66 rows in `.history[[t]]`... 
#> ... ... ... Filtering 31 dead ends (69 remain)... 
#> ... ... ... Output: 69 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[9]]`... 
#>   |=====================================================================           | 86% elapsed=00s, remaining~00s... Time step 8: 
#> ... ... Reading `.history[[7]]`... 
#> ... ... Cleaning `.history[[7]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 69 rows in `.history[[t]]`... 
#> ... ... ... Filtering 44 dead ends (56 remain)... 
#> ... ... ... Output: 56 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[8]]`... 
#>   |=======================================================================         | 88% elapsed=00s, remaining~00s... Time step 7: 
#> ... ... Reading `.history[[6]]`... 
#> ... ... Cleaning `.history[[6]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 56 rows in `.history[[t]]`... 
#> ... ... ... Filtering 32 dead ends (68 remain)... 
#> ... ... ... Output: 68 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[7]]`... 
#>   |========================================================================        | 90% elapsed=00s, remaining~00s... Time step 6: 
#> ... ... Reading `.history[[5]]`... 
#> ... ... Cleaning `.history[[5]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 68 rows in `.history[[t]]`... 
#> ... ... ... Filtering 32 dead ends (68 remain)... 
#> ... ... ... Output: 68 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[6]]`... 
#>   |==========================================================================      | 92% elapsed=00s, remaining~00s... Time step 5: 
#> ... ... Reading `.history[[4]]`... 
#> ... ... Cleaning `.history[[4]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 68 rows in `.history[[t]]`... 
#> ... ... ... Filtering 52 dead ends (48 remain)... 
#> ... ... ... Output: 48 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[5]]`... 
#>   |============================================================================    | 94% elapsed=00s, remaining~00s... Time step 4: 
#> ... ... Reading `.history[[3]]`... 
#> ... ... Cleaning `.history[[3]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 48 rows in `.history[[t]]`... 
#> ... ... ... Filtering 46 dead ends (54 remain)... 
#> ... ... ... Output: 54 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[4]]`... 
#>   |=============================================================================   | 96% elapsed=00s, remaining~00s... Time step 3: 
#> ... ... Reading `.history[[2]]`... 
#> ... ... Cleaning `.history[[2]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 54 rows in `.history[[t]]`... 
#> ... ... ... Filtering 36 dead ends (64 remain)... 
#> ... ... ... Output: 64 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[3]]`... 
#>   |=============================================================================== | 98% elapsed=00s, remaining~00s... Time step 2: 
#> ... ... Reading `.history[[1]]`... 
#> ... ... Cleaning `.history[[1]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 64 rows in `.history[[t]]`... 
#> ... ... ... Filtering 43 dead ends (57 remain)... 
#> ... ... ... Output: 57 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[2]]`... 
#>   |================================================================================| 100% elapsed=00s, remaining~00s... Time step 1: 
#> ... ... Recording (cleaned) outputs for `.history[[1]]`... 
#> 
#> patter::pf_backward_*() call ended (@ 2023-12-12 21:41:24.023933). 

# List files from pf_backward_*()
pf_files(pfb_folder)
#> [[1]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/1.parquet"
#> 
#> [[2]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/2.parquet"
#> 
#> [[3]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/3.parquet"
#> 
#> [[4]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/4.parquet"
#> 
#> [[5]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/5.parquet"
#> 
#> [[6]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/6.parquet"
#> 
#> [[7]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/7.parquet"
#> 
#> [[8]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/8.parquet"
#> 
#> [[9]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/9.parquet"
#> 
#> [[10]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/10.parquet"
#> 
#> [[11]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/11.parquet"
#> 
#> [[12]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/12.parquet"
#> 
#> [[13]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/13.parquet"
#> 
#> [[14]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/14.parquet"
#> 
#> [[15]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/15.parquet"
#> 
#> [[16]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/16.parquet"
#> 
#> [[17]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/17.parquet"
#> 
#> [[18]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/18.parquet"
#> 
#> [[19]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/19.parquet"
#> 
#> [[20]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/20.parquet"
#> 
#> [[21]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/21.parquet"
#> 
#> [[22]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/22.parquet"
#> 
#> [[23]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/23.parquet"
#> 
#> [[24]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/24.parquet"
#> 
#> [[25]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/25.parquet"
#> 
#> [[26]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/26.parquet"
#> 
#> [[27]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/27.parquet"
#> 
#> [[28]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/28.parquet"
#> 
#> [[29]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/29.parquet"
#> 
#> [[30]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/30.parquet"
#> 
#> [[31]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/31.parquet"
#> 
#> [[32]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/32.parquet"
#> 
#> [[33]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/33.parquet"
#> 
#> [[34]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/34.parquet"
#> 
#> [[35]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/35.parquet"
#> 
#> [[36]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/36.parquet"
#> 
#> [[37]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/37.parquet"
#> 
#> [[38]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/38.parquet"
#> 
#> [[39]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/39.parquet"
#> 
#> [[40]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/40.parquet"
#> 
#> [[41]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/41.parquet"
#> 
#> [[42]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/42.parquet"
#> 
#> [[43]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/43.parquet"
#> 
#> [[44]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/44.parquet"
#> 
#> [[45]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/45.parquet"
#> 
#> [[46]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/46.parquet"
#> 
#> [[47]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/47.parquet"
#> 
#> [[48]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/48.parquet"
#> 
#> [[49]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/49.parquet"
#> 
#> [[50]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpWxc8gS/patter/pf/backward/50.parquet"
#> 

# Clean up
unlink(file.path(con, "patter"), recursive = TRUE)