PF: set up movement datasets

This function builds a timeline of observations for particle filtering. The forward filter (pf_forward()) iterates over this timeline. At each time step, the algorithm proposes candidate locations for an individual, evaluates the likelihood of each observation (e.g., acoustic and/or archival data) in the timeline at the current time step and (re)samples valid locations using likelihood-based weights.

pf_setup_obs(
  .dlist,
  .trim = TRUE,
  .step,
  .period = NULL,
  .mobility,
  .receiver_range
)

Arguments

.dlist

A named list of data and parameters from pat_setup_data(). This function requires:

(optional) .dlist$data$acoustics, with the following columns: timestamp and receiver_id.

(optional) .dlist$data$archival, with the following columns: timestamp and depth.

At least one dataset must be provided. Data should be given for a single individual. Other datasets are not currently integrated within this function but can be included afterwards (see Details).

.trim

If both acoustic and archival data are supplied, .trim is logical variable that defines whether or not to trim the time series to the time period for which they overlap.

.step

An character, passed to lubridate::period(), lubridate::round_date() and seq() that defines the duration between sequential time steps (e.g., "2 mins").

.period

(optional) A length-two POSIXct vector that defines the start and end time for the timeline. If unprovided, this is defined internally according to inputted movement datasets and .trim:

If only acoustics = .dlist$data$acoustics is provided, .period is taken as range(acoustics$timestamp).

If only archival = .dlist$data$archival is provided, .period is taken as range(archival$timestamp).

If both datasets are provided and .trim = FALSE, .period is defined as range(c(acoustics$timestamp, archival$timestamp)).

If both datasets are provided and .trim = TRUE, .period is defined as the range in timestamps for the overlapping region.

.mobility

A constant that defines the maximum (Euclidean) distance the individual could move in .step. See the flapper::get_mvt_mobility_*() functions to estimate .mobility from acoustic and/or archival data.

.receiver_range

A constant that defines the receiver detection range. At the time of writing, a constant value across all receivers and time steps is currently assumed (although receiver-specific detection kernels within this range are supported by acs_setup_detection_kernels()).

Value

The function returns a data.table with the following columns:

timestep---an integer that defines the time step;

timestamp---a regular sequences of POSIXct time stamps;

date---a character that defines the date;

mobility---a double that defines .mobility;

If acoustic data is provided, the following column(s) are also included:

detection_id---an integer vector that uniquely defines each detection;

detection---an integer that distinguishes the time steps at which detections were (1) or were not (0) recorded;

receiver_id---a list that defines the receiver(s) that recorded detection(s) at each time step;

receiver_id_next---a list that defines the receiver(s) that recorded the next detection(s);

buffer_past---a double that controls container growth from the past to the present;

buffer_future---a double that controls container shrinkage from the future to the present;

buffer_future_incl_gamma---a double (buffer_future + .receiver_range) that controls container shrinkage for pf_forward();

These columns are required to calculate the likelihood of acoustic observations (the detection or lack thereof at each operational receiver) in pf_forward() by the default pf_lik routines.

If archival data is provided, the following column(s) are also included:

depth---a number that defines the individual's depth (m) at each time step;

To calculate the likelihood of archival observations in pf_forward() by the default pf_lik_dc() routine, a depth envelope (as defined by depth_deep and depth_shallow columns) is also required (see pf_lik_dc()).

Details

This function defines the timeline of observations over which pf_forward() is implemented. The function implements the following routines:

Acoustic progressing (if applicable);

Acoustic time series are rounded to the nearest .step;

All receivers that recorded detection(s) in each time interval are listed;

Duplicate detections (at the same receiver in the same time interval) are dropped;

Archival processing (if applicable);

Archival time series are rounded to the nearest .step;

Multiple records (in the same time interval) are not permitted (in this instance, you need to aggregate observations manually or reduce .step);

Alignment (if applicable)

Acoustic and archival time series are optionally trimmed to the time period for which they overlap;

Time series

Acoustic and/or archival time series are mapped onto a regular time series;

Information required by pf_forward() is added;

At the time of writing, this function only supports acoustic and archival data. However, ancillary observations can easily be included in pf_forward() as extra columns in this data.table.

See also

Author

Edward Lavender

Examples

#### Define example datasets
# Define movement datasets for a single individual
acc  <- dat_acoustics[dat_acoustics$individual_id == 25, ]
arc  <- dat_archival[dat_archival$individual_id == 25, ]
# For example speed, focus on a subset of data
acc <- acc[as.Date(timestamp) == as.Date("2016-03-17")]
arc <- arc[as.Date(timestamp) == as.Date("2016-03-17")]
# Setup datasets as usual
dlist     <- pat_setup_data(.acoustics = acc,
                            .moorings = dat_moorings,
                            .archival = arc,
                            .bathy = dat_gebco(),
                            .lonlat = FALSE)
#> Warning: `.bathy` name updated from 'layer' to 'bathy'.

#### Example (1): Implement the function for acoustic time series only
dlist_acpf <- dlist
dlist_acpf$data$archival <- NULL
obs <- pf_setup_obs(.dlist = dlist_acpf,
                    .step = "2 mins",
                    .mobility = 500,
                    .receiver_range = 750)
head(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:        1 2016-03-17 01:50:00 2016-03-17            1         1          26
#> 2:        2 2016-03-17 01:52:00 2016-03-17            2         1          26
#> 3:        3 2016-03-17 01:54:00 2016-03-17            3         1          26
#> 4:        4 2016-03-17 01:56:00 2016-03-17            3         0            
#> 5:        5 2016-03-17 01:58:00 2016-03-17            4         1          26
#> 6:        6 2016-03-17 02:00:00 2016-03-17            5         1          26
#>    receiver_id_next mobility buffer_past buffer_future buffer_future_incl_gamma
#> 1:               26      500         500           500                     1250
#> 2:               26      500         500           500                     1250
#> 3:               26      500         500          1000                     1750
#> 4:               26      500         500           500                     1250
#> 5:               26      500         500           500                     1250
#> 6:               26      500         500          1000                     1750

#### Example (2): Use alternative parameters
obs <- pf_setup_obs(.dlist = dlist_acpf,
                    .step = "2 mins",
                    .mobility = 1000,
                    .receiver_range = 500)
head(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:        1 2016-03-17 01:50:00 2016-03-17            1         1          26
#> 2:        2 2016-03-17 01:52:00 2016-03-17            2         1          26
#> 3:        3 2016-03-17 01:54:00 2016-03-17            3         1          26
#> 4:        4 2016-03-17 01:56:00 2016-03-17            3         0            
#> 5:        5 2016-03-17 01:58:00 2016-03-17            4         1          26
#> 6:        6 2016-03-17 02:00:00 2016-03-17            5         1          26
#>    receiver_id_next mobility buffer_past buffer_future buffer_future_incl_gamma
#> 1:               26     1000        1000          1000                     1500
#> 2:               26     1000        1000          1000                     1500
#> 3:               26     1000        1000          2000                     2500
#> 4:               26     1000        1000          1000                     1500
#> 5:               26     1000        1000          1000                     1500
#> 6:               26     1000        1000          2000                     2500

#### Example (3): Implement the function for acoustic & archival time series
obs <- pf_setup_obs(.dlist = dlist,
                    .step = "2 mins",
                    .mobility = 500,
                    .receiver_range = 750)
head(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:        1 2016-03-17 01:50:00 2016-03-17            1         1          26
#> 2:        2 2016-03-17 01:52:00 2016-03-17            2         1          26
#> 3:        3 2016-03-17 01:54:00 2016-03-17            3         1          26
#> 4:        4 2016-03-17 01:56:00 2016-03-17            3         0            
#> 5:        5 2016-03-17 01:58:00 2016-03-17            4         1          26
#> 6:        6 2016-03-17 02:00:00 2016-03-17            5         1          26
#>    receiver_id_next mobility buffer_past buffer_future buffer_future_incl_gamma
#> 1:               26      500         500           500                     1250
#> 2:               26      500         500           500                     1250
#> 3:               26      500         500          1000                     1750
#> 4:               26      500         500           500                     1250
#> 5:               26      500         500           500                     1250
#> 6:               26      500         500          1000                     1750
#>    depth
#> 1: 73.78
#> 2: 73.32
#> 3: 73.32
#> 4: 73.32
#> 5: 73.55
#> 6: 68.70

#### Example (4): Include full acoustic & archival time series
obs <- pf_setup_obs(.dlist = dlist,
                    .trim = FALSE,
                    .step = "2 mins",
                    .mobility = 500,
                    .receiver_range = 750)
head(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:        1 2016-03-17 00:00:00 2016-03-17           NA         0            
#> 2:        2 2016-03-17 00:02:00 2016-03-17           NA         0            
#> 3:        3 2016-03-17 00:04:00 2016-03-17           NA         0            
#> 4:        4 2016-03-17 00:06:00 2016-03-17           NA         0            
#> 5:        5 2016-03-17 00:08:00 2016-03-17           NA         0            
#> 6:        6 2016-03-17 00:10:00 2016-03-17           NA         0            
#>    receiver_id_next mobility buffer_past buffer_future buffer_future_incl_gamma
#> 1:               26      500         500         27500                    28250
#> 2:               26      500         500         27000                    27750
#> 3:               26      500         500         26500                    27250
#> 4:               26      500         500         26000                    26750
#> 5:               26      500         500         25500                    26250
#> 6:               26      500         500         25000                    25750
#>     depth
#> 1: 134.61
#> 2: 135.77
#> 3: 135.53
#> 4: 136.00
#> 5: 135.07
#> 6: 134.38
tail(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:      715 2016-03-17 23:48:00 2016-03-17           92         0            
#> 2:      716 2016-03-17 23:50:00 2016-03-17           92         0            
#> 3:      717 2016-03-17 23:52:00 2016-03-17           92         0            
#> 4:      718 2016-03-17 23:54:00 2016-03-17           92         0            
#> 5:      719 2016-03-17 23:56:00 2016-03-17           92         0            
#> 6:      720 2016-03-17 23:58:00 2016-03-17           92         0            
#>    receiver_id_next mobility buffer_past buffer_future buffer_future_incl_gamma
#> 1:                       500         500          3000                     3750
#> 2:                       500         500          2500                     3250
#> 3:                       500         500          2000                     2750
#> 4:                       500         500          1500                     2250
#> 5:                       500         500          1000                     1750
#> 6:               NA      500         500           500                     1250
#>     depth
#> 1: 156.61
#> 2: 156.61
#> 3: 156.61
#> 4: 156.61
#> 5: 156.84
#> 6: 156.61