PF set up: list files from AC* for PF

This function creates an ordered list of 'record' files derived from an AC* algorithm (e.g., acs()) for particle filtering (via pf()).

pf_setup_record(.root, ...)

Arguments

.root

A string that defines the directory in which files are located.

...

Additional arguments passed to list.files(), such as pattern, excluding full.names.

Value

The function returns an ordered list of file paths.

See also

Author

Edward Lavender

Examples

# Quick implementation of AC algorithm
acoustics <- dat_acoustics[individual_id == 25, ]
archival <- dat_archival[individual_id == 25, ]
obs <- acs_setup_obs(acoustics, archival, "2 mins", 500)
obs <- obs[1:10, ]
gebco <- dat_gebco()
dat_moorings$receiver_range <- 500
containers <- acs_setup_detection_containers(gebco, dat_moorings)
# (For speed only, ignore overlaps)
# overlaps <- acs_setup_detection_overlaps(containers, dat_moorings)
kernels <-
  acs_setup_detection_kernels(dat_moorings,
                              .calc_detection_pr = acs_setup_detection_pr,
                              .bathy = gebco)
#> patter::acs_setup_detection_kernels() called (@ 2023-09-29 20:23:40.311755)... 
#> ... Setting up function... 
#> ... Getting receiver-specific kernels (for detection)... 
#> ... Getting receiver-specific inverse kernels... 
#> ... Getting area-wide kernels (for non-detection)... 
#> ... ... Get unique array designs... 
#> ... ... Get area wide kernels for each array design... 
#> 
#> ... ... ... For design 1/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 2/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 3/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 4/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 5/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 6/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 7/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 8/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 9/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 10/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 11/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 12/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 13/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 14/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 15/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 16/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 17/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 18/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> 
#> ... ... ... For design 19/19... 
#> ... ... ... ... Extract detection probability kernels for active receivers... 
#> ... ... ... ... Combining detection kernels to calculate the background detection probability surfaces (this is a slow step)... 
#> ... Process detection probability kernels ... 
#> ... patter::acs_setup_detection_kernels() call completed (@ 2023-09-29 20:23:43.272885) after ~0.05 minutes. 
folder <- file.path(tempdir(), "ac")
dir.create(folder)
out_ac <-
  acs(obs,
      .bathy = gebco,
      .detection_kernels = kernels,
      .write_record = list(filename = folder, overwrite = TRUE))
#> patter::acs() called (@ 2023-09-29 20:23:43.278442)... 
#> ... Time step 1: 
#> ... ... Detection 1: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 2: 
#> ... ... Detection 2: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 3: 
#> ... ... Detection 3: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 4: 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 5: 
#> ... ... Detection 4: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 6: 
#> ... ... Detection 5: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 7: 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 8: 
#> ... ... Detection 6: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 9: 
#> ... ... Detection 7: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 10: 
#> ... ... Detection 8: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> patter::acs() call ended (@ 2023-09-29 20:23:43.913034). 

# List files
pf_setup_record(folder)
#> [[1]]
#> [1] "/tmp/Rtmpd2Yoxr/ac/1.tif"
#> 
#> [[2]]
#> [1] "/tmp/Rtmpd2Yoxr/ac/2.tif"
#> 
#> [[3]]
#> [1] "/tmp/Rtmpd2Yoxr/ac/3.tif"
#> 
#> [[4]]
#> [1] "/tmp/Rtmpd2Yoxr/ac/4.tif"
#> 
#> [[5]]
#> [1] "/tmp/Rtmpd2Yoxr/ac/5.tif"
#> 
#> [[6]]
#> [1] "/tmp/Rtmpd2Yoxr/ac/6.tif"
#> 
#> [[7]]
#> [1] "/tmp/Rtmpd2Yoxr/ac/7.tif"
#> 
#> [[8]]
#> [1] "/tmp/Rtmpd2Yoxr/ac/8.tif"
#> 
#> [[9]]
#> [1] "/tmp/Rtmpd2Yoxr/ac/9.tif"
#> 
#> [[10]]
#> [1] "/tmp/Rtmpd2Yoxr/ac/10.tif"
#> 

# Clean up
unlink(folder, recursive = TRUE)