Simulation: helper functions

These convenience functions support the generation of animal movement paths and observations in de novo simulations (sim_*() functions) and simulation-based reconstructions of movement paths (pf_forward() and pf_backward_sampler()).

r*() functions simulate random variates;

c*() functions calculate outcomes from random-variate inputs;

d*() functions return densities;

r*() and c*() functions are used in de novo simulations (via sim_*() functions such as sim_path_walk()) and the forward simulation-based reconstruction of movement paths (in pf_forward()).

d*() functions are primarily used in the simulation-based reconstruction of movement paths as part of the backward sampler via pf_backward_sampler().

rtruncgamma(.n = 1, .shape = 15, .scale = 15, .mobility = 500, ...)

dtruncgamma(.x, .shape = 15, .scale = 15, .mobility = 500, ...)

rwn(.n = 1, .mu = 0, .rho = 0, .sd = 1, ...)

rlen(.n = 1, .prior = NULL, .t = NULL, ...)

rangrw(.n = 1, .prior = NULL, .t = NULL, ...)

rangcrw(.n = 1, .prior = NULL, .t = NULL, ...)

cstep(
  .xy_now,
  .xy_next = matrix(NA, nrow = nrow(.xy_now), ncol = 2L),
  .lonlat = FALSE,
  .length = rtruncgamma(nrow(.xy_now)),
  .angle = rwn(nrow(.xy_now))
)

dstep(.data_now, .data_past, ...)

Arguments

.n, .x, .shape, .scale, .mobility

Arguments for rtruncgamma() and dtruncgamma():

.n is an integer that defines the number of simulated outcome(s);

.x is a numeric vector that defines step length(s) (in the context of animal movement);

.shape is a numeric value that defines the shape parameter of a Gamma distribution (see stats::rgamma()).

.scale is a numeric value that defines the scale parameter of a Gamma distribution (see stats::rgamma()).

.mobility is a numeric value that defines the maximum possible value (step length) (see truncdist::rtrunc()).

...

Arguments passed to/from functions.

.mu, .rho, .sd

Arguments for rwn() for the simulation of turning angles, passed to the mu, rho and sd arguments of circular::rwrappednormal().

.prior, .t

Arguments for rlen(), rangrw() and rangcrw(), as used in top-level functions (e.g, sim_path_walk()):

.n---an integer that defines the number of simulated outcome(s);

.prior---a numeric vector that defines the simulated value(s) from the previous time step;

.t---an integer that defines the time step;

...---additional arguments, if needed;

.xy_now, .xy_next, .lonlat, .length, .angle

Arguments for cstep().

.xy_now is a two-column matrix of x and y coordinates;

(optional) .xy_next is a blank two-column matrix that is updated with future coordinates;

.lonlat is a logical variable that defines whether or not coordinates are in longitude/latitude;

.length is numeric vector of step lengths;

.angle is a numeric vector of turning angles;

.data_now, .data_past

Arguments for dstep().

.data_now is a data.table with coordinates in x_now and y_now columns;

.data_past is a data.table as above;

Movement

Core functions

Step lengths

rtruncgamma() simulates step length(s) from a truncated gamma distribution with a .mobility parameter.

dtruncgamma() returns the densit(ies) of specified step length(s).

Turning angles

rwn() simulates turning angle(s) from a wrapped normal distribution.

dwn() is not currently implemented.

Wrappers

The following wrapper functions are provided in the form required by front-end functions (e.g., sim_path_walk() and pf_backward_sampler()):

rlen() is a wrapper for rtruncgamma(). The corresponding function dlen() is not currently implemented.

rangrw(), rangcrw() are wrappers for rwn() for random walks and correlated random walks. The corresponding functions dangrw() and dangcrw() are not currently implemented.

Extensions

cstep() steps from previous location(s) into future location(s), given simulated step length(s) and turning angle(s).

dstep() returns the density of steps between two locations. This is only suitable for random walks (it wraps dtruncgamma() but corresponding functions for turning angles are not yet implemented).

See also

Author

Edward Lavender