Skill: evaluation metrics

These functions are standard model skill metrics. In patter, they support comparisons of simulated and reconstructed patterns of space use.

skill_mb(.obs, .mod, .summarise = "mean")

skill_me(.obs, .mod, .summarise = "mean")

skill_rmse(.obs, .mod)

skill_R(.obs, .mod)

skill_d(.obs, .mod)

Arguments

.obs

The 'observed' (true) pattern of space use, as a SpatRaster.

.mod

The 'modelled' (reconstructed) pattern of space use, as a SpatRaster.

.summarise

A function, passed to terra::global(), used to summarise SpatRaster values.

Value

The functions return a number.

Details

We follow the mathematical definitions in Lavender et al. (2022) Supplementary Information Sect. 3.2.1 (see here).

skill_mb() computes mean bias (if .summarise = "mean").

skill_me() computes mean error (if .summarise = "mean").

skill_rmse() computes root mean squared error.

skill_R() computes Spearman's rank correlation coefficient.

skill_d() computes the index of agreement.

These functions are not memory safe.

References

Lavender, E. et al. (2022). Benthic animal-borne sensors and citizen science combine to validate ocean modelling. Sci. Rep. 12: 16613. https://www.doi.org/1038/s41598-022-20254-z

See also

Author

Edward Lavender

Examples

# Define template SpatRasters
mod <- obs <- spatTemplate(.res = 1)
n <- terra::ncell(mod)
mod[] <- runif(n)
obs[] <- mod[] + rnorm(n)

# Calculate skill metrics
skill_mb(mod, obs)
#> [1] -0.002253598
skill_me(mod, obs)
#> [1] 0.7971842
skill_rmse(mod, obs)
#> [1] 0.9992181
skill_R(mod, obs)
#> [1] 0.273344
skill_d(mod, obs)
#> [1] -0.3729872