AC* set up: define detection container overlaps

This function identifies receivers with overlapping detection containers in space and time for the AC* algorithms.

acs_setup_detection_overlaps(.moorings, .services = NULL)

Source

This function supersedes the get_detection_containers_overlaps function in the flapper package.

Arguments

.moorings

A data.table that defines receiver deployments and associated information (see dat_moorings for an example). At a minimum, this must contain the following columns:

receiver_id---an integer vector of receiver IDs;

receiver_easting and receiver_northing or receiver_lon and receiver_lat---doubles that define receiver coordinates on a planar coordinate reference system or in longitude/latitude;

receiver_start---a Date vector that defines receiver deployment dates;

receiver_end---a Date vector that defines receiver retrieval dates;

receiver_range---a double that defines the detection range (m) for each receiver;

.services

(optional) A data.table that defines receiver IDs and servicing Dates (times during the deployment period of a receiver when it was not active due to servicing) (see make_matrix_receivers()). If provided, this must contain the following columns:

receiver_id---an integer vector of receiver IDs;

service_start---a Date vector that defines receiver servicing start dates;

service_end---a Date vector that defines receiver servicing completion dates;

Value

The function returns a nested list, with one element for all integers from 1:max(.moorings$receiver_id). Any elements that do not correspond to receivers contain a NULL element. List elements that correspond to receivers contain a NULL or a list that defines, for each deployment date with overlapping receiver(s), a vector of overlapping receiver(s).

Details

In the AC* algorithms, at the moment of detection, the set of possible locations depends on the receiver(s) at which an individual is, and is not, detected. The outputs of this function are used to restrict the probability calculations to the set of receivers that overlap with the receiver(s) at which an individual is detected for improved efficiency.

At the time of writing (November 2023), the function permits receiver ranges to differ between receivers, but assumes they are constant in time.

See also

Author

Edward Lavender

Examples

#### Example (1): Basic implementation
overlaps <- acs_setup_detection_overlaps(dat_moorings)
#> Warning: UTM coordinates used (both UTM and lon/lat coordinates detected).
summary(overlaps)
#>       Length Class  Mode
#>  [1,]   0    -none- NULL
#>  [2,]   0    -none- NULL
#>  [3,] 225    -none- list
#>  [4,] 371    -none- list
#>  [5,]   0    -none- NULL
#>  [6,]   0    -none- NULL
#>  [7,] 109    -none- list
#>  [8,]   0    -none- NULL
#>  [9,] 225    -none- list
#> [10,]   0    -none- NULL
#> [11,]   0    -none- NULL
#> [12,] 371    -none- list
#> [13,]   0    -none- NULL
#> [14,] 371    -none- list
#> [15,]   0    -none- NULL
#> [16,]   0    -none- NULL
#> [17,]   0    -none- NULL
#> [18,] 261    -none- list
#> [19,]   0    -none- NULL
#> [20,] 370    -none- list
#> [21,] 427    -none- list
#> [22,]   0    -none- NULL
#> [23,] 206    -none- list
#> [24,] 261    -none- list
#> [25,]   0    -none- NULL
#> [26,] 230    -none- list
#> [27,]   0    -none- NULL
#> [28,]   0    -none- NULL
#> [29,]   0    -none- NULL
#> [30,] 245    -none- list
#> [31,] 262    -none- list
#> [32,]   0    -none- NULL
#> [33,] 144    -none- list
#> [34,]   0    -none- NULL
#> [35,] 116    -none- list
#> [36,]   0    -none- NULL
#> [37,] 115    -none- list
#> [38,] 116    -none- list
#> [39,]   0    -none- NULL
#> [40,] 109    -none- list
#> [41,] 165    -none- list
#> [42,]   0    -none- NULL
#> [43,] 165    -none- list
#> [44,]  79    -none- list
#> [45,] 165    -none- list
#> [46,] 165    -none- list
#> [47,] 165    -none- list
#> [48,] 165    -none- list
#> [49,] 165    -none- list
#> [50,] 165    -none- list
#> [51,] 165    -none- list
#> [52,]  77    -none- list
#> [53,]  77    -none- list
#> [54,]  77    -none- list
#> [55,]  75    -none- list
#> [56,]  75    -none- list
#> [57,]  75    -none- list