AC* set up: detection container overlaps

This function is part of a set of acs_setup_*() functions that prepare the layers required to evaluate the likelihood of acoustic observations in an AC*PF algorithm. The function identifies receivers with overlapping detection containers in space and time.

acs_setup_detection_overlaps(.dlist)

Source

This function supersedes the get_detection_containers_overlaps function in the flapper package.

Arguments

.dlist

A named list of data and parameters from pat_setup_data(). This function requires:

.dlist$data$moorings, with the following columns: receiver_id, receiver_x, receiver_y, receiver_start, receiver_end and receiver_range.

(optional) .dlist$data$services, with the following columns: receiver_id, service_start and service_end.

Value

The function returns a nested list, with one element for all integers from 1:max(.moorings$receiver_id). Any elements that do not correspond to receivers contain a NULL element. List elements that correspond to receivers contain a NULL or a list that defines, for each deployment date with overlapping receiver(s), a vector of overlapping receiver(s).

Details

An AC*PF algorithm is a particle filtering algorithm that incorporates acoustic observations to reconstruct the possible movements of an individual. In such an algorithm, at the moment of detection, the likelihood of detection(s) is evaluated, given particle samples. The outputs of this function are used to restrict the likelihood calculations to the set of receivers that overlap with the receiver(s) at which an individual is detected. This improves efficiency.

See also

Author

Edward Lavender

Examples

# Set up data
dlist <- pat_setup_data(.moorings = dat_moorings,
                        .bathy = dat_gebco(),
                        .lonlat = FALSE)
# Prepare detection overlaps
# * Store detection overlaps in dlist$algorithms
# * It is expected in this slot by `pf_lik_ac()`
dlist$algorithm$overlaps <- acs_setup_detection_overlaps(dlist)