AC* set up: set up movement datasets

This function proccess passive acoustic telemetry detections and (optionally) archival time series for use in AC-branch algorithms.

acs_setup_obs(.acoustics, .archival = NULL, .step, .mobility)

Arguments

.acoustics

A data.table that defines passive acoustic telemetry detections (see dat_acoustics for an example) for a single individual. At a minimum, this must contain a timestamp column (an ordered, POSIXct vector that defines the times of detections) and receiver_id (an integer vector that defines the receiver(s) that recorded detections).

.archival

(optional) A data.table that defines depth (m) observations (see dat_archival for an example) for the same individual. At a minimum, this must contain a timestamp column (as in .acoustics) and a depth column (a positive-valued numeric vector that defines the individual's depth (m) below the surface at each time step).

.step

An character, passed to lubridate::period(), lubridate::round_date() and seq() that defines the duration between sequential time steps (e.g., "2 mins"). If .archival is supplied, .step should be the duration between sequential depth observations.

.mobility

A number that defines the maximum (Euclidean) distance the individual could move in .step.

Value

The function returns a data.table with the following columns:

timestep---an integer that defines the time step;

timestamp---a regular sequences of POSIXct time stamps;

date---a character that defines the date;

detection_id---an integer vector that uniquely defines each detection;

detection---an integer that distinguishes the time steps at which detections were (1) or were not (0) recorded;

receiver_id---a list that defines the receiver(s) that recorded detections at each time step;

buffer_past---a double that controls container growth from the past to the present;

buffer_future---a double that controls container shrinkage from the future to the present;

depth---if .archival is provided, depth is a number that defines the individual's depth (m) at each time step;

Details

This function implements the following routines:

Acoustic time series are rounded to the nearest .step;

All receivers that recorded detection(s) in each time interval are listed;

Duplicate detections (at the same receiver in the same time interval) are dropped;

Archival time series, if provided, are defined over the same time interval and both sets of time series are aligned;

Acoustic and archival time series are merged and ordered by time stamp;

Information required by the AC* algorithms (ultimately acs()) is added;

See also

Author

Edward Lavender

Examples

#### Define example datasets
acoustics <- dat_acoustics[dat_acoustics$individual_id == 25, ]
archival  <- dat_archival[dat_archival$individual_id == 25, ]

#### Example (1): Implement the function for acoustic time series only
obs <- acs_setup_obs(acoustics, .step = "2 mins", .mobility = 500)
utils::head(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:        1 2016-03-17 01:50:00 2016-03-17            1         1          26
#> 2:        2 2016-03-17 01:52:00 2016-03-17            2         1          26
#> 3:        3 2016-03-17 01:54:00 2016-03-17            3         1          26
#> 4:        4 2016-03-17 01:56:00 2016-03-17            3         0            
#> 5:        5 2016-03-17 01:58:00 2016-03-17            4         1          26
#> 6:        6 2016-03-17 02:00:00 2016-03-17            5         1          26
#>    buffer_past buffer_future
#> 1:         500           500
#> 2:         500           500
#> 3:         500          1000
#> 4:         500           500
#> 5:         500           500
#> 6:         500          1000

#### Example (2): Use alternative step lengths & mobilities
obs <- acs_setup_obs(acoustics, .step = "4 mins", .mobility = 1000)
utils::head(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:        1 2016-03-17 01:52:00 2016-03-17            1         1          26
#> 2:        2 2016-03-17 01:56:00 2016-03-17            2         1          26
#> 3:        3 2016-03-17 02:00:00 2016-03-17            3         1          26
#> 4:        4 2016-03-17 02:04:00 2016-03-17            4         1          26
#> 5:        5 2016-03-17 02:08:00 2016-03-17            5         1          26
#> 6:        6 2016-03-17 02:12:00 2016-03-17            6         1          26
#>    buffer_past buffer_future
#> 1:        1000          1000
#> 2:        1000          1000
#> 3:        1000          1000
#> 4:        1000          1000
#> 5:        1000          1000
#> 6:        1000          1000

#### Example (3): Implement the function for acoustic & archival time series
obs <- acs_setup_obs(acoustics, archival, .step = "2 mins", .mobility = 500)
utils::head(obs)
#>    timestep           timestamp       date detection_id detection receiver_id
#> 1:        1 2016-03-17 01:50:00 2016-03-17            1         1          26
#> 2:        2 2016-03-17 01:52:00 2016-03-17            2         1          26
#> 3:        3 2016-03-17 01:54:00 2016-03-17            3         1          26
#> 4:        4 2016-03-17 01:56:00 2016-03-17            3         0            
#> 5:        5 2016-03-17 01:58:00 2016-03-17            4         1          26
#> 6:        6 2016-03-17 02:00:00 2016-03-17            5         1          26
#>    buffer_past buffer_future depth
#> 1:         500           500 73.78
#> 2:         500           500 73.32
#> 3:         500          1000 73.32
#> 4:         500           500 73.32
#> 5:         500           500 73.55
#> 6:         500          1000 68.70