Utilities: check data

These functions validate user datasets. See Details for patter requirements.

check_bathy(.bathy)

check_acoustics(.acoustics, .moorings = NULL)

check_moorings(.moorings, .lonlat, .bathy)

check_services(.services, .moorings)

check_archival(.archival)

check_dlist(
  .dlist,
  .dataset = NULL,
  .spatial = NULL,
  .par = NULL,
  .algorithm = NULL
)

Arguments

.bathy

A (bathymetry) SpatRaster. This is widely used as a generic grid in patter functions.

.acoustics, .moorings, .services, .archival

Movement data.tables.

.lonlat

A logical variable that defines whether spatial data is in longitude/latitude or planar format. If unsupplied, this is defined (if possible) based on the columns in .moorings (via is.lonlat()).

.dlist, .dataset, .spatial, .par, .algorithm

Arguments for check_dlist().

Value

Each function returns the input dataset, as described, formatted in line with patter's requirements.

Details

These are internal functions that ensure that input dataset(s) meet patter requirements. Users should implement pat_setup_data() to prepare datasets for patter functions. Downstream functions (often) silently assume that input datasets meet all requirements, without subsequent checks. This simplifies internal code and documentation.

check_bathy() checks the bathymetry SpatRaster:

Class: SpatRaster;

Name(s): "bathy";

Properties:

A square grid is recommended;

Absolute values (m) are recommended;

At the time of writing (December 2023), planar grids are better tested;

check_acoustics() checks detection time series:

Class: data.table;

Columns:

timestamp---an ordered, POSIXct vector of time stamps with a defined tzone attribute;

receiver_id---see .moorings$receiver_id;

Properties:

Functions assume accurate clock alignment between receivers;

For most functions, a single time series (for one individual) is required;

check_services() checks acoustic servicing records:

Class: data.table;

Columns:

receiver_id---see .moorings$receiver_id;

receiver_start,receiver_end---Date vectors of service start and end dates. Before/after service events, receivers are assumed to have been deployed in the same locations; receiver deployments in different locations before/after servicing should be treated as distinct deployments in .moorings.

Properties:

Serviced receivers should be found in the .moorings dataset;

check_moorings() checks receiver deployment records:

Class: data.table;

Columns:

receiver_id---an integer vector of unique receiver deployments (1, ..., N);

receiver_easting and receiver_northing and/or receiver_lon and receiver_lat---Receiver coordinates. receiver_easting and receiver_northing must be present if .lonlat = FALSE and receiver_lon and receiver_lat must be present otherwise. The selected coordinate columns are copied, renamed to receiver_x and receiver_y and coerced onto the .bathy grid. Note that .bathy must be in the specified projection.

Properties:

receiver_start should precede receiver_end (not currently validated);

Receiver coordinates must be valid (not currently validated);

check_archival() checks archival depth time series data:

Class: data.table;

Columns:

timestamp---an ordered, POSIXct vector of time stamps with a defined tzone attribute;

depth---a numeric vector of (absolute) depths (m);

Properties:

Functions assume accurate clock alignment between acoustic and archival tags (if applicable);

For most functions, a single time series (for one individual) is required;

It is convenient if archival time stamps are regularly spaced;

check_dlist() is an internal function used to confirm that the data list passed to a patter function contains the elements required for that function.

For all datasets, NAs may cause unexpected errors and produce a message. NAs in columns other than those described above are safe.

NULL inputs are permitted in all check_{data} functions, in which case NULL is returned.

See also

Author

Edward Lavender