Parallelisation helpers

A set of internal wrappers for parallel::parallel functions that facilitate the implementation of parallel routines in functions via pbapply::pblapply().

cl_lapply(
  .x,
  .fun,
  ...,
  .cl = NULL,
  .varlist = NULL,
  .envir = .GlobalEnv,
  .use_chunks = FALSE,
  .combine = NULL
)

cl_check(.cl = NULL, .varlist = NULL)

cl_cores(.cl = NULL)

cl_chunks(.cl = NULL, .length)

cl_export(.cl = NULL, .varlist = NULL, .envir = .GlobalEnv)

cl_stop(.cl = NULL)

Arguments

.x

A list over which to iterate.

.fun, ...

A function that is applied to elements of .x and any additional named arguments.

.cl

(optional) The cluster argument passed to pbapply::pblapply(), supplied as:

A cluster object from parallel::makeCluster() or a sister function;

An integer that defines the number of child processes;

.varlist

(optional) A character vector of objects for export (see parallel::clusterExport()).

.envir

The environment from which to export variables (see parallel::clusterExport()).

.use_chunks

A logical vector that defines whether to parallelise or .x or batches of .x (chunks).

If .use_chunks = FALSE, function behaviour matches pbapply::pblapply().

If .use_chunks = TRUE, .x is split into a series of chunks that are processed in parallel; within each chunk .x is updated iteratively. This reduces the parallelisation overhead.

.combine

If .use_chunks = TRUE, .combine is function that defines how to combine the results for each chunk (e.g., purrr::list_flatten()). NULL is permitted, in which case the chunk-specific outputs are not combined.

.length

For cl_chunks(), .length is an integer that defines the number of elements in the iteration.

Value

cl_lapply() returns a list.

cl_cores() returns an integer.

cl_chunks() returns a list of integers.

cl_check(), cl_export() and cl_stop() return invisible(NULL).

Details

cl_lapply() is a wrapper for pbapply::pblapply() that handles cluster checking, set up, batch processing and cluster closure, using the following functions:

cl_check() checks .cl and .varlist arguments, as inputted to a parent function. For example, if .cl = NULL, .varlist should also be NULL.

cl_cores() identifies the number of cores specified.

cl_chunks() defines a list, with one element for core specified x 4, that contains an integer vector of the positions of an object over which to iterate serially in each chunk.

cl_export() implements parallel::clusterExport() if both .cl and .varlist are specified.

cl_stop() implements parallel::stopCluster() if .cl is a cluster object from parallel::makeCluster().

These routines evolved from the cl_*() functions in flapper.

See also

Author

Edward Lavender