Datasets: example algorithm outputs

These functions load example outputs from key patter functions. They are included in the package to streamline function examples and tests.

Usage

dat_path()

dat_coa()

dat_pff()

dat_pfb()

dat_tff()

Source

For full details on the algorithm parameters used to generate these datasets, see https://github.com/edwardlavender/patter/blob/main/data-raw/003-add-data-algorithms.R.

Value

The functions return a dataset or a character string that defines the path to a dataset. See the corresponding function documentation for full details.

Details

dat_path() reads an example output from sim_path_walk();

dat_coa() reads an example output from coa();

dat_pff() reads an example output from a forward run of pf_filter();

dat_pfb() reads an example output from a backward run of pf_filter();

dat_tff() reads an example output from pf_smoother_two_filter();

See also

Author

Edward Lavender

Examples

# Load inbuilt datasets
dat_path() |> summary()
#>     path_id     timestep        timestamp                     map_value      
#>  Min.   :1   Min.   :  1.00   Min.   :2016-01-01 00:00:00   Min.   :  5.661  
#>  1st Qu.:1   1st Qu.: 90.75   1st Qu.:2016-01-01 02:59:30   1st Qu.: 27.332  
#>  Median :1   Median :180.50   Median :2016-01-01 05:59:00   Median : 47.209  
#>  Mean   :1   Mean   :180.50   Mean   :2016-01-01 05:59:00   Mean   : 73.360  
#>  3rd Qu.:1   3rd Qu.:270.25   3rd Qu.:2016-01-01 08:58:30   3rd Qu.:116.657  
#>  Max.   :1   Max.   :360.00   Max.   :2016-01-01 11:58:00   Max.   :196.003  
#>        x                y          
#>  Min.   :706015   Min.   :6259156  
#>  1st Qu.:707440   1st Qu.:6261929  
#>  Median :710191   Median :6263235  
#>  Mean   :709700   Mean   :6263451  
#>  3rd Qu.:711246   3rd Qu.:6265161  
#>  Max.   :711852   Max.   :6267903  
dat_coa() |> summary()
#>     timestep      timestamp                     map_value            x         
#>  Min.   :1.00   Min.   :2016-01-01 00:00:00   Min.   : 24.39   Min.   :706900  
#>  1st Qu.:2.25   1st Qu.:2016-01-01 02:30:00   1st Qu.: 31.60   1st Qu.:708216  
#>  Median :3.50   Median :2016-01-01 05:00:00   Median : 52.49   Median :710228  
#>  Mean   :3.50   Mean   :2016-01-01 05:00:00   Mean   : 83.81   Mean   :709568  
#>  3rd Qu.:4.75   3rd Qu.:2016-01-01 07:30:00   3rd Qu.:134.28   3rd Qu.:711013  
#>  Max.   :6.00   Max.   :2016-01-01 10:00:00   Max.   :187.29   Max.   :711228  
#>        y          
#>  Min.   :6259666  
#>  1st Qu.:6262546  
#>  Median :6263398  
#>  Mean   :6263395  
#>  3rd Qu.:6264844  
#>  Max.   :6266321  
dat_pff() |> summary()
#>             Length Class      Mode
#> states      6      data.table list
#> diagnostics 4      data.table list
#> callstats   7      data.table list
dat_pfb() |> summary()
#>             Length Class      Mode
#> states      6      data.table list
#> diagnostics 4      data.table list
#> callstats   7      data.table list
dat_tff() |> summary()
#>             Length Class      Mode
#> states      6      data.table list
#> diagnostics 4      data.table list
#> callstats   7      data.table list