Datasets: example algorithm outputs

These functions load example outputs from key patter functions. They are included in the package to streamline function examples and tests.

dat_obs()

dat_pff()

dat_pfbk()

dat_pfp()

Source

For full details on the algorithm parameters used to generate these datasets, see https://github.com/edwardlavender/patter/blob/main/data-raw/002-add-data-algorithms.R.

Details

dat_pff() reads an example output from acs_setup_obs();

dat_pff() reads an example output from pf_forward();

dat_pfbk() reads an example output from pf_backward_killer();

dat_pfp() reads an example output from pf_path();

Author

Edward Lavender

Examples

dat_obs() |> summary()
#>     timestep       timestamp                       date          
#>  Min.   : 1.00   Min.   :2016-03-17 01:50:00   Length:50         
#>  1st Qu.:13.25   1st Qu.:2016-03-17 02:14:30   Class :character  
#>  Median :25.50   Median :2016-03-17 02:39:00   Mode  :character  
#>  Mean   :25.50   Mean   :2016-03-17 02:39:00                     
#>  3rd Qu.:37.75   3rd Qu.:2016-03-17 03:03:30                     
#>  Max.   :50.00   Max.   :2016-03-17 03:28:00                     
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>   detection_id     detection   
#>  Min.   : 1.00   Min.   :0.00  
#>  1st Qu.:11.25   1st Qu.:0.00  
#>  Median :18.00   Median :1.00  
#>  Mean   :17.94   Mean   :0.72  
#>  3rd Qu.:24.75   3rd Qu.:1.00  
#>  Max.   :36.00   Max.   :1.00  
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>  receiver_id.Length  receiver_id.Class  receiver_id.Mode
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  0        -none-   NULL                                 
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  0        -none-   NULL                                 
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  2        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  0        -none-   NULL                                 
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  receiver_id_next.Length  receiver_id_next.Class  receiver_id_next.Mode
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  2        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>     mobility    buffer_past  buffer_future  buffer_future_incl_gamma
#>  Min.   :500   Min.   :500   Min.   : 500   Min.   :1250            
#>  1st Qu.:500   1st Qu.:500   1st Qu.: 500   1st Qu.:1250            
#>  Median :500   Median :500   Median : 500   Median :1250            
#>  Mean   :500   Mean   :500   Mean   :1010   Mean   :1760            
#>  3rd Qu.:500   3rd Qu.:500   3rd Qu.:1000   3rd Qu.:1750            
#>  Max.   :500   Max.   :500   Max.   :5000   Max.   :5750            
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>      depth      
#>  Min.   :65.23  
#>  1st Qu.:66.15  
#>  Median :72.39  
#>  Mean   :70.71  
#>  3rd Qu.:73.32  
#>  Max.   :76.78  
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
#>                 
dat_pff() |> summary()
#>             Length Class      Mode   
#> history     50     -none-     list   
#> diagnostics  8     data.table list   
#> internal     1     -none-     list   
#> convergence  1     -none-     logical
#> time         1     -none-     list   
dat_pfbk() |> summary()
#>         Length Class  Mode
#> history 50     -none- list
#> time     3     -none- list
dat_pfp() |> summary()
#>     path_id          timestep       cell_id          cell_x      
#>  Min.   :  1.00   Min.   : 1.0   Min.   :35857   Min.   :708242  
#>  1st Qu.: 25.75   1st Qu.:13.0   1st Qu.:37943   1st Qu.:708842  
#>  Median : 50.50   Median :25.5   Median :38512   Median :709142  
#>  Mean   : 50.50   Mean   :25.5   Mean   :38411   Mean   :709125  
#>  3rd Qu.: 75.25   3rd Qu.:38.0   3rd Qu.:38901   3rd Qu.:709442  
#>  Max.   :100.00   Max.   :50.0   Max.   :39662   Max.   :709742  
#>      cell_y            cell_z      
#>  Min.   :6252207   Min.   : 36.50  
#>  1st Qu.:6252607   1st Qu.: 44.89  
#>  Median :6252807   Median : 52.99  
#>  Mean   :6252863   Mean   : 62.99  
#>  3rd Qu.:6253107   3rd Qu.: 80.03  
#>  Max.   :6254207   Max.   :140.85