Datasets: example algorithm outputs

These functions load example outputs from key patter functions. They are included in the package to streamline function examples and tests.

dat_obs()

dat_containers()

dat_overlaps()

dat_kernels()

dat_ac()

dat_pff()

dat_pfb()

dat_pfp()

Source

For full details on the algorithm parameters used to generate these datasets, see https://github.com/edwardlavender/patter/blob/main/data-raw/002-add-data-algorithms.R.

Details

dat_obs() reads an example input dataset from acs_setup_obs();

dat_containers() reads an example output from acs_setup_detection_containers();

dat_overlaps() reads an example output from acs_setup_detection_overlaps();

dat_kernels() reads an example output from acs_setup_detection_kernels();

dat_ac() reads an example output from acs();

dat_pff() reads an example output from pf_forward_1();

dat_pfb() reads an example output from pf_backward();

dat_pfp() reads an example output from pf_path();

Author

Edward Lavender

Examples

dat_obs() |> summary()
#>     timestep       timestamp                       date          
#>  Min.   : 1.00   Min.   :2016-03-17 01:50:00   Length:50         
#>  1st Qu.:13.25   1st Qu.:2016-03-17 02:14:30   Class :character  
#>  Median :25.50   Median :2016-03-17 02:39:00   Mode  :character  
#>  Mean   :25.50   Mean   :2016-03-17 02:39:00                     
#>  3rd Qu.:37.75   3rd Qu.:2016-03-17 03:03:30                     
#>  Max.   :50.00   Max.   :2016-03-17 03:28:00                     
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>   detection_id     detection   
#>  Min.   : 1.00   Min.   :0.00  
#>  1st Qu.:11.25   1st Qu.:0.00  
#>  Median :18.00   Median :1.00  
#>  Mean   :17.94   Mean   :0.72  
#>  3rd Qu.:24.75   3rd Qu.:1.00  
#>  Max.   :36.00   Max.   :1.00  
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>                                
#>  receiver_id.Length  receiver_id.Class  receiver_id.Mode
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  0        -none-   NULL                                 
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  0        -none-   NULL                                 
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  0        -none-   NULL                                 
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  2        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  0        -none-   NULL                                 
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  1        -none-   numeric                              
#>  receiver_id_next.Length  receiver_id_next.Class  receiver_id_next.Mode
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  2        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>  1        -none-   numeric                                             
#>   buffer_past  buffer_future  buffer_future_incl_gamma     depth      
#>  Min.   :500   Min.   : 500   Min.   :1250             Min.   :65.23  
#>  1st Qu.:500   1st Qu.: 500   1st Qu.:1250             1st Qu.:66.15  
#>  Median :500   Median : 500   Median :1250             Median :72.39  
#>  Mean   :500   Mean   :1010   Mean   :1760             Mean   :70.71  
#>  3rd Qu.:500   3rd Qu.:1000   3rd Qu.:1750             3rd Qu.:73.32  
#>  Max.   :500   Max.   :5000   Max.   :5750             Max.   :76.78  
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
dat_containers() |> summary()
#>    Length Class      Mode
#> 1  0      -none-     NULL
#> 2  0      -none-     NULL
#> 3  1      SpatRaster S4  
#> 4  1      SpatRaster S4  
#> 5  0      -none-     NULL
#> 6  0      -none-     NULL
#> 7  1      SpatRaster S4  
#> 8  0      -none-     NULL
#> 9  1      SpatRaster S4  
#> 10 0      -none-     NULL
#> 11 1      SpatRaster S4  
#> 12 1      SpatRaster S4  
#> 13 0      -none-     NULL
#> 14 1      SpatRaster S4  
#> 15 0      -none-     NULL
#> 16 0      -none-     NULL
#> 17 0      -none-     NULL
#> 18 1      SpatRaster S4  
#> 19 0      -none-     NULL
#> 20 1      SpatRaster S4  
#> 21 1      SpatRaster S4  
#> 22 0      -none-     NULL
#> 23 1      SpatRaster S4  
#> 24 1      SpatRaster S4  
#> 25 0      -none-     NULL
#> 26 1      SpatRaster S4  
#> 27 1      SpatRaster S4  
#> 28 0      -none-     NULL
#> 29 1      SpatRaster S4  
#> 30 1      SpatRaster S4  
#> 31 1      SpatRaster S4  
#> 32 0      -none-     NULL
#> 33 1      SpatRaster S4  
#> 34 0      -none-     NULL
#> 35 1      SpatRaster S4  
#> 36 1      SpatRaster S4  
#> 37 1      SpatRaster S4  
#> 38 1      SpatRaster S4  
#> 39 1      SpatRaster S4  
#> 40 1      SpatRaster S4  
#> 41 1      SpatRaster S4  
#> 42 0      -none-     NULL
#> 43 1      SpatRaster S4  
#> 44 1      SpatRaster S4  
#> 45 1      SpatRaster S4  
#> 46 1      SpatRaster S4  
#> 47 1      SpatRaster S4  
#> 48 1      SpatRaster S4  
#> 49 1      SpatRaster S4  
#> 50 1      SpatRaster S4  
#> 51 1      SpatRaster S4  
#> 52 1      SpatRaster S4  
#> 53 1      SpatRaster S4  
#> 54 1      SpatRaster S4  
#> 55 1      SpatRaster S4  
#> 56 1      SpatRaster S4  
#> 57 1      SpatRaster S4  
dat_overlaps() |> summary()
#>                  Length Class  Mode
#> list_by_receiver  57    -none- list
#> list_by_date     508    -none- list
dat_kernels() |> summary()
#>                               Length Class      Mode
#> receiver_specific_kernels      57    -none-     list
#> receiver_specific_inv_kernels  57    -none-     list
#> array_design                    4    data.frame list
#> array_design_by_date          508    -none-     list
#> bkg_surface_by_design          19    -none-     list
#> bkg_inv_surface_by_design      19    -none-     list
dat_ac() |> summary()
#>        Length Class  Mode
#> record 50     -none- list
#> map     0     -none- NULL
#> time    3     -none- list
dat_pff() |> summary()
#>         Length Class  Mode
#> history 50     -none- list
#> time     3     -none- list
dat_pfb() |> summary()
#>         Length Class  Mode
#> history 50     -none- list
#> time     3     -none- list
dat_pfp() |> summary()
#>     path_id          timestep       cell_id          cell_x      
#>  Min.   :   1.0   Min.   : 1.0   Min.   :36043   Min.   :708242  
#>  1st Qu.: 250.8   1st Qu.:13.0   1st Qu.:37376   1st Qu.:708642  
#>  Median : 500.5   Median :25.5   Median :37758   Median :708842  
#>  Mean   : 500.5   Mean   :25.5   Mean   :37842   Mean   :708925  
#>  3rd Qu.: 750.2   3rd Qu.:38.0   3rd Qu.:38511   3rd Qu.:709142  
#>  Max.   :1000.0   Max.   :50.0   Max.   :39651   Max.   :709542  
#>      cell_y            cell_z      
#>  Min.   :6252207   Min.   : 40.22  
#>  1st Qu.:6252807   1st Qu.: 60.09  
#>  Median :6253207   Median : 82.97  
#>  Mean   :6253161   Mean   : 87.30  
#>  3rd Qu.:6253407   3rd Qu.:108.30  
#>  Max.   :6254107   Max.   :150.05