Datasets: example algorithm outputs

These functions load example outputs from key patter functions, including those required to implement the forward filtering–backward sampling algorithm. They are included in the package to streamline function examples and tests.

dat_dlist()

dat_obs()

dat_pff()

dat_pff_src(.folder = c("history", "diagnostics"))

dat_pfbk()

dat_pfbk_src()

dat_pfbs()

dat_pfbs_src()

dat_pfp()

dat_coa()

Source

For full details on the algorithm parameters used to generate these datasets, see https://github.com/edwardlavender/patter/blob/main/data-raw/003-add-data-algorithms.R.

Arguments

.folder

For dat_pff_src(), .folder is NULL or a character that defines the name of the folder in which outputs are stored (history for particle samples or diagnostics for particle diagnostics). NULL returns the path to the directory containing both history/ and diagnostics/. See the documentation for pf_forward() and pf_particles objects for further details.

Value

The functions return a dataset or a character string that defines the path to a dataset. See the corresponding function documentation for full details.

Details

dat_dlist() reads example output from pat_setup_data();

dat_pff() reads an example output from pf_setup_obs();

dat_pff() reads an example output from pf_forward();

dat_pff_src() is the directory in which example .parquet files from pf_forward() are stored;

dat_pfbk() reads an example output from pf_backward_killer();

dat_pff_src() is the directory in which example .parquet files from pf_backward_killer() are stored;

dat_pfbs() reads an example output from pf_backward_sampler_v();

dat_pfbs_src() is the directory in which example .parquet files from pf_backward_sampler_v() are stored;

dat_pfp() reads an example output from pf_path();

dat_coa() reads an example output from coa();

See also

Author

Edward Lavender

Examples

# Load inbuilt datasets
dat_dlist() |> summary()
#>           Length Class  Mode
#> data      3      -none- list
#> spatial   1      -none- list
#> pars      2      -none- list
#> algorithm 0      -none- list
dat_obs() |> dplyr::glimpse()
#> Rows: 25
#> Columns: 12
#> $ timestep                 <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14…
#> $ timestamp                <dttm> 2016-03-17 01:50:00, 2016-03-17 01:52:00, 20…
#> $ date                     <chr> "2016-03-17", "2016-03-17", "2016-03-17", "20…
#> $ detection_id             <int> 1, 2, 3, 3, 4, 5, 5, 6, 7, 8, 9, 10, 11, 12, …
#> $ detection                <int> 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ receiver_id              <list> 26, 26, 26, <NULL>, 26, 26, <NULL>, 26, 26, …
#> $ receiver_id_next         <list> 26, 26, 26, 26, 26, 26, 26, 26, 26, 26, 26, …
#> $ mobility                 <dbl> 500, 500, 500, 500, 500, 500, 500, 500, 500, …
#> $ buffer_past              <dbl> 500, 500, 500, 500, 500, 500, 500, 500, 500, …
#> $ buffer_future            <dbl> 500, 500, 1000, 500, 500, 1000, 500, 500, 500…
#> $ buffer_future_incl_gamma <dbl> 1250, 1250, 1750, 1250, 1250, 1750, 1250, 125…
#> $ depth                    <dbl> 73.78, 73.32, 73.32, 73.32, 73.55, 68.70, 72.…
dat_pff() |> summary()
#>             Length Class      Mode   
#> history     25     -none-     list   
#> path         0     -none-     NULL   
#> diagnostics  7     data.table list   
#> internal     1     -none-     list   
#> convergence  1     -none-     logical
#> time         1     -none-     list   
dat_pfbk() |> summary()
#>         Length Class  Mode
#> history 25     -none- list
#> path     0     -none- NULL
#> time     3     -none- list
dat_pfbs() |> summary()
#>         Length Class  Mode
#> history 25     -none- list
#> path     0     -none- NULL
#> time     3     -none- list
dat_pfp() |> summary()
#>     path_id          timestep     cell_id          cell_x      
#>  Min.   :  1.00   Min.   : 1   Min.   :35857   Min.   :708142  
#>  1st Qu.: 25.75   1st Qu.: 7   1st Qu.:37374   1st Qu.:709042  
#>  Median : 50.50   Median :13   Median :37756   Median :709242  
#>  Mean   : 50.50   Mean   :13   Mean   :37720   Mean   :709219  
#>  3rd Qu.: 75.25   3rd Qu.:19   3rd Qu.:38139   3rd Qu.:709342  
#>  Max.   :100.00   Max.   :25   Max.   :39852   Max.   :710142  
#>      cell_y            cell_z      
#>  Min.   :6252107   Min.   : 35.75  
#>  1st Qu.:6253007   1st Qu.: 49.43  
#>  Median :6253207   Median : 68.53  
#>  Mean   :6253227   Mean   : 75.07  
#>  3rd Qu.:6253407   3rd Qu.: 96.39  
#>  Max.   :6254207   Max.   :150.05  
dat_coa() |> summary()
#>    timestamp                       coa_x            coa_y        
#>  Min.   :2016-03-17 01:00:00   Min.   :702842   Min.   :6252991  
#>  1st Qu.:2016-03-24 01:00:00   1st Qu.:706442   1st Qu.:6253107  
#>  Median :2016-03-27 21:00:00   Median :709057   Median :6253174  
#>  Mean   :2016-03-28 04:02:24   Mean   :707749   Mean   :6257412  
#>  3rd Qu.:2016-03-29 13:00:00   3rd Qu.:709242   3rd Qu.:6265407  
#>  Max.   :2016-04-16 05:00:00   Max.   :709430   Max.   :6267707  

# Directory of inbuilt parquet files from `pf_forward()`
dat_pff_src()
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history"
dat_pff_src(.folder = "history")
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history"
dat_pff_src(.folder = "diagnostics")
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/diagnostics"
pf_files(dat_pff_src())
#> [[1]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/1.parquet"
#> 
#> [[2]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/2.parquet"
#> 
#> [[3]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/3.parquet"
#> 
#> [[4]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/4.parquet"
#> 
#> [[5]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/5.parquet"
#> 
#> [[6]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/6.parquet"
#> 
#> [[7]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/7.parquet"
#> 
#> [[8]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/8.parquet"
#> 
#> [[9]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/9.parquet"
#> 
#> [[10]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/10.parquet"
#> 
#> [[11]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/11.parquet"
#> 
#> [[12]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/12.parquet"
#> 
#> [[13]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/13.parquet"
#> 
#> [[14]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/14.parquet"
#> 
#> [[15]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/15.parquet"
#> 
#> [[16]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/16.parquet"
#> 
#> [[17]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/17.parquet"
#> 
#> [[18]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/18.parquet"
#> 
#> [[19]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/19.parquet"
#> 
#> [[20]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/20.parquet"
#> 
#> [[21]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/21.parquet"
#> 
#> [[22]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/22.parquet"
#> 
#> [[23]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/23.parquet"
#> 
#> [[24]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/24.parquet"
#> 
#> [[25]]
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/forward/history/25.parquet"
#> 

# Directory of inbuilt parquet files from `pf_backward_killer()`
dat_pfbk_src()
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/backward/killer"

# Directory of inbuilt parquet files frim `pf_backward_sampler_v()`
dat_pfbs_src()
#> [1] "/private/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T/RtmpITfXS3/temp_libpath671648e30fc/patter/extdata/acpf/backward/sampler"