Datasets: example algorithm outputs

These functions load example outputs from key patter functions. They are included in the package to streamline function examples and tests.

dat_obs()

dat_pff()

dat_pfbk()

dat_pfp()

dat_coa()

Source

For full details on the algorithm parameters used to generate these datasets, see https://github.com/edwardlavender/patter/blob/main/data-raw/003-add-data-algorithms.R.

Details

dat_pff() reads an example output from pf_setup_obs();

dat_pff() reads an example output from pf_forward();

dat_pfbk() reads an example output from pf_backward_killer();

dat_pfp() reads an example output from pf_path();

dat_coa() reads an example output from coa();

Author

Edward Lavender

Examples

dat_obs() |> dplyr::glimpse()
#> Rows: 25
#> Columns: 12
#> $ timestep                 <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14…
#> $ timestamp                <dttm> 2016-03-17 01:50:00, 2016-03-17 01:52:00, 20…
#> $ date                     <chr> "2016-03-17", "2016-03-17", "2016-03-17", "20…
#> $ detection_id             <int> 1, 2, 3, 3, 4, 5, 5, 6, 7, 8, 9, 10, 11, 12, …
#> $ detection                <int> 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ receiver_id              <list> 26, 26, 26, <NULL>, 26, 26, <NULL>, 26, 26, …
#> $ receiver_id_next         <list> 26, 26, 26, 26, 26, 26, 26, 26, 26, 26, 26, …
#> $ mobility                 <dbl> 500, 500, 500, 500, 500, 500, 500, 500, 500, …
#> $ buffer_past              <dbl> 500, 500, 500, 500, 500, 500, 500, 500, 500, …
#> $ buffer_future            <dbl> 500, 500, 1000, 500, 500, 1000, 500, 500, 500…
#> $ buffer_future_incl_gamma <dbl> 1250, 1250, 1750, 1250, 1250, 1750, 1250, 125…
#> $ depth                    <dbl> 73.78, 73.32, 73.32, 73.32, 73.55, 68.70, 72.…
dat_pff() |> summary()
#>             Length Class      Mode   
#> history     25     -none-     list   
#> diagnostics  8     data.table list   
#> internal     1     -none-     list   
#> convergence  1     -none-     logical
#> time         1     -none-     list   
dat_pfbk() |> summary()
#>         Length Class  Mode
#> history 25     -none- list
#> time     3     -none- list
dat_pfp() |> summary()
#>     path_id          timestep     cell_id          cell_x      
#>  Min.   :  1.00   Min.   : 1   Min.   :36043   Min.   :708142  
#>  1st Qu.: 25.75   1st Qu.: 7   1st Qu.:37757   1st Qu.:709142  
#>  Median : 50.50   Median :13   Median :37949   Median :709242  
#>  Mean   : 50.50   Mean   :13   Mean   :38032   Mean   :709220  
#>  3rd Qu.: 75.25   3rd Qu.:19   3rd Qu.:38330   3rd Qu.:709342  
#>  Max.   :100.00   Max.   :25   Max.   :40036   Max.   :710242  
#>      cell_y            cell_z      
#>  Min.   :6252007   Min.   : 14.66  
#>  1st Qu.:6252907   1st Qu.: 46.54  
#>  Median :6253107   Median : 56.54  
#>  Mean   :6253063   Mean   : 64.10  
#>  3rd Qu.:6253207   3rd Qu.: 68.53  
#>  Max.   :6254107   Max.   :149.88  
dat_coa() |> summary()
#>    timestamp                       coa_x            coa_y        
#>  Min.   :2016-03-17 01:00:00   Min.   :702842   Min.   :6252991  
#>  1st Qu.:2016-03-24 01:00:00   1st Qu.:706442   1st Qu.:6253107  
#>  Median :2016-03-27 21:00:00   Median :709057   Median :6253174  
#>  Mean   :2016-03-28 04:02:24   Mean   :707749   Mean   :6257412  
#>  3rd Qu.:2016-03-29 13:00:00   3rd Qu.:709242   3rd Qu.:6265407  
#>  Max.   :2016-04-16 05:00:00   Max.   :709430   Max.   :6267707