patter set up: datasets

This function is used to set up movement datasets, spatial datasets and associated parameters for all patter functions. In general, pat_setup_data() should be called before other routines.

pat_setup_data(
  .acoustics = NULL,
  .moorings = NULL,
  .services = NULL,
  .archival = NULL,
  .bathy = NULL,
  .lonlat = NULL
)

Arguments

.acoustics, .services, .archival, .moorings

(optional) Movement datasets.

.acoustics contains the detection time series (see dat_acoustics for an example);

.moorings contains receiver deployments (see dat_moorings) for an example);

.services contains receiver servicing information (times during the deployment period of a receiver when it was not active due to servicing);

.archival contains archival (depth) time series (see dat_archival for an example);

.bathy

(optional) A SpatRaster that defines a (bathymetry) grid over the study area.

.lonlat

(optional) A logical variable that defines whether spatial data (.moorings coordinates, .bathy) is in longitude/latitude format (TRUE) or planar (FALSE).

Value

The function returns a named list with three elements:

data---a named list, with one element for each movement dataset;

spatial---a named list, with one element for each spatial dataset;

pars---a named list, with required parameters;

algorithm---an empty list that serves as a placeholder for algorithm layers required by pf_forward() and associated routines;

Details

The preparation of datasets for patter is a one-off inconvenience. See the check_dlist() function documentation for full details as to the required properties of each input dataset. All requirements are kept to a minimum and straightforward to address. To minimise inconvenience, all inputs are optional in pat_setup_data(). If you are only interested in specific routine(s), check the associated documentation to see which inputs are essential for those routine(s). The output of pat_setup_data() is widely passed to downstream patter functions, which streamlines the API, documentation and internal code.

If .lonlat = TRUE, downstream functions may require the geosphere package.

Author

Edward Lavender

Examples

dlist <- pat_setup_data(.acoustics = dat_acoustics,
                        .services = NULL,
                        .archival = dat_archival,
                        .moorings = dat_moorings,
                        .bathy = dat_gebco(),
                        .lonlat = FALSE)
#> Multiple individuals detected in acoustic data.
#> Acoustic time stamps should be ordered chronologically.
#> Multiple individuals detected in archival data.
#> Archival time stamps should be ordered chronologically.
#> Archival time steps are not regularly spaced.
summary(dlist)
#>           Length Class  Mode
#> data      3      -none- list
#> spatial   1      -none- list
#> pars      2      -none- list
#> algorithm 0      -none- list