PF: list files

pf_files() function creates an ordered list of the parquet files contains particle samples (e.g., from pf_forward() or pf_backward_killer()). pf_files_size() calculates the total size of all files.

pf_files(.sink, ...)

pf_files_size(.sink, .unit = c("MB", "GB", "TB"))

Arguments

.sink

A character string that defines the directory in which files are located.

...

For pf_files(), ... is a placeholder for additional arguments passed to list.files(), such as pattern, excluding full.names.

.unit

For pf_files_size(), .unit is a character string that defines the units of the output (MB, GB, TB).

Value

pf_files() returns an ordered list of file paths.

pf_files_size() returns a number.

Author

Edward Lavender

Examples

# Use `pf_files()` to list files from `pf_forward()`
pff_folder <- system.file("extdata", "acpf", "forward", "history",
                          package = "patter", mustWork = TRUE)
files <- pf_files(pff_folder)

# Use `pf_files()` to list files from `pf_backward_killer()`
pfbk_folder <- system.file("extdata", "acpf", "backward", "killer",
                           package = "patter", mustWork = TRUE)
files <- pf_files(pfbk_folder)

# Use `pf_files_size()` to calculate file size
pf_files_size(pfbk_folder)
#> [1] 0.101539
pf_files_size(pfbk_folder, .unit = "GB")
#> [1] 0.000101539
pf_files_size(pfbk_folder, .unit = "TB")
#> [1] 1.01539e-07