PF: calculate particle likelihoods

PF: calculate particle likelihoods

acs_filter_land(.particles, .bathy)

acs_filter_container(.particles, .moorings, .receivers, .threshold)

pf_lik_ac(.particles, .obs, .t, .detection_overlaps, .detection_kernels)

pf_lik_update(.particles, .obs, .t, .bathy, .update_ac)

pf_lik(
  .particles,
  .obs,
  .t,
  .bathy,
  .is_land,
  .moorings,
  .detection_overlaps,
  .detection_kernels,
  .update_ac,
  .trial = NA_integer_
)

Arguments

.moorings

For acs_filter_container,

.detection_overlaps, .detection_kernels,

For pf_lik_ac

.update_ac

For pf_lik_update

.is_land

For acs_filter_land,

Details

TO DO provide likelihood so you can stack them how you like e.g. filters first then DC or AC each function should accept particles, return filtered particles & likelihood column

Author

Edward Lavender