PF: pf_particles objects

An S3 class that defines the named list returned by pf_forward() and pf_backward_killer().

Structure

pf_particles is a label used to denote outputs from selected functions in patter. The structure of this class is not strictly defined and primarily exists to streamline documentation. At the time of writing, pf_particles objects may comprise the following elements:

history---a list of particle samples;

diagnostics---a data.table of diagnostics;

internal---a list of internal objects;

convergence---a logical value that defines convergence success;

time---a list of timings;

history

history is a list with one element for each time step. Each element is a data.table that contains location (particle) samples. Each row is a specific sample. The columns in this table depend on pf_forward()'s .likelihood function(s) and the .return$cols argument (see pf_opt_return()). By default, the following columns are computed:

timestep---an integer vector that defines the time step;

cell_past---an integer vector that identifies the grid cells of previous particle samples;

x_past, y_past---numeric vectors that define the coordinates of previous particle samples;

cell_now---an integer vector that identifies the grid cells of current particle samples;

x_now, y_now---numeric vectors that define the coordinates of current particle samples;

lik---a numeric vector of likelihoods;

weight---a numeric vector of sampling weights;

Coordinates are defined on the grid by the default proposal functions pf_rpropose_kick() and pf_rpropose_reachable().

The likelihood column (lik) is the product of the likelihood scores from each likelihood function. (Successive likelihood functions should update this column.)

The weights column (weight) is equal to the lik column, for particles generated by stochastic kicks, or the normalised product of the lik and dens columns, in the case of directed sampling (see below).

Additional columns may be included in history if computed by inputted likelihood functions. At the time of writing, most default likelihood functions calculate likelihoods without defining additional columns. However, pf_lik_dc() adds bathy column, with values extracted from .dlist$spatial$bathy. If directed sampling is used, bathy and dens columns are included where required (after the first time step). dens is numeric vector that defines the probability density of moving from (x_past, y_past) to (x_now, y_now).

Since columns are recorded only when required, not all columns are recorded at each time step. To combine data.tables, use .pf_history_dt().

If .record$save = FALSE, .history is an empty list and the individual data.tables are written to file in {.record$sink}/history/ (see pf_opt_record()).

pf_backward_killer() simply drops dead-ends (rows) from the time series and leaves the columns from pf_forward() intact (unless .record$cols is supplied).

diagnostics

In pf_forward(), we track metrics of particle diversity (hereafter, 'particle diagnostics') through time for analysis of convergence issues, sampling sufficiency and other properties (see pf_diag). Particle diagnostics are recorded for each trial and component of a stochastic process.

In all cases, at the start of each time step, we calculate baseline diagnostics.

Thereafter, for stochastic kicks (implemented via .pf_particles_kick()), we record:

Proposal diagnostics, following stochastic kicks;

Likelihood diagnostics, following the evaluation of each likelihood function;

Sampling diagnostics, following sampling;

For directed sampling (implemented via .pf_particles_sampler()), we record:

Proposal/likelihood diagnostics, following a combined proposal plus likelihood step;

Sampling diagnostics, following sampling;

Diagnostics are tracked for all trials, reversions and reruns (see pf_opt_trial()) and never replaced. (This is so we can check, for instance, how many trials of a process, such as stochastic kick-based sampling, are required at each time step to generate a sufficient number of valid particles and, in so doing, improve algorithm efficiency.)

If .record$save = TRUE, particle diagnostics are collated in a single data.table, with the following columns:

iter_m---an integer that identifies 'manual' iterations. iter_m is incremented when .rerun is used.

iter_i---an integer that identifies 'internal' iterations. iter_i is incremented when the algorithm jumps back to an earlier time step (see pf_opt_record()).

timestep---an integer vector that defines time steps.

component---a character vector that defines algorithm components:

base---the start of each time step (a baseline), after the first time step;

proposal---the proposal step (at the first time step and for stochastic kicks only);

lik_*---the likelihood step;

For stochastic kicks, particle diagnostics are given following each (i) likelihood evaluation and the components are named by paste0("lik-", names(.likelihood)[i]);

For directed sampling, particle diagnostics for the proposal and likelihood stages are combined under a single lik-sampler label and not split by likelihood component;

sample_*---the sampling step, denoted as:

sample-origin at the first time step;

sample-kick for stochastic kick-based sampling;

sample-directed for directed sampling;

trial---an integer vector that defines the trial of a stochastic process, if applicable (see pf_opt_trial()).

n, nu, ess---particle diagnostics immediately following completion of each algorithm component, computed by internal .pf_diag_*() functions:

n---an integer vector that defines the number of locations;

nu---an integer vector that defines the number of unique locations (see .pf_diag_nu()).

ess---a numeric vector that defines the effective sample size (see .pf_diag_ess()).

In pf_forward(), if .record$save = FALSE, diagnostics is NULL and the individual data.tables are written to file in {.record$sink}/diagnostics/ (see pf_opt_record()). To collate diagnostics from file, use pf_forward_diagnostics().

At the time of writing, diagnostics are not computed on the fly by pf_backward_*() and are instead calculated post-hoc by pf_backward_killer_diagnostics().

internal

For pf_forward(), internal is named list of internal objects for algorithm reruns (see .rerun and .pf_startup()).

convergence

For pf_forward(), convergence is a logical variable that defines whether or not the algorithm successfully converged (i.e., reached the end of the time series).

time

A list that records the start, end and duration of algorithm runs (as POSIXct and difftime objects). For pf_forward(), this is a nested list, with one element for each manual algorithm run (see .rerun).

See also

Author

Edward Lavender