PF: proposal functions

These are proposal functions for pf_forward().

pf_rpropose_kick(
  .particles,
  .obs,
  .t,
  .dlist,
  .sim_length = rlen,
  .sim_angle = rangrw,
  ...
)

pf_rpropose_reachable(.particles, .obs, .t, .dlist)

pf_dpropose(.particles, .obs, .t, .dlist, ...)

Arguments

.particles, .obs, .t, .dlist

Required arguments for all propose functions.

.particles---a data.table of particle samples from the previous time step;

.obs---the .obs data.table from pf_forward();

.t---an integer that defines the time step;

.dlist---the .dlist list from pf_forward();

.sim_length, .sim_angle

Additional arguments for pf_rpropose_kick(). These are functions that simulate step lengths and turning angles respectively. Each function must accept a .n argument that defines the number of simulated values.

...

Additional arguments.

For pf_rpropose_kick(), ... is passed to .sim_length and .sim_angle.

For pf_dpropose(), ... is passed to dstep().

Details

In pf_forward(), proposal functions are used to generate (propose) new, candidate locations for the individual's position, contingent upon previous positions (particle samples). Proposal locations are generated from previous locations via stochastic kicks and directed sampling.

The .rpropose argument in pf_forward() expects a stochastic-kick routine and pf_rpropose_kick() is the default. This is used to simulate proposal locations by 'kicking' particles into new locations as specified by a movement model. Custom functions that accept the .particles, .obs, .t and .dlist arguments can be provided.

If stochastic kicks fail to produce a sufficient number of valid particle samples, pf_rpropose_reachable() may be called under-the-hood for directed sampling. For selected particles, this function identifies the set of reachable locations, for which likelihood calculations and sampling are then implemented (see pf_forward()).

pf_dpropose() calculates the probability density of movements into proposal locations and is required for directed sampling.

See also

Author

Edward Lavender