PF: proposal functions

These are proposal functions and supporting routines for pf_forward().

rkick(.xy0, .rstep = rstep, ..., .obs, .t, .dlist)

dkick(.xy0, .xy1, .dstep = dstep, ..., .obs, .t, .dlist)

pf_rpropose_kick(.particles, .obs, .t, .dlist, .rkick = rkick, ...)

pf_rpropose_reachable(.particles, .obs, .t, .dlist)

pf_dpropose(.particles, .obs, .t, .dlist, .dkick = dkick, ...)

Arguments

.xy0, .xy1, .rstep, .dstep

Arguments for .rkick (and rkick()) and .dkick (and dkick()).

.xy0---a two-column object (matrix, data.frame, data.table) of accepted (x, y) coordinates from the previous time step. In .rkick, movement is simulated from .xy0 into new locations;

.xy1---For .dkick and dkick(), .xy1 is a two-column matrix of accepted coordinates for the current time step;

.rstep or .dstep and ...---functions that simulate new locations or calculate the probability density of movements between locations (such as rstep() and dstep()) and additional arguments passed to those functions. Both functions must accept:

.xy0 and .xy1---as above;

...---arguments passed from above;

.lonlat---a logical variable that defines whether or not coordinates are in longitude/latitude format or planar (extracted from .dlist$pars$lonlat by default);

.obs, .t, .dlist

Required arguments for all functions.

.obs---the .obs data.table from pf_forward();

.t---an integer that defines the time step;

.dlist---the .dlist list from pf_forward();

For rkick() and dkick(), .dlist must contain .dlist$spatial$bathy;

For pf_rpropose_kick(), .dlist must contain .dlist$spatial$bathy;

For pf_rpropose_reachable(), .dlist must contain .dlist$spatial$bathy;

.particles

Additional required arguments for all .rpropose and .dpropose functions:

.particles---a data.table of particle samples from the previous time step;

.rkick, .dkick, ...

Additional arguments for pf_rpropose_kick() and pf_dpropose() respectively.

.rkick is a function, like rkick(), that simulates new locations;

.dkick is a function, like dkick(), that calculates the probability density of movements between locations;

At the time of writing, these functions must accept the .xy0, .xy1, .rstep/.dstep and ... arguments.

Details

In pf_forward(), proposal functions are used to generate (propose) new, candidate locations for the individual's position, contingent upon previous positions (particle samples). Proposal locations are generated from previous locations via stochastic kicks and directed sampling.

The .rpropose argument in pf_forward() expects a stochastic-kick routine and pf_rpropose_kick() is the default. This is used to simulate proposal locations by 'kicking' particles into new locations as specified by a movement model (.rkick).

pf_rpropose_kick() is a simple wrapper for rkick() that passes the .particles data.table to the function as required and then updates .particles with simulated locations. Simulated coordinates are redefined on the grid (.dlist$spatial$bathy) and grid cell IDs are included in the output. (Proposals beyond the grid are silently dropped.) rkick() itself is a wrapper for .rstep = rstep that accepts (but silently ignores) the .obs, .t and .dlist objects (except .dlist$pars$lonlat, which is passed to the .lonlat argument of .rstep). In pf_rpropose_kick, arguments passed via ... are passed to .rkick, which under default settings means .rstep = = rstep (i.e., .rlen and .rang or additional arguments passed to those arguments). This facilitates simulation of a wide variety of random walks. At the time of writing, correlated random walks are not easy for users to implement, but this should improve in future.

In pf_forward(), if stochastic kicks fail to produce a sufficient number of valid particle samples, pf_rpropose_reachable() may be called under-the-hood for directed sampling (see pf_opt_trial()). For selected particles, this function identifies the set of reachable locations. We evaluate the likelihood of reachable locations and the probability density of moving into each location and then sample locations according to the (normalised) product of these two variables. The .dpropose argument in pf_forward() is required to calculate the probability density of moving between locations. .dpropose is a function that must accept the usual .particles, .obs, .t, .dlist and ... arguments and return a data.table, for the subset of valid locations, with a dens column that defines probability densities. The function must be able to handle empty data.tables, which are passed down the call stack if all proposal (reachable) locations have zero likelihood.

pf_dpropose() is the default .dpropose routine. Under default settings, this is a simple wrapper for dkick() that handles empty data.tables or passes the relevant coordinate columns, for the accepted locations from the previous time step ((x_past, y_past) in .particles) and the proposal locations for the current time step ((x_now, y_now) in .particles) to .dkick. Under default settings, .dkick = dkick. dkick() wraps itself wraps a .dstep function such as dstep(), accepting (but silently ignoring) the .obs, .t and .dlist arguments (except .dlist$pars$lonlat). In this situation, ... arguments are passed to .dstep.

In pf_forward(), use .rpropose and .dpropose to write fully custom routines, if required. Use .rargs and .dargs to customise the default routines. For movement models that require .obs, .t and .dlist, use custom .rkick and .dkick functions. Otherwise, you can simply customise .rstep and .dstep (for instance, by revising the models used to simulate step lengths and turning angles, or the parameters passed to those models).

See also

Author

Edward Lavender