PF set up: list files for PF

This function creates an ordered list of files for PF, either from an AC* algorithm (i.e., acs() or dc()) for the forward simulation (pf_forward()), or from the forward simulation (pf_forward()) for the backward pass (pf_backward()).

pf_setup_files(.root, ...)

Arguments

.root

A string that defines the directory in which files are located.

...

Additional arguments passed to list.files(), such as pattern, excluding full.names.

Value

The function returns an ordered list of file paths.

Author

Edward Lavender

Examples

# Quick implementation of AC algorithm
obs    <- dat_obs()[1:10, ]
gebco  <- dat_gebco()
con    <- tempdir()
ac_folder <- file.path(con, "patter", "ac")
dir.create(ac_folder, recursive = TRUE)
out_ac <-
  acs(obs,
      .bathy = gebco,
      .detection_overlaps = dat_overlaps(),
      .detection_kernels = dat_kernels(),
      .write_record = list(filename = ac_folder, overwrite = TRUE))
#> patter::acs() called (@ 2023-10-13 16:47:20)... 
#> ... Time step 1: 
#> ... ... Detection 1: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 2: 
#> ... ... Detection 2: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 3: 
#> ... ... Detection 3: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 4: 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 5: 
#> ... ... Detection 4: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 6: 
#> ... ... Detection 5: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 7: 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 8: 
#> ... ... Detection 6: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 9: 
#> ... ... Detection 7: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 10: 
#> ... ... Detection 8: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> patter::acs() call ended (@ 2023-10-13 16:47:20). 

# List files for pf_forward()
files <- pf_setup_files(ac_folder)

# Quick implementation of pf_forward()
pff_folder <- file.path(con, "patter", "pf", "forward")
dir.create(pff_folder, recursive = TRUE)
out_pff <- pf_forward(obs,
                      .record = files,
                      .kick = pf_kick,
                      .bathy = gebco,
                      .write_history = list(sink = pff_folder))
#> patter::pf_forward() called (@ 2023-10-13 16:47:20)... 
#> ... Time step 1: 
#> ... ... Reading record... 
#> ... ... Sampling starting particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 2: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 3: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 4: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 5: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 6: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 7: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 8: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 9: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 10: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> patter::pf_forward() call ended (@ 2023-10-13 16:47:21). 

# List files for pf_backward()
files <- pf_setup_files(pff_folder)

# Quick implementation of pf_backward()
pfb_folder <- file.path(con, "patter", "pf", "backward")
dir.create(pfb_folder, recursive = TRUE)
out_pfb <- pf_backward(files,
                       .write_history = list(sink = pfb_folder))
#> patter::pf_backward() called (@ 2023-10-13 16:47:21)... 
#> ... Time step 10: 
#> ... ... Reading `.history[[10]]`... 
#> ... ... Reading `.history[[9]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[10]]`... 
#> ... Time step 9: 
#> ... ... Reading `.history[[8]]`... 
#> ... ... Cleaning `.history[[8]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 100 rows in `.history[[t]]`... 
#> ... ... ... Filtering 20 dead ends (80 remain)... 
#> ... ... ... Output: 80 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[9]]`... 
#> ... Time step 8: 
#> ... ... Reading `.history[[7]]`... 
#> ... ... Cleaning `.history[[7]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 80 rows in `.history[[t]]`... 
#> ... ... ... Filtering 44 dead ends (56 remain)... 
#> ... ... ... Output: 56 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[8]]`... 
#> ... Time step 7: 
#> ... ... Reading `.history[[6]]`... 
#> ... ... Cleaning `.history[[6]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 56 rows in `.history[[t]]`... 
#> ... ... ... Filtering 43 dead ends (57 remain)... 
#> ... ... ... Output: 57 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[7]]`... 
#> ... Time step 6: 
#> ... ... Reading `.history[[5]]`... 
#> ... ... Cleaning `.history[[5]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 57 rows in `.history[[t]]`... 
#> ... ... ... Filtering 38 dead ends (62 remain)... 
#> ... ... ... Output: 62 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[6]]`... 
#> ... Time step 5: 
#> ... ... Reading `.history[[4]]`... 
#> ... ... Cleaning `.history[[4]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 62 rows in `.history[[t]]`... 
#> ... ... ... Filtering 58 dead ends (42 remain)... 
#> ... ... ... Output: 42 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[5]]`... 
#> ... Time step 4: 
#> ... ... Reading `.history[[3]]`... 
#> ... ... Cleaning `.history[[3]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 42 rows in `.history[[t]]`... 
#> ... ... ... Filtering 50 dead ends (50 remain)... 
#> ... ... ... Output: 50 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[4]]`... 
#> ... Time step 3: 
#> ... ... Reading `.history[[2]]`... 
#> ... ... Cleaning `.history[[2]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 50 rows in `.history[[t]]`... 
#> ... ... ... Filtering 33 dead ends (67 remain)... 
#> ... ... ... Output: 67 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[3]]`... 
#> ... Time step 2: 
#> ... ... Reading `.history[[1]]`... 
#> ... ... Cleaning `.history[[1]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 67 rows in `.history[[t]]`... 
#> ... ... ... Filtering 45 dead ends (55 remain)... 
#> ... ... ... Output: 55 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[2]]`... 
#> ... Time step 1: 
#> ... ... Recording (cleaned) outputs for `.history[[1]]`... 
#> patter::pf_backward() call ended (@ 2023-10-13 16:47:21). 

# List files from pf_backward()
pf_setup_files(pfb_folder)
#> [[1]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//Rtmp8CETPa/patter/pf/backward/1.parquet"
#> 
#> [[2]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//Rtmp8CETPa/patter/pf/backward/2.parquet"
#> 
#> [[3]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//Rtmp8CETPa/patter/pf/backward/3.parquet"
#> 
#> [[4]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//Rtmp8CETPa/patter/pf/backward/4.parquet"
#> 
#> [[5]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//Rtmp8CETPa/patter/pf/backward/5.parquet"
#> 
#> [[6]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//Rtmp8CETPa/patter/pf/backward/6.parquet"
#> 
#> [[7]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//Rtmp8CETPa/patter/pf/backward/7.parquet"
#> 
#> [[8]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//Rtmp8CETPa/patter/pf/backward/8.parquet"
#> 
#> [[9]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//Rtmp8CETPa/patter/pf/backward/9.parquet"
#> 
#> [[10]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//Rtmp8CETPa/patter/pf/backward/10.parquet"
#> 

# Clean up
unlink(file.path(con, "patter"), recursive = TRUE)