PF set up: list files for PF

This function creates an ordered list of files for PF, either from an AC* algorithm (i.e., acs() or dc()) for the forward simulation (pf_forward_1()), or from the forward simulation (pf_forward_*()) for the backward pass (pf_backward()).

pf_setup_files(.root, ...)

Arguments

.root

A string that defines the directory in which files are located.

...

Additional arguments passed to list.files(), such as pattern, excluding full.names.

Value

The function returns an ordered list of file paths.

See also

Author

Edward Lavender

Examples

# Quick implementation of AC algorithm
obs    <- dat_obs()[1:10, ]
gebco  <- dat_gebco()
con    <- tempdir()
ac_folder <- file.path(con, "patter", "ac")
dir.create(ac_folder, recursive = TRUE)
out_ac <-
  acs(obs,
      .bathy = gebco,
      .detection_overlaps = dat_overlaps(),
      .detection_kernels = dat_kernels(),
      .write_record = list(filename = ac_folder, overwrite = TRUE))
#> patter::acs() called (@ 2023-11-24 20:46:55.160022)... 
#> ... Time step 1: 
#> ... ... Detection 1: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 2: 
#> ... ... Detection 2: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 3: 
#> ... ... Detection 3: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 4: 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 5: 
#> ... ... Detection 4: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 6: 
#> ... ... Detection 5: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 7: 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 8: 
#> ... ... Detection 6: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 9: 
#> ... ... Detection 7: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining location given future... 
#> ... ... - Defining future kernel... 
#> ... ... - Defining future buffer... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> ... Time step 10: 
#> ... ... Detection 8: 
#> ... ... * Defining location given data... 
#> ... ... * Defining location given past... 
#> ... ... * Defining present... 
#> ... ... * Updating record ... 
#> ... ... * Moving on... 
#> patter::acs() call ended (@ 2023-11-24 20:46:55.609948). 

# List files for pf_forward_*()
files <- pf_setup_files(ac_folder)

# Quick implementation of pf_forward_1()
pff_folder <- file.path(con, "patter", "pf", "forward")
dir.create(pff_folder, recursive = TRUE)
out_pff <- pf_forward_1(obs,
                        .record = files,
                        .kick = pf_kick,
                        .bathy = gebco,
                        .write_history = list(sink = pff_folder))
#> patter::pf_forward_1() called (@ 2023-11-24 20:46:55.615143)... 
#> ... Time step 1: 
#> ... ... Reading record... 
#> ... ... Sampling starting particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 2: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 3: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 4: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 5: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 6: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 7: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 8: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 9: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> ... ... Kicking particles into new locations... 
#> ... ... Moving on... 
#> ... Time step 10: 
#> ... ... Reading record... 
#> ... ... Extracting particle weights... 
#> ... ... Resampling particles... 
#> patter::pf_forward_1() call ended (@ 2023-11-24 20:46:55.697462). 

# List files for pf_backward()
files <- pf_setup_files(pff_folder)

# Quick implementation of pf_backward()
pfb_folder <- file.path(con, "patter", "pf", "backward")
dir.create(pfb_folder, recursive = TRUE)
#> Warning: '/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward' already exists
out_pfb <- pf_backward(files,
                       .write_history = list(sink = pfb_folder))
#> Warning: `.write_history$sink` ('/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward') is not an empty directory.
#> patter::pf_backward() called (@ 2023-11-24 20:46:55.7019)... 
#> ... Time step 10: 
#> ... ... Reading `.history[[10]]`... 
#> ... ... Reading `.history[[9]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[10]]`... 
#> ... Time step 9: 
#> ... ... Reading `.history[[8]]`... 
#> ... ... Cleaning `.history[[8]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 100 rows in `.history[[t]]`... 
#> ... ... ... Filtering 14 dead ends (86 remain)... 
#> ... ... ... Output: 86 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[9]]`... 
#> ... Time step 8: 
#> ... ... Reading `.history[[7]]`... 
#> ... ... Cleaning `.history[[7]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 86 rows in `.history[[t]]`... 
#> ... ... ... Filtering 39 dead ends (61 remain)... 
#> ... ... ... Output: 61 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[8]]`... 
#> ... Time step 7: 
#> ... ... Reading `.history[[6]]`... 
#> ... ... Cleaning `.history[[6]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 61 rows in `.history[[t]]`... 
#> ... ... ... Filtering 27 dead ends (73 remain)... 
#> ... ... ... Output: 73 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[7]]`... 
#> ... Time step 6: 
#> ... ... Reading `.history[[5]]`... 
#> ... ... Cleaning `.history[[5]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 73 rows in `.history[[t]]`... 
#> ... ... ... Filtering 21 dead ends (79 remain)... 
#> ... ... ... Output: 79 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[6]]`... 
#> ... Time step 5: 
#> ... ... Reading `.history[[4]]`... 
#> ... ... Cleaning `.history[[4]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 79 rows in `.history[[t]]`... 
#> ... ... ... Filtering 41 dead ends (59 remain)... 
#> ... ... ... Output: 59 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[5]]`... 
#> ... Time step 4: 
#> ... ... Reading `.history[[3]]`... 
#> ... ... Cleaning `.history[[3]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 59 rows in `.history[[t]]`... 
#> ... ... ... Filtering 39 dead ends (61 remain)... 
#> ... ... ... Output: 61 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[4]]`... 
#> ... Time step 3: 
#> ... ... Reading `.history[[2]]`... 
#> ... ... Cleaning `.history[[2]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 61 rows in `.history[[t]]`... 
#> ... ... ... Filtering 29 dead ends (71 remain)... 
#> ... ... ... Output: 71 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[3]]`... 
#> ... Time step 2: 
#> ... ... Reading `.history[[1]]`... 
#> ... ... Cleaning `.history[[1]]`... 
#> ... ... Identifying `cell_now` (for the previous step) that match `cell_past` (for the current step)... 
#> ... ... ... Input: 71 rows in `.history[[t]]`... 
#> ... ... ... Filtering 42 dead ends (58 remain)... 
#> ... ... ... Output: 58 rows in `.history[[t - 1]]`... 
#> ... ... Recording (cleaned) outputs for `.history[[2]]`... 
#> ... Time step 1: 
#> ... ... Recording (cleaned) outputs for `.history[[1]]`... 
#> patter::pf_backward() call ended (@ 2023-11-24 20:46:55.736344). 

# List files from pf_backward()
pf_setup_files(pfb_folder)
#> [[1]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/1.parquet"
#> 
#> [[2]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/2.parquet"
#> 
#> [[3]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/3.parquet"
#> 
#> [[4]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/4.parquet"
#> 
#> [[5]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/5.parquet"
#> 
#> [[6]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/6.parquet"
#> 
#> [[7]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/7.parquet"
#> 
#> [[8]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/8.parquet"
#> 
#> [[9]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/9.parquet"
#> 
#> [[10]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/10.parquet"
#> 
#> [[11]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/11.parquet"
#> 
#> [[12]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/12.parquet"
#> 
#> [[13]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/13.parquet"
#> 
#> [[14]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/14.parquet"
#> 
#> [[15]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/15.parquet"
#> 
#> [[16]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/16.parquet"
#> 
#> [[17]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/17.parquet"
#> 
#> [[18]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/18.parquet"
#> 
#> [[19]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/19.parquet"
#> 
#> [[20]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/20.parquet"
#> 
#> [[21]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/21.parquet"
#> 
#> [[22]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/22.parquet"
#> 
#> [[23]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/23.parquet"
#> 
#> [[24]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/24.parquet"
#> 
#> [[25]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/25.parquet"
#> 
#> [[26]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/26.parquet"
#> 
#> [[27]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/27.parquet"
#> 
#> [[28]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/28.parquet"
#> 
#> [[29]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/29.parquet"
#> 
#> [[30]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/30.parquet"
#> 
#> [[31]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/31.parquet"
#> 
#> [[32]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/32.parquet"
#> 
#> [[33]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/33.parquet"
#> 
#> [[34]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/34.parquet"
#> 
#> [[35]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/35.parquet"
#> 
#> [[36]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/36.parquet"
#> 
#> [[37]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/37.parquet"
#> 
#> [[38]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/38.parquet"
#> 
#> [[39]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/39.parquet"
#> 
#> [[40]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/40.parquet"
#> 
#> [[41]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/41.parquet"
#> 
#> [[42]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/42.parquet"
#> 
#> [[43]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/43.parquet"
#> 
#> [[44]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/44.parquet"
#> 
#> [[45]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/45.parquet"
#> 
#> [[46]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/46.parquet"
#> 
#> [[47]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/47.parquet"
#> 
#> [[48]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/48.parquet"
#> 
#> [[49]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/49.parquet"
#> 
#> [[50]]
#> [1] "/var/folders/nl/ygb3g6tj24z06jddbqqhj6hw0000gn/T//RtmpT0Pq16/patter/pf/backward/50.parquet"
#> 

# Clean up
unlink(file.path(con, "patter"), recursive = TRUE)