PF: set up movement datasets

This function builds a timeline of observations for particle filtering. The forward filter (pf_forward()) iterates over this timeline. At each time step, the algorithm proposes candidate locations for an individual, evaluates the likelihood of each observation (e.g., acoustic and/or archival data) in the timeline at the current time step and optionally (re)samples valid locations using a set of weights that incorporate the likelihood.

pf_setup_obs(
  .dlist,
  .trim = TRUE,
  .step,
  .period = NULL,
  .mobility,
  .verbose = getOption("patter.verbose")
)

Arguments

.dlist

A named list of data and parameters from pat_setup_data(). This function requires:

(optional) .dlist$data$acoustics, with the following columns: timestamp and receiver_id.

(optional) .dlist$data$archival, with the following columns: timestamp and depth.

(optional) If .dlist$data$acoustics is provided, .dlist$algorithm$detection_overlaps is required if there are receivers with overlapping detection containers;

At least one dataset must be provided. Data should be given for a single individual. Other datasets are not currently integrated within this function but can be included afterwards (see Details).

.trim

If both acoustic and archival data are supplied, .trim is logical variable that defines whether or not to trim the time series to the time period for which they overlap.

.step

An character, passed to lubridate::period(), lubridate::round_date() and seq() that defines the duration between sequential time steps (e.g., "2 mins").

.period

(optional) A length-two POSIXct vector that defines the start and end time for the timeline. If unprovided, this is defined internally according to inputted movement datasets and .trim:

If only acoustics = .dlist$data$acoustics is provided, .period is taken as range(acoustics$timestamp).

If only archival = .dlist$data$archival is provided, .period is taken as range(archival$timestamp).

If both datasets are provided and .trim = FALSE, .period is defined as range(c(acoustics$timestamp, archival$timestamp)).

If both datasets are provided and .trim = TRUE, .period is defined as the range in timestamps for the overlapping region.

.mobility

A constant that defines the maximum (Euclidean) distance the individual could move in .step. See the flapper::get_mvt_mobility_*() functions to estimate .mobility from acoustic and/or archival data.

.verbose

User output control (see patter-progress for supported options).

Value

The function returns a data.table with the following columns:

timestep---an integer that defines the time step;

timestamp---a regular sequences of POSIXct time stamps;

If acoustic data is provided, the following column(s) are also included:

array_id---an integer vector that uniquely distinguishes each array design;

detection---an integer that distinguishes the time steps at which detections were (1) or were not (0) recorded;

acoustics---a list of one-row acoustic observation matrices. In pf_forward(), this information is used to calculate the likelihood of acoustic observations at proposal locations (see pf_lik_ac()).

If detection = 1, the matrix contains one named column for each receiver that recorded a detection and each overlapping receiver. (It is not necessary to include all receivers in this matrix.) The entries 1 and 0 distinguish the receivers at which detections were recorded/not recorded. This information is used to calculate the likelihood of acoustic observations at time steps with detection(s).

If detection = 0, the list element is NULL. We do not need to store empty detection matrices because the likelihood of non detection at all operational receivers in a given array design is extracted from pre-computed layers (see acs_setup_detection_kernels()).

container---a list of acoustic container information. In pf_forward(), this information is used to eliminate proposal locations that are too far away from the location(s) of the next detection(s) (see acs_filter_container()). Each list contains:

coord---a two-column matrix of coordinates for the receiver(s) that recorded the next detection;

buffer---a numeric vector of the (Euclidean) distance from each pair of receiver coordinates within which the individual must have been located at each time step. The distance depends on the time between detections, the individual's .mobility and the receiver-specific detection range;

If archival data is provided, the following column(s) are also included:

depth---a number that defines the individual's depth (m) at each time step;

To calculate the likelihood of archival observations in pf_forward() by the default pf_lik_dc() routine, a depth envelope (controlled by depth_deep_eps and depth_shallow_eps columns) is also required (see pf_lik_dc()).

Details

This function defines the timeline of observations over which pf_forward() is implemented. The function implements the following routines:

Acoustic progressing (if applicable):

Acoustic time series are rounded to the nearest .step;

Duplicate detections (at the same receiver in the same time interval) are dropped;

Archival processing (if applicable):

Archival time series are rounded to the nearest .step;

Multiple records (in the same time interval) are not permitted (in this instance, you need to aggregate observations manually or reduce .step);

Alignment (if applicable):

Acoustic and archival time series are optionally trimmed to the time period for which they overlap;

Time series:

Acoustic and/or archival time series are mapped onto a regular time series;

Information required by pf_forward() (and built-in likelihood functions) is added;

At the time of writing, this function only supports acoustic and archival data. However, ancillary observations can easily be included in pf_forward() as extra columns in this data.table.

See also

Author

Edward Lavender

Examples

#### Define example datasets
# Define movement datasets for a single individual
acc  <- dat_acoustics[dat_acoustics$individual_id == 25, ]
arc  <- dat_archival[dat_archival$individual_id == 25, ]
# For example speed, focus on a subset of data
acc <- acc[as.Date(timestamp) == as.Date("2016-03-17")]
arc <- arc[as.Date(timestamp) == as.Date("2016-03-17")]
# Set up data list(see `?pat_setup_data()`)
dlist     <- pat_setup_data(.acoustics = acc,
                            .moorings = dat_moorings,
                            .archival = arc,
                            .bathy = dat_gebco(),
                            .lonlat = FALSE)
#> Warning: The deployment period(s) of some receiver(s) (`35`, `36`, `37`, `38`, `39`, `40`, `41`, `43`, `44`, `45`, `46`, `47`, `48`, `49`, `50`, `51`, `52`, `53`, `54`, `55`, `56`, `57`) in `.moorings` are entirely outside the range of acoustic observations. Consider excluding these receivers for improved efficiency (e.g., in `acs_setup_detection_kernels()`, which pre-calculates the likelihood of non detection at all operational receivers _for each array design_.
# Include detection overlaps
dlist$algorithm$detection_overlaps <-
  acs_setup_detection_overlaps(dlist)

#### Example (1): Implement the function for acoustic time series only
dlist_acpf <- dlist
dlist_acpf$data$archival <- NULL
obs <- pf_setup_obs(.dlist = dlist_acpf,
                    .step = "2 mins",
                    .mobility = 500)
#> `patter::pf_setup_obs()` called @ 2024-03-18 14:28:09... 
#> ... Checking user inputs... 
#> ... Processing time series... 
#> ... Defining observations timeline... 
#> ... Adding acoustic observations: detection IDs... 
#> ... Adding acoustic observations: array IDs... 
#> ... Adding acoustic observations: matrices... 
#> ... Adding acoustic observations: containers... 
#> ... Completing call... 
#> `patter::pf_setup_obs()` call ended @ 2024-03-18 14:28:09 (duration: ~0 sec(s)). 
head(obs)
#>    timestep           timestamp mobility array_id detection acoustics container
#>       <int>              <POSc>    <num>    <int>     <int>    <list>    <list>
#> 1:        1 2016-03-17 01:50:00      500        1         1   1,0,0,0 <list[2]>
#> 2:        2 2016-03-17 01:52:00      500        1         1   1,0,0,0 <list[2]>
#> 3:        3 2016-03-17 01:54:00      500        1         1   1,0,0,0 <list[2]>
#> 4:        4 2016-03-17 01:56:00      500        1         0           <list[2]>
#> 5:        5 2016-03-17 01:58:00      500        1         1   1,0,0,0 <list[2]>
#> 6:        6 2016-03-17 02:00:00      500        1         1   1,0,0,0 <list[2]>

#### Example (2): Use alternative parameters
obs <- pf_setup_obs(.dlist = dlist_acpf,
                    .step = "2 mins",
                    .mobility = 1000)
#> `patter::pf_setup_obs()` called @ 2024-03-18 14:28:09... 
#> ... Checking user inputs... 
#> ... Processing time series... 
#> ... Defining observations timeline... 
#> ... Adding acoustic observations: detection IDs... 
#> ... Adding acoustic observations: array IDs... 
#> ... Adding acoustic observations: matrices... 
#> ... Adding acoustic observations: containers... 
#> ... Completing call... 
#> `patter::pf_setup_obs()` call ended @ 2024-03-18 14:28:09 (duration: ~0 sec(s)). 
head(obs)
#>    timestep           timestamp mobility array_id detection acoustics container
#>       <int>              <POSc>    <num>    <int>     <int>    <list>    <list>
#> 1:        1 2016-03-17 01:50:00     1000        1         1   1,0,0,0 <list[2]>
#> 2:        2 2016-03-17 01:52:00     1000        1         1   1,0,0,0 <list[2]>
#> 3:        3 2016-03-17 01:54:00     1000        1         1   1,0,0,0 <list[2]>
#> 4:        4 2016-03-17 01:56:00     1000        1         0           <list[2]>
#> 5:        5 2016-03-17 01:58:00     1000        1         1   1,0,0,0 <list[2]>
#> 6:        6 2016-03-17 02:00:00     1000        1         1   1,0,0,0 <list[2]>

#### Example (3): Implement the function for acoustic & archival time series
obs <- pf_setup_obs(.dlist = dlist,
                    .step = "2 mins",
                    .mobility = 500)
#> `patter::pf_setup_obs()` called @ 2024-03-18 14:28:09... 
#> ... Checking user inputs... 
#> ... Processing time series... 
#> ... Aligning time series... 
#> ... Defining observations timeline... 
#> ... Adding acoustic observations: detection IDs... 
#> ... Adding acoustic observations: array IDs... 
#> ... Adding acoustic observations: matrices... 
#> ... Adding acoustic observations: containers... 
#> ... Adding archival observations... 
#> ... Completing call... 
#> `patter::pf_setup_obs()` call ended @ 2024-03-18 14:28:09 (duration: ~0 sec(s)). 
head(obs)
#>    timestep           timestamp mobility array_id detection acoustics container
#>       <int>              <POSc>    <num>    <int>     <int>    <list>    <list>
#> 1:        1 2016-03-17 01:50:00      500        1         1   1,0,0,0 <list[2]>
#> 2:        2 2016-03-17 01:52:00      500        1         1   1,0,0,0 <list[2]>
#> 3:        3 2016-03-17 01:54:00      500        1         1   1,0,0,0 <list[2]>
#> 4:        4 2016-03-17 01:56:00      500        1         0           <list[2]>
#> 5:        5 2016-03-17 01:58:00      500        1         1   1,0,0,0 <list[2]>
#> 6:        6 2016-03-17 02:00:00      500        1         1   1,0,0,0 <list[2]>
#>    depth
#>    <num>
#> 1: 73.78
#> 2: 73.32
#> 3: 73.32
#> 4: 73.32
#> 5: 73.55
#> 6: 68.70

#### Example (4): Include full acoustic & archival time series
obs <- pf_setup_obs(.dlist = dlist,
                    .trim = FALSE,
                    .step = "2 mins",
                    .mobility = 500)
#> `patter::pf_setup_obs()` called @ 2024-03-18 14:28:10... 
#> ... Checking user inputs... 
#> ... Processing time series... 
#> ... Defining observations timeline... 
#> ... Adding acoustic observations: detection IDs... 
#> ... Adding acoustic observations: array IDs... 
#> ... Adding acoustic observations: matrices... 
#> ... Adding acoustic observations: containers... 
#> ... Adding archival observations... 
#> ... Completing call... 
#> `patter::pf_setup_obs()` call ended @ 2024-03-18 14:28:10 (duration: ~0 sec(s)). 
head(obs)
#>    timestep           timestamp mobility array_id detection acoustics container
#>       <int>              <POSc>    <num>    <int>     <int>    <list>    <list>
#> 1:        1 2016-03-17 00:00:00      500        1         0           <list[2]>
#> 2:        2 2016-03-17 00:02:00      500        1         0           <list[2]>
#> 3:        3 2016-03-17 00:04:00      500        1         0           <list[2]>
#> 4:        4 2016-03-17 00:06:00      500        1         0           <list[2]>
#> 5:        5 2016-03-17 00:08:00      500        1         0           <list[2]>
#> 6:        6 2016-03-17 00:10:00      500        1         0           <list[2]>
#>     depth
#>     <num>
#> 1: 134.61
#> 2: 135.77
#> 3: 135.53
#> 4: 136.00
#> 5: 135.07
#> 6: 134.38
tail(obs)
#>    timestep           timestamp mobility array_id detection acoustics container
#>       <int>              <POSc>    <num>    <int>     <int>    <list>    <list>
#> 1:      715 2016-03-17 23:48:00      500        1         0                    
#> 2:      716 2016-03-17 23:50:00      500        1         0                    
#> 3:      717 2016-03-17 23:52:00      500        1         0                    
#> 4:      718 2016-03-17 23:54:00      500        1         0                    
#> 5:      719 2016-03-17 23:56:00      500        1         0                    
#> 6:      720 2016-03-17 23:58:00      500        1         0                    
#>     depth
#>     <num>
#> 1: 156.61
#> 2: 156.61
#> 3: 156.61
#> 4: 156.61
#> 5: 156.84
#> 6: 156.61